% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{ADbreak}
\alias{ADbreak}
\title{NIMBLE language function to break tracking of derivatives}
\usage{
ADbreak(x)
}
\arguments{
\item{x}{scalar value}
}
\description{
This function is used in a method of a nimbleFunction that has derivatives enabled.  It returns its value but breaks tracking of derivatives.
}
\details{
This funcion only works with scalars.
}
