% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{ADNimbleList}
\alias{ADNimbleList}
\title{EXPERIMENTAL Data type for the return value of \code{\link{nimDerivs}}}
\format{An object of class \code{list} of length 1.}
\usage{
ADNimbleList
}
\description{
\code{\link{nimbleList}} definition for the type of \code{\link{nimbleList}} returned by \code{\link{nimDerivs}}.
}
\section{Fields}{

\describe{
\item{\code{value}}{The value of the function evaluated at the given input arguments.}

\item{\code{gradient}}{The gradient of the function evaluated at the given input arguments.}

\item{\code{hessian}}{The Hessian of the function evaluated at the given input arguments.}

\item{\code{thirdDerivs}}{Currently unused.}
}}
\seealso{
\code{\link{nimDerivs}}
}
\keyword{datasets}

