% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{any_na}
\alias{any_na}
\alias{any_nan}
\title{Determine if any values in a vector are NA or NaN}
\usage{
any_na(x)

any_nan(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
NIMBLE language functions that can be used in either compiled or uncompiled
nimbleFunctions to detect if there are any NA or NaN values in a vector.
}
\author{
NIMBLE Development Team
}

