% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d64b7a1f83
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497657d3b1a5
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871776a8110
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c347686f848
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b7dc6e865
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba69801dd5
>>>>>>> fixed up some roxygen stuff and
\name{asRow}
\alias{asCol}
\alias{asRow}
\title{Turn a numeric vector into a single-row or single-column matrix}
\usage{
asRow(x)
}
\arguments{
\item{x}{Numeric to be turned into a single row or column matrix}}\description{
Turns a numeric vector into a matrix that has 1 row or 1 column.  Part of NIMBLE language.
}
\details{
In the NIMBLE language, some automatic determination of how to turn vectors into single-row or single-column matrices is done.
For example, in \code{A \%*\% x}, where A is a matrix and x a vector, x will be turned into a single-column matrix unless
it is known at compile time that A is a single column, in which case x will be turned into a single-row matrix.
However, if it is desired that x be turned into a single row but A cannot be determined at compile time to be a single column,
then one can use \code{A \%*\% asRow(x)} to force this conversion.
}
\author{
Perry de Valpine
}

