% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmps9T35v/file3fb3776e765b
\name{autoBlock}
\alias{autoBlock}
\title{Automated parameter blocking for efficient MCMC sampling}
\usage{
autoBlock(Rmodel, autoIt = 20000, run = list("all", "default"),
  setSeed = TRUE, verbose = FALSE, makePlots = FALSE, round = TRUE)
}
\value{
info
}
\description{
Automated blocking description
}
\details{
Automated blocking details
}
\examples{
ab <- autoBlock(Rmodel)
}
\author{
Daniel Turek
}

