% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb867a91ce51
\name{buildConjugateSamplerFunctions}
\alias{buildConjugateSamplerFunctions}
\title{Rebuild conjugate sampler functions}
\usage{
buildConjugateSamplerFunctions(writeToFile = NULL)
}
\description{
Rebuild conjugate sampler functions
}

