% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{buildLaplace}
\alias{buildLaplace}
\title{Placeholder for buildLaplace}
\usage{
buildLaplace(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleQuad` package.
}
