% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_implementations.R
\name{CAR-Normal}
\alias{CAR-Normal}
\alias{dcar_normal}
\alias{rcar_normal}
\title{The CAR-Normal Distribution}
\usage{
dcar_normal(x, adj, weights, num, tau, c = CAR_calcNumIslands(adj, num),
  zero_mean = 0, log = FALSE)

rcar_normal(n = 1, adj, weights, num, tau, c = CAR_calcNumIslands(adj, num),
  zero_mean = 0)
}
\arguments{
\item{x}{vector of values.}

\item{adj}{vector of indicies of the adjacent locations (neighbors) of each spatial location.  This is a sparse representation of the full adjacency matrix.}

\item{weights}{vector of symmetric unnormalized weights associated with each pair of adjacent locations, of the same length as adj.  If omitted, all weights are taken as unity.}

\item{num}{vector giving the number of neighbors of each spatial location, with length equal to the total number of locations.}

\item{tau}{scalar precision of the Gaussian CAR prior.}

\item{c}{integer number of constraints to impose on the improper density function.  If omitted, c is calculated as the number of disjoint groups of spatial locations in the adjacency structure.}

\item{zero_mean}{integer specifying whether to impose a zero-mean constraint (during MCMC sampling) to all spatial regions (default 0).  If equal to 0, the overall process mean is included in the value of each location; if equal to 1, the mean of all locations is enforced to be zero and a separate intercept term should be included in the model.}

\item{log}{logical; if TRUE, probability density is returned on the log scale.}

\item{n}{number of observations.}
}
\value{
\code{dcar_normal} gives the density, and \code{rcar_normal} returns the current process values, since this distribution is improper.
}
\description{
Density function and random generation for the improper
  Gaussian conditional autoregressive (CAR) distribution.
}
\details{

}
\examples{
x <- c(1, 3, 3, 4)
num <- c(1, 2, 2, 1)
adj <- c(2, 1,3, 2,4, 3)
weights <- c(1, 1, 1, 1, 1, 1)
lp <- dcar_normal(x, adj, weights, num, tau = 1)
}
\author{
Daniel Turek
}
\references{
Banerjee, S., Carlin, B.P., and Gelfand, E.G. (2015). \emph{Hierarchical Modeling and Analysis for Spatial Data}, 2nd ed. Chapman and Hall/CRC.
}
\seealso{
\link{Distributions} for other standard distributions
}

