% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR.R
\name{CAR_calcNumIslands}
\alias{CAR_calcNumIslands}
\title{Calculate number of islands (non-recursive nimbleFunction version).}
\usage{
CAR_calcNumIslands(adj, num)
}
\description{
Calculate number of islands (non-recursive nimbleFunction version).
}

