% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_implementations.R
\name{Categorical}
\alias{Categorical}
\alias{dcat}
\alias{rcat}
\title{The Categorical Distribution}
\usage{
dcat(x, prob, log = FALSE)

rcat(n = 1, prob)
}
\arguments{
\item{x}{non-negative integer-value numeric value.}

\item{prob}{vector of probabilities, internally normalized to sum to one.}

\item{log}{logical; if TRUE, probability density is returned on the log scale.}

\item{n}{number of observations.}
}
\value{
\code{dcat} gives the density and \code{rcat} generates random deviates.
}
\description{
Density and random generation for the categorical distribution
}
\details{
See the BUGS manual for mathematical details.
}
\examples{
probs <- c(1/4, 1/10, 1 - 1/4 - 1/10)
x <- rcat(n = 30, probs)
dcat(x, probs)

}
\seealso{
\link{Distributions} for other standard distributions
}
\author{
Christopher Paciorek
}
