% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_distributions.R
\name{ChinesseRestaurantProcess}
\alias{ChinesseRestaurantProcess}
\alias{dCRP}
\alias{rCRP}
\title{The Chinesse Restaurant Process Distribution}
\usage{
rCRP(n, conc = 1)

dCRP(x, conc = 1, log = 0)
}
\arguments{
\item{n}{number of observations.}

\item{conc}{scalar with concentration parameter.}

\item{x}{vector of values.}

\item{log}{logical; if TRUE, probability density is returned on the log scale.}
}
\value{
\code{dCRP} gives the density, and \code{rCRP} gives random generation.
}
\description{
Density and random generation for the Chinesse
  Restaurant Process distribution with concentration 
  parameter \code{conc}.
}
\details{
The chineese restaurant process distribution is a distribution
on the space of partitions of the positive integers. 
The chineese restaurant process
distribution with concentration parameter \eqn{=\alpha}{= conc} has 
probability function 
\deqn{f(x_i \mid x_1, \ldots, x_{i-1})=\frac{1}{i-1+\alpha}\sum_{j=1}^{i-1}\delta_{x_j}+
\frac{\alpha}{i-1+\alpha}\delta_{x^{new}},}
where \eqn{x^{new}} is a new integer not in \eqn{x_1, \ldots, x_{i-1}}.

If \code{conc} is not specified, it assumes the default value of 1.
}
\examples{
x <- rCRP(10, conc = 1)
dCRP(x, conc = 1)
}
\author{
Claudia Wehrhahn
}
\references{
Blackwell, D., & MacQueen, J. B. (1973) Ferguson distributions via 
Pólya urn schemes \emph{The annals of statistics}, 353-355.

Aldous, D. J. (1985) Exchangeability and related topics \emph{In École d'Été de 
Probabilités de Saint-Flour XIII} 1983 (pp. 1-198) Springer, Berlin, Heidelberg.

Pitman, Jim. Some developments of the Blackwell-MacQueen urn scheme 
\emph{Statistics, probability and game theory}, 245--267, Institute of
 Mathematical Statistics, Hayward, CA, 1996.
}

