% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d27be14bb
\docType{class}
\name{codeBlockClass-class}
\alias{codeBlockClass}
\alias{codeBlockClass-class}
\title{Class \code{codeBlockClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

