% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmp3xd4YE/filefb2408a832a
\name{combine_MCMC_comparison_results}
\alias{combine_MCMC_comparison_results}
\title{Combine multiple objects returned by compareMCMCs}
\usage{
combine_MCMC_comparison_results(..., name = "MCMCresults")
}
\arguments{
\item{...}{objects returned by \code{\link{compareMCMCs}}}\item{name}{(default "MCMCresults") name to be given to the resulting set of comparisons.  This will be used in the html generated by \code{\link{make_MCMC_comparison_pages}}.  It is simply a list name and thus can easily be modified later.}}\value{
An object in the same format as returned by \code{\link{compareMCMCs}} with \code{summary} = TRUE.
}
\description{
Useful for running different cases separately and combining them.
}
\details{
Only the summary, timing, and efficiency elements will survive the combination.  MCMC samples will not.
}
\seealso{
\code{\link{compareMCMCs}}, \code{\link{rename_MCMC_comparison_method}}, \code{\link{make_MCMC_comparison_pages}}, \code{\link{reshape_comparison_results}}.
}

