% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee45ff4b22
=======
% Please edit documentation in Rtmps9T35v/file3fb31bd37e83
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b048d0d3dc
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d69a87d69
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976666923e
>>>>>>> updated suite and autoblock roxygen
\name{dconstraint}
\alias{dconstraint}
\alias{rconstraint}
\title{Constraint calculations in NIMBLE}
\usage{
dconstraint(x, cond, log = FALSE)
}
\arguments{
  \item{x}{value indicating whether \code{cond} is TRUE or
  FALSE} \item{cond}{logical value} \item{log}{logical; if
  TRUE, probability density is returned on the log scale.}
  \item{n}{number of observations (only \code{n=1} is
  handled currently).}
}
\value{
\code{dconstraint} gives the density and \code{rconstraint}
generates random deviates, but these are unusual as the
density is 1 if \code{x} matches \code{cond} and 0 otherwise
and the deviates are simply the value of \code{cond}
}
\description{
Calculations to handle censoring
}
\details{
Used for working with constraints in BUGS code. See the
NIMBLE manual for additional details.
}
\examples{
constr <- 3 > 2 && 4 > 0
x <- rconstraint(1, constr)
dconstraint(x, constr)
dconstraint(0, 3 > 4)
dconstraint(1, 3 > 4)
rconstraint(1, 3 > 4)
}
\author{
Christopher Paciorek
}
\seealso{
\link{Distributions} for other standard distributions
}

