<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MCMC_utils.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb866314149a
>>>>>>> upstream/devel
\name{decide}
\alias{decide}
\title{Makes the Metropolis-Hastings acceptance decision, based upon the input (log) Metropolis-Hastings ratio}
\usage{
decide(logMetropolisRatio)
}
\arguments{
<<<<<<< HEAD
\item{logMetropolisRatio}{The log of the Metropolis-Hastings ratio, which is calculated from model probabilities and forward/reverse transition probabilities}
}
\description{
This function returns a logical TRUE/FALSE value, indicating whether the proposed transition should be accepted (TRUE) or rejected (FALSE).
}
\details{
The Metropolis-Hastings accept/reject decisions is made as follows.  If \code{logMetropolisRatio} is greater than 0, accept (return \code{TRUE}).  Otherwise draw a uniform random number between 0 and 1 and accept if it is less that \code{exp(logMetropolisRatio}.  The proposed transition will be rejected (return \code{FALSE}). If \code{logMetropolisRatio} is NA, NaN, or -Inf, a reject (\code{FALSE}) decision will be returned.
=======
  \item{logMetropolisRatio}{The log of the
  Metropolis-Hastings ratio, which is calculated from model
  probabilities and forward/reverse transition
  probabilities}
}
\description{
This function returns a logical TRUE/FALSE value,
indicating whether the proposed transition should be
accepted (TRUE) or rejected (FALSE).
}
\details{
The Metropolis-Hastings accept/reject decisions is made as
follows.  If \code{logMetropolisRatio} is greater than 0,
accept (return \code{TRUE}).  Otherwise draw a uniform
random number between 0 and 1 and accept if it is less that
\code{exp(logMetropolisRatio}.  The proposed transition
will be rejected (return \code{FALSE}). If
\code{logMetropolisRatio} is NA, NaN, or -Inf, a reject
(\code{FALSE}) decision will be returned.
>>>>>>> upstream/devel
}
\examples{
jump <- decide(lMHr)
}
\author{
Daniel Turek
}

