% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_processInputList.R
\name{deregisterDistributions}
\alias{deregisterDistributions}
\title{Remove user-supplied distributions from use in NIMBLE BUGS models}
\usage{
deregisterDistributions(
  distributionsNames,
  userEnv = parent.frame(),
  warn = TRUE
)
}
\arguments{
\item{distributionsNames}{a character vector giving the names of the distributions to be deregistered.}

\item{userEnv}{environment in which to look for the nimbleFunctions that provide the distribution; this will generally not need to be set by the user as it will default to the environment from which this function was called.}
}
\description{
Deregister distributional information originally supplied by the user
for use in BUGS model code
}
\author{
Christopher Paciorek
}
