% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee45ff4b22
=======
% Please edit documentation in RtmpLwENQY/file61612f4fe2c0
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb31bd37e83
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b048d0d3dc
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d69a87d69
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976666923e
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file787133664cae
>>>>>>> added line to remove .o files after create
\name{dinterval}
\alias{dinterval}
\alias{rinterval}
\title{Interval calculations}
\usage{
dinterval(x, t, c, log = FALSE)
}
\arguments{
  \item{x}{vector of interval indices.} \item{t}{vector of
  values.} \item{c}{vector of one or more values delineating
  the intervals.} \item{log}{logical; if TRUE, probability
  density is returned on the log scale.} \item{n}{number of
  observations.}
}
\value{
\code{dinterval} gives the density and \code{rinterval}
generates random deviates, but these are unusual as the
density is 1 if \code{x} indicates the interval in which
\code{t} falls and 0 otherwise and the deviates are simply
the interval(s) in which \code{t} falls.
}
\description{
Calculations to handle censoring
}
\details{
Used for working with censoring in BUGS code. Taking
\code{c} to define the endpoints of two or more intervals
(with implicit endpoints of plus/minus infinity), \code{x}
(or the return value of \code{rinterval}) gives the non-
negative integer valued index of the interval in which
\code{t} falls. See the NIMBLE manual for additional
details.
}
\examples{
endpoints <- c(-3, 0, 3)
vals <- c(-4, -1, 1, 5)
x <- rinterval(4, vals, endpoints)
dinterval(x, vals, endpoints)
dinterval(c(1, 5, 2, 3), vals, endpoints)
}
\author{
Christopher Paciorek
}
\seealso{
\link{Distributions} for other standard distributions
}

