% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee45ff4b22
=======
% Please edit documentation in RtmpLwENQY/file61612f4fe2c0
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb31bd37e83
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b048d0d3dc
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d69a87d69
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976666923e
>>>>>>> updated suite and autoblock roxygen
\name{dmulti}
\alias{dmulti}
\alias{rmulti}
\title{The Multinomial Distribution}
\usage{
dmulti(x, size = sum(x), prob, log = FALSE)
}
\arguments{
  \item{x}{vector of values.} \item{prob}{vector of
  probabilities, summing to one, of same length as
  \code{x}} \item{log}{logical; if TRUE, probability
  density is returned on the log scale.} \item{n}{number of
  observations (only \code{n=1} is handled currently).}
}
\value{
\code{dmulti} gives the density and \code{rmulti} generates
random deviates.
}
\description{
Density and random generation for the multinomial
distribution
}
\details{
See Gelman et al., Appendix A or the BUGS manual for
mathematical details.
}
\examples{
size <- 30
probs <- c(1/4, 1/10, 1 - 1/4 - 1/10)
x <- rmulti(1, size, probs)
dmulti(x, size, probs)
}
\author{
Christopher Paciorek
}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., and Rubin, D.B.
(2004) \emph{Bayesian Data Analysis}, 2nd ed. Chapman and
Hall/CRC.
}
\seealso{
\link{Distributions} for other standard distributions
}

