% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee45ff4b22
=======
% Please edit documentation in RtmpLwENQY/file61612f4fe2c0
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb31bd37e83
>>>>>>> various updates to roxygen
\name{dt_nonstandard}
\alias{dt_nonstandard}
\alias{pt_nonstandard}
\alias{qt_nonstandard,}
\alias{rt_nonstandard,}
\title{The t Distribution}
\usage{
dt_nonstandard(x, df = 1, mu = 0, sigma = 1, log = FALSE)
}
\arguments{
  \item{x}{vector of values.} \item{df}{vector of degrees
  of freedom values.} \item{mu}{vector of location values.}\item{sigma}{vector of scale values.} \item{log}{logical;
  if TRUE, probability density is returned on the log
  scale.} \item{n}{number of observations.} \item{log.p}
  {logical; if TRUE, probabilities p are given by user as
  log(p).} \item{lower.tail}{logical; if TRUE (default)
  probabilities are \eqn{P[X \le x]}; otherwise, \eqn{P[X >
  x]}.}
}
\value{
\code{dt_nonstandard} gives the density,
\code{pt_nonstandard} gives the distribution function,
\code{qt_nonstandard} gives the quantile function, and
\code{rt_nonstandard} generates random deviates.
}
\description{
Density, distribution function, quantile function and random
generation for the t distribution with \code{df} degrees of
freedom, allowing non-zero location, \code{mu}, and non-unit
scale, \code{sigma}
}
\details{
See Gelman et al., Appendix A or the BUGS manual for
mathematical details.
}
\examples{
x <- rt_nonstandard(50, df = 1, mu = 5, sigma = 1)
dt_nonstandard(x, 3, 5, 1)
}
\author{
Christopher Paciorek
}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., and Rubin, D.B.
(2004) \emph{Bayesian Data Analysis}, 2nd ed. Chapman and
Hall/CRC.
}
\seealso{
\link{Distributions} for other standard distributions
}

