% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscAlgorithms.r
\name{expmA}
\alias{expmA}
\title{Matrix Exponential}
\usage{
expmA(A, tol = 1e-08)
}
\arguments{
\item{A}{Infinitesimal Generator Matrix.}

\item{tol}{level of accuracy required (default = 1e-8).}
}
\value{
\code{expmA} gives a matrix that is ans = exp(A).
}
\description{
Compute the the matrix exponential expm(A)
  by scaling and squaring.
}
\details{
A, the generator matrix must have negative value diagonals.

This function follows the scaling and squaring algorithm from Sherlock (2021), except that we compute the full
matrix exponential. It differs from the standard Taylor scaling and squaring algorithm reviewed by Ruiz et al (2016) and found in common texts. 
If using the matrix exponential to create a transition probability matrix in a HMM context just once, 
this function is good. If dimension is large, we recommend avoiding the matrix exponential and using `expmAv` instead. 
Note that for computation efficiency matrix uniformization is always done by A* = A + rho I, where rho = max(abs(diag(A))). 
Derivatives are currently not supported.
}
\examples{
A <- rbind(c(-1, 0.25, 0.75), c(0, -2, 2), c(0.25, 0.25, -0.5))
Lambda <- diag(c(0.25, 0.1, 0))
expmA((A-Lambda)*2.5)
}
\references{
Sherlock, C. (2021). Direct statistical inference for finite Markov jump processes via the matrix exponential. Computational Statistics, 36(4), 2863-2887.

Ruiz, P., Sastre, J., Ibáñez, J., & Defez, E. (2016). High performance computing of the matrix exponential. 
Journal of computational and applied mathematics, 291, 370-379.
}
\author{
Paul van Dam-Bates
}
