% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscAlgorithms.r
\name{expmAv}
\alias{expmAv}
\title{Matrix Exponential times a vector}
\usage{
expmAv(A, v, tol = 1e-08, rescaleFreq = 10, Nmax = 10000, sparse = FALSE)
}
\arguments{
\item{A}{Infinitesimal Generator Matrix.}

\item{v}{vector to multiply by the matrix exponential exp(A) %*% v.}

\item{tol}{level of accuracy required (default = 1e-8).}

\item{rescaleFreq}{How frequently should the terms be scaled to avoid underflow/overflow (default = 10).}

\item{Nmax}{maximum number of iterations to compute (default = 10000).}

\item{sparse}{(logical) specify if the matrix may be sparse and to do sparse computation (default = FALSE).}
}
\value{
\code{expmAv} gives a vector that is ans = exp(A) %*% v.
}
\description{
Compute the combined term expm(A) %*% v
  to avoid a full matrix exponentiation.
}
\details{
For large matrix exponentials it is much more efficient to compute exp(A) %*% v, than to actually compute the entire matrix exponential. A, the generator matrix must have
negative value diagonals.

This function follows the function `expAv` from the R package RTMB (Kristensen, 2025), and theory outlined in Sherlock (2021). It is developed for working with continuous times
Markov chains. If using the matrix exponential to create a transition probability matrix in a HMM context just once, 
this function may be slower than the one time call to compute the full matrix exponentiation. If a full matrix exponentiation is required, refer to `nimbleRcall` and 
functions such as `pracma::expm` or `expm::expm` to compute. Choosing sparse = TRUE will then check which matrix A values are non-zero and do sparse linear 
algebra on the non-zero terms only. Note that for computation efficiency matrix uniformization is always done by A* = A + rho I, where rho = max(abs(diag(A))), see Algorithm 2#' in Sherlock (2021).
}
\examples{
A <- rbind(c(-1, 0.25, 0.75), c(0, -2, 2), c(0.25, 0.25, -0.5))
v <- c(0.35, 0.25, 0.1)
expmAv(A, v)
}
\references{
Sherlock, C. (2021). Direct statistical inference for finite Markov jump processes via the matrix exponential. Computational Statistics, 36(4), 2863-2887.

Kristensen K (2025). _RTMB: 'R' Bindings for 'TMB'_. R package version 1.7, commit 6bd7a16403ccb4d3fc13ff7526827540bf27b352, 
<https://github.com/kaskr/RTMB>.
}
\author{
Paul van Dam-Bates
}
