% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{getBound}
\alias{getBound}
\title{Get value of bound of a stochastic node in a model}
\usage{
getBound(model, node, bound, nodeFunctionIndex)
}
\arguments{
\item{model}{A NIMBLE model object}

\item{node}{The name of a stochastic node in the model}

\item{bound}{Either \code{'lower'} or \code{'upper'} indicating the desired bound for the node}

\item{nodeFunctionIndex}{For internal NIMBLE use only}
}
\description{
Part of the NIMBLE language
}
\details{
For nodes that do not involve truncation of the distribution
this will return the lower or upper bound of the distribution, which
may be a constant or for a limited number of distributions a parameter
or functional of a parameter (at the moment in NIMBLE, the only case
where a bound is a parameter is for the uniform distribution. For nodes
that are truncated, this will return the desired bound, which may be
a functional of other quantities in the model or may be a constant.
}
