% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpOrtTO2/file33512e205b8b
\name{getBUGSexampleDir}
\alias{getBUGSexampleDir}
\title{Get the directory path to one of the classic BUGS examples installed with NIMBLE package}
\usage{
getBUGSexampleDir(example)
}
\arguments{
\item{example}{The name of the classic BUGS example.}
}
\value{
Character string of the fully pathed directory of the BUGS example.
}
\description{
NIMBLE comes with some of the classic BUGS examples.  \code{getBUGSexampleDir} looks up the location of an example from its name.
}
\author{
Christopher Paciorek
}

