% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp9zTnuX/file4e27c86f778
\name{getDefinition}
\alias{getDefinition}
\title{Get nimbleFunction definition}
\usage{
getDefinition(nf)
}
\arguments{
\item{nf}{A nimbleFunction generator, or a compiled or un-compiled specialized nimbleFunction.}
}
\description{
Returns a list containing the nimbleFunction definition components (setup function, run function, and other member methods) for the supplied nimbleFunction generator or specialized instance.
}
\author{
Daniel Turek
}

