<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nimbleProject.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb867b867383
>>>>>>> upstream/devel
\name{getNimbleProject}
\alias{getNimbleProject}
\title{return the nimble project, if any, associated with a model or nimbleFunction object.}
\usage{
getNimbleProject(project, stopOnNull = FALSE)
}
\description{
<<<<<<< HEAD
return the nimble project, if any, associated with a model or nimbleFunction object.
=======
return the nimble project, if any, associated with a model
or nimbleFunction object.
>>>>>>> upstream/devel
}

