% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpVY8pk7/file3afa418b6638
\name{identityMatrix}
\alias{identityMatrix}
\title{Create an Identity matrix}
\usage{
identityMatrix(d)
}
\arguments{
\item{d}{The size of the identity matrix to return, will return a d-by-d matrix}
}
\value{
A d-by-d identity matrix
}
\description{
Returns a d-by-d identity matrix (square matrix of 0's, with 1's on the main diagnol).
}
\details{
This function can be used in the NIMBLE DSL, i.e. in the run function and member methods of nimbleFunctions.
}
\examples{
Id <- identityMatrix(d = 3)
}
\author{
Daniel Turek
}

