% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{is.na.vec}
\alias{is.na.vec}
\alias{is.nan.vec}
\title{Determine if any values in a vector are NA or NaN}
\usage{
\method{is.na}{vec}(x)

\method{is.nan}{vec}(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
NIMBLE language functions that can be used in either compiled or uncompiled
nimbleFunctions to detect if there are any NA or NaN values in a vector.
}
\author{
NIMBLE Development Team
}
