% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\name{is.nl}
\alias{is.nl}
\title{check if a nimbleList}
\usage{
is.nl(l)
}
\arguments{
\item{l}{object to be tested}
}
\description{
Checks an object to determine if it is a nimbleList (i.e., a list created by \code{nlDef$new()}).
}
\seealso{
\code{\link{nimbleList}} for how to create a nimbleList
}
