% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{makeBoundInfo}
\alias{makeBoundInfo}
\title{Make an object of information about a model-bound pairing for getBound.  Used internally}
\usage{
makeBoundInfo(model, nodes, bound)
}
\arguments{
\item{model}{A model such as returned by \code{\link{nimbleModel}}.}

\item{nodes}{A character string naming a stochastic nodes, such as \code{'mu'}.}

\item{bound}{A character string naming a bound of the distribution, either \code{'lower'} or \code{'upper'}.}
}
\description{
Creates a simple getBound_info object, which has a list with a boundID and a type.
Unlike \code{makeParamInfo} this is more bare-bones, but keeping it for parallelism
with \code{getParam}.
}
\details{
This is used internally by \code{\link{getBound}}.  It is not intended for direct use by a user or even a nimbleFunction programmer.
}
