% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmpgvo3o8/file189d2b52d551
\name{makeParamInfo}
\alias{makeParamInfo}
\title{Make an object of information about a model-parameter pairing for getParam.  Used internally}
\usage{
makeParamInfo(model, nodes, param)
}
\arguments{
\item{model}{A model such as returned by \link{nimbleModel}.}

\item{param}{A character string naming a parameter of the distribution followed by node, such as "mean", "rate", "lambda", or whatever parameter names are relevant for the distribution of the node.}

\item{node}{A character string naming a stochastic node, such as "mu", "beta[2]", or "eta[1:3, 2]"}
}
\description{
Creates a simple getParam_info object, which has a list with a paramID and a type
}
\details{
This is used internally by \link{getParam}.  It is not intended for direct use by a user or even a nimbleFunction programmer.
}

