% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_utils.R
\name{mcmc_createModelObject}
\alias{mcmc_createModelObject}
\title{Create model object using arguments provided to nimbleMCMC or runMCMC}
\usage{
mcmc_createModelObject(
  model,
  inits,
  nchains,
  setSeed,
  code,
  constants,
  data,
  dimensions,
  check
)
}
\description{
For internal use only
}
