% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpXAHF9Y/file39b66cc1ff95
\name{MCMCsuite}
\alias{MCMCsuite}
\title{Executes multiple MCMC algorithms and organizes results.}
\usage{
MCMCsuite(...)
}
\arguments{
\item{...}{Arguments that will be passed to \code{MCMCsuiteClass()}}}\description{
Creates, runs, and organizes output from a suite of MCMC algorithms, all applied to the same model, data, and initial values.
This can include WinBUGS, OpenBUGS, JAGS and Stan MCMCs, as well as NIMBLE MCMC algorithms.
Trace plots and density plots for the MCMC samples may also be generated and saved.
}
\details{
Please see \code{help(MCMCsuiteClass)} for details on how to use this function.
}

