% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modelCode}
\alias{modelCode}
\title{Turn BUGS model code into an object for use in \code{nimbleModel} or \code{readBUGSmodel}}
\usage{
modelCode(code)
}
\arguments{
  \item{code}{expression providing the code for the model}
}
\description{
Simply keeps model code as an R call object, the form
needed by \code{nimbleModel} and optionally usable by
\code{readBUGSmodel}
}
\details{
It is equivalent to use the R function \code{quote}.
\code{modelCode} is simply provided as a more readable
alternative for NIMBLE users not familiar with
\code{quote}.
}
\examples{
modelCode <- BUGScode({
 x ~ dnorm(mu, sd = 1)
 mu ~ dnorm(0, sd = prior_sd)
})
}
\author{
Daniel Turek
}

