% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGS_model.R
\name{modelInitialization}
\alias{modelInitialization}
\title{Information on initial values in a nimbleModel}
\description{
Having uninitialized nodes in a nimbleModel can potentially cause some algorithms to fail, and can lead to poor performance in others.  Here are some
 general guidelines on  how non-intitialized variables can affect performance:
 \itemize{
   \item MCMC will atuo-initialize, but will do so from the prior distribution.  This can cause slow convergence, especially in the case of diffuse priors.
   \item Likewise, particle filtering methods will initialize top-level parameters from their prior distributions, which can lead to errors or poor performance in these methods.
}
}

