% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/types_modelVariableAccessor.R
\docType{class}
\name{modelVariableAccessorVector-class}
\alias{modelVariableAccessorVector}
\alias{modelVariableAccessorVector-class}
\title{##                   first = 'ANY',numeric',
##                   last  = 'ANY',numeric',
##                   length = 'ANY'numeric'}
\description{
##                   first = 'ANY',numeric',
##                   last  = 'ANY',numeric',
##                   length = 'ANY'numeric'
}

