% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpSlIkmN/file236958e547a2
\name{modelVariableAccessorVector}
\alias{modelVariableAccessorVector}
\title{##                   first = 'ANY',numeric',
##                   last  = 'ANY',numeric',
##                   length = 'ANY'numeric'}
\usage{
modelVariableAccessorVector(mv, nodeNames, logProb = FALSE)
}
\description{
## first = 'ANY',numeric', ## last = 'ANY',numeric', ##
length = 'ANY'numeric'
}

