% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp6hGOZh/file12a1f5ae2d6b1
\docType{data}
\name{ModifiedRmmParseKeywords2}
\alias{ModifiedRmmParseKeywords2}
\title{[[' = 'outputCppArrayIndex2',}
\format{An object of class \code{list} of length 40.}
\usage{
ModifiedRmmParseKeywords2
}
\description{
[[' = 'outputCppArrayIndex2',
}
\keyword{datasets}

