<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nfMethod}
\alias{nfMethod}
\title{access a member function (or method) of a nimbleFunction}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb8615493aea
\name{nfMethod}
\alias{nfMethod}
\title{Internal function for accessing a member function (method) of a nimbleFunction.  Normally a user will write \code{nf$method(x)} instead of \code{nfMethod(nf, method)(x)}.}
>>>>>>> upstream/devel
\usage{
nfMethod(nf, methodName)
}
\arguments{
<<<<<<< HEAD
\item{nf}{a specialized nimbleFunction, i.e. one that has already had setup parameters processed}

\item{methodName}{a character string giving the name of the member function to call}
=======
  \item{nf}{a specialized nimbleFunction, i.e. one that has
  already had setup parameters processed}

  \item{methodName}{a character string giving the name of
  the member function to call}
>>>>>>> upstream/devel
}
\value{
a function that can be called.
}
\description{
<<<<<<< HEAD
access (call) a member function of a nimbleFunction other than \code{run}, which calling the nimbleFunction directly uses.  Works in R and NIMBLE.
}
\details{
nimbleFunctions have a default member function called \code{run}, and may have other member functions provided via the \code{methods} argument to \code{nimbleFunction}.
\code{nfMethod} is primarily used to call the latter, although it could be used with \code{methodName = 'run'}.  Normally arguments will be provided after \code{nfMethod}, e.g. \code{nfMethod(myNimbleFunction, 'foo')(x)}.
=======
access (call) a member function of a nimbleFunction,
including \code{run}.
}
\details{
nimbleFunctions have a default member function called
\code{run}, and may have other member functions provided
via the \code{methods} argument to \code{nimbleFunction}.
As an internal step, the NIMBLE compiler turns
\code{nf$method(x)} into \code{nfMethod(nf, method)(x)},
but a NIMBLE user or programmer would not normally need to
use \code{nfMethod} directly.
>>>>>>> upstream/devel
}
\author{
NIMBLE development team
}

