\name{nimble-math}
\title{Mathematical functions for BUGS and nimbleFunction programming}
\alias{besselK}
\alias{cloglog}
\alias{cube}
\alias{expit}
\alias{icloglog}
\alias{ilogit}
\alias{inprod}
\alias{inverse}
\alias{iprobit}
\alias{logdet}
\alias{logfact}
\alias{loggam}
\alias{logit}
\alias{nimEquals}
\alias{nimRound}
\alias{nimStep}
\alias{nimSwitch}
\alias{phi}
\alias{pow}
\alias{probit}
\description{
Mathematical functions for use in BUGS code and in nimbleFunction programming (i.e., nimbleFunction run code). See Chapter 5 of the User Manual for more details.
}
\author{NIMBLE Development Team}