% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-package.r
\docType{package}
\name{nimble-package}
\alias{nimble}
\alias{nimble-package}
\title{nimble: MCMC, Particle Filtering, and Programmable Hierarchical Modeling}
\description{
A system for writing hierarchical statistical models largely compatible with 'BUGS' and 'JAGS', writing nimbleFunctions to operate models and do basic R-style math, and compiling both models and nimbleFunctions via custom-generated C++. 'NIMBLE' includes default methods for MCMC, Laplace Approximation, deterministic nested approximations, Monte Carlo Expectation Maximization, and some other tools. The nimbleFunction system makes it easy to do things like implement new MCMC samplers from R, customize the assignment of samplers to different parts of a model from R, and compile the new samplers automatically via C++ alongside the samplers 'NIMBLE' provides. 'NIMBLE' extends the 'BUGS'/'JAGS' language by making it extensible: New distributions and functions can be added, including as calls to external compiled code. Although most people think of MCMC as the main goal of the 'BUGS'/'JAGS' language for writing models, one can use 'NIMBLE' for writing arbitrary other kinds of model-generic algorithms as well. A full User Manual is available at \url{https://r-nimble.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-nimble.org}
  \item \url{https://github.com/nimble-dev/nimble}
  \item Report bugs at \url{https://github.com/nimble-dev/nimble/issues}
}

}
\author{
\strong{Maintainer}: Christopher Paciorek \email{paciorek@stat.berkeley.edu}

Authors:
\itemize{
  \item Perry de Valpine
  \item Daniel Turek
  \item Nick Michaud
  \item Cliff Anderson-Bergman
  \item Fritz Obermeyer
  \item Claudia Wehrhahn Cortes (Bayesian nonparametrics system)
  \item Abel Rodríguez (Bayesian nonparametrics system)
  \item Duncan Temple Lang (packaging configuration)
  \item Wei Zhang (Laplace approximation)
  \item Sally Paganin (reversible jump MCMC)
  \item Joshua Hug (WAIC)
  \item Paul van Dam-Bates (Pólya-Gamma sampler, nimIntegrate, matrix exponential)
}

Other contributors:
\itemize{
  \item Jagadish Babu (code for the compilation system for an early version of NIMBLE) [contributor]
  \item Lauren Ponisio (contributions to the cross-validation code) [contributor]
  \item Peter Sujan (multivariate t distribution code) [contributor]
}

}
\keyword{internal}
