\name{nimble}
\alias{nimble}
\alias{nimble-package}
\title{Flexible BUGS-Compatible System for Hierarchical Statistical
  Modeling and Algorithm Development}
\description{
Flexible application of algorithms to models specified in the BUGS language. Algorithms can be written in the NIMBLE language and made available to any model.
}
\details{
  Please see the NIMBLE home page (\url{http://r-nimble.org}) and the
  NIMBLE User Manual. The User Manual is available at
  \url{http://r-nimble.org/manuals/NimbleUserManual.pdf}  or available
  from the \name{doc} directory of the installed R package.}
  \author{NIMBLE Development Team}


