% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nimbleDim}
\alias{nimbleDim}
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nimbleDim(obj)
}
\arguments{
  \item{obj}{objects for which the sizes are requested}
}
\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.
It is useful to have this function that treats a vector
similarly to a matrix or array.  Works in R and NIMBLE.  In
NIMBLE \code{dim} is identical to \code{nimbleDim}, not to
R's \code{dim}
}
\examples{
x <- rnorm(4)
dim(x)
nimbleDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nimbleDim(y)
}
\author{
NIMBLE development team
}

