% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee3048d1a7
=======
% Please edit documentation in RtmpLwENQY/file6161341da6ca
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb34bf85ace
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b06807e4a
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d641d1a21a
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976516d1f67
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file787144043461
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c34231b7291
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b70a6902d
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba2d9e88a6
>>>>>>> fixed up some roxygen stuff and
=======
% Please edit documentation in Rtmpzdod9z/file79ee3048d1a7
>>>>>>> added man files from devel to ease rebase
=======
% Please edit documentation in Rtmpzdod9z/file79ee3048d1a7
>>>>>>> 6f0bc2e7a0885e03527d78f21be97fad5c54074a
\name{nimbleFunctionVirtual}
\alias{nimbleFunctionVirtual}
\title{create a virtual nimbleFunction, a base class for other nimbleFunctions}
\usage{
nimbleFunctionVirtual(contains = NULL, run = function() { },
  methods = list(), name = NA)
}
\arguments{
<<<<<<< HEAD
\item{contains}{Not yet functional}\item{run}{A NIMBLE function that will only be used to inspect its argument types and returnType.}\item{methods}{An optional named list of NIMBLE functions that will also only be used for inspecting argument types and returnTypes.}\item{name}{An optional name used internally by the NIMBLE compiled.  This is usually omitted and NIMBLE provides one.}}\value{
An object that can be passed as the \code{contains} argument to \code{nimbleFunction} or as the argument to \code{nimbleFunctionList}
}
\description{
define argument types and returnType for the \code{run} function and any \code{methods}, to be used in the \code{contains} argument of \code{nimbleFunction}
}
\details{
See the NIMBLE User Manual section on nimbleFunctionLists for explanation of how to use a virtual nimbleFunction.
=======
  \item{contains}{Not yet functional} \item{run}{A NIMBLE
  function that will only be used to inspect its argument
  types and returnType.} \item{methods}{An optional named
  list of NIMBLE functions that will also only be used for
  inspecting argument types and returnTypes.} \item{name}{An
  optional name used internally by the NIMBLE compiled. This
  is usually omitted and NIMBLE provides one.}
}
\value{
An object that can be passed as the \code{contains}
argument to \code{nimbleFunction} or as the argument to
\code{nimbleFunctionList}
}
\description{
define argument types and returnType for the \code{run}
function and any \code{methods}, to be used in the
\code{contains} argument of \code{nimbleFunction}
}
\details{
See the NIMBLE User Manual section on nimbleFunctionLists
for explanation of how to use a virtual nimbleFunction.
>>>>>>> added man files from devel to ease rebase
}
\author{
NIMBLE development team
}
\seealso{
\link{nimbleFunction}
}

