% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee1e8d0966
=======
% Please edit documentation in RtmpLwENQY/file61612b823b49
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb35b4fa7ed
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b079b14439
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6518728f4
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file49762eb6cc9
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78717038320b
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c344ed94e0b
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b6904a114
>>>>>>> removed top level tests dir
\name{nimbleOptions}
\alias{nimbleOptions}
\title{NIMBLE Options Settings}
\usage{
nimbleOptions(...)
}
\arguments{
  \item{...}{any options to be defined as one or more 'name
  = value' pairs. Options can also be passed by giving a
  single unnamed argument that is a named list.}
}
\value{
When invoked with no arguments, a list with the current
values of all options.
}
\description{
Allow the user to set and examine a variety of global
_options_ that affect the way in which NIMBLE operates
}
\details{
\code{nimbleOptions} mimics \code{options}. Invoking
\code{nimbleOptions()} with no arguments returns a list with
the current values of the options. To access the value of a
single option, one should use \code{getNimbleOption()}.
}
\examples{
nimbleOptions(verifyConjugatePosteriors = FALSE)
}
\author{
Christopher Paciorek
}

