<<<<<<< HEAD:packages/nimble/man/nimbleDim.Rd
% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/all_utils.R
=======
% Please edit documentation in RtmpgVE0pU/file7ba03a6f273b
>>>>>>> upstream/devel
\name{nimbleDim}
\alias{nimbleDim}
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb86392fb2e1
\name{nimDim}
\alias{nimDim}
>>>>>>> upstream/devel:packages/nimble/man/nimDim.Rd
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nimDim(obj)
}
\arguments{
\item{obj}{objects for which the sizes are requested}
}
\value{
a vector of sizes in each dimension
}
\description{
<<<<<<< HEAD
R's regular \code{dim} function returns NULL for a vector.  It is useful to have this function that treats a vector similarly to a matrix or array.  Works in R and NIMBLE.  In NIMBLE \code{dim} is identical to \code{nimbleDim}, not to R's \code{dim}
=======
R's regular \code{dim} function returns NULL for a vector.
It is useful to have this function that treats a vector
<<<<<<< HEAD:packages/nimble/man/nimbleDim.Rd
similarly to a matrix or array. Works in R and NIMBLE. In
NIMBLE \code{dim} is identical to \code{nimbleDim}, not to
R's \code{dim}
>>>>>>> upstream/devel
=======
similarly to a matrix or array.  Works in R and NIMBLE.  In
NIMBLE \code{dim} is identical to \code{nimDim}, not to R's
\code{dim}
>>>>>>> upstream/devel:packages/nimble/man/nimDim.Rd
}
\examples{
x <- rnorm(4)
dim(x)
nimDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nimDim(y)
}
\author{
NIMBLE development team
}

