% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nimOptim}
\alias{nimOptim}
\title{Nimble wrapper around R's builtin \code{\link{optim}}.}
\usage{
nimOptim(par, fn)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{fn}{A function to be minimized (or maximized), with first argument the
vector of parameters over which minimization is to take place. It
should return a scalar result.}
}
\value{
\code{\link{optimResultNimbleList}}
}
\description{
Nimble wrapper around R's builtin \code{\link{optim}}.
}
\seealso{
\code{\link{optim}}
}
