% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nimOptim_model}
\alias{nimOptim_model}
\title{Interface for optimizing model logProb with respect to parameters}
\usage{
nimOptim_model(
  model,
  wrt,
  nodes,
  use.gr = TRUE,
  method = "BFGS",
  lower = -Inf,
  upper = Inf,
  control = nimOptimDefaultControl(),
  hessian = FALSE
)
}
\description{
For internal use.
}
