% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nimOptimDefaultControl}
\alias{nimOptimDefaultControl}
\title{Creates a deafult \code{control} argument for \code{\link{nimOptim}}.}
\usage{
nimOptimDefaultControl()
}
\value{
\code{\link{optimControlNimbleList}}
}
\description{
Creates a deafult \code{control} argument for \code{\link{nimOptim}}.
}
\seealso{
\code{\link{nimOptim}}, \code{\link{optim}}
}
