% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in RtmpLwENQY/file6161636d898a
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d64b7a1f83
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497657d3b1a5
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871776a8110
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c347686f848
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b7dc6e865
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba69801dd5
>>>>>>> fixed up some roxygen stuff and
=======
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
>>>>>>> added man files from devel to ease rebase
=======
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
>>>>>>> 6f0bc2e7a0885e03527d78f21be97fad5c54074a
\name{nimPrint}
\alias{nimPrint}
\title{print function for use in nimbleFunctions, where it is identical to \code{print}, but not R's \code{print}.}
\usage{
nimPrint(...)
}
\arguments{
<<<<<<< HEAD
\item{...}{an abitrary set of arguments that will be printed in sequence}}\description{
print function for use in nimbleFunctions, where it is identical to \code{print}.  Works in R or NIMBLE, not quite identically.
}
\details{
The keyword \code{print} in nimbleFunction run-time code will be automatically turned into \code{nimPrint}.  This is a function that prints its arguments int order using \code{cat} in R, or using \code{std::cout} in C++ code generated from compiling nimbleFunctions.
Non-scalar numeric objects can be included, although their output will be formatted slightly different in uncompiled and compiled nimbleFunctions.
=======
  \item{...}{an abitrary set of arguments that will be
  printed in sequence}
}
\description{
print function for use in nimbleFunctions, where it is
identical to \code{print}. Works in R or NIMBLE, not quite
identically.
}
\details{
The keyword \code{print} in nimbleFunction run-time code
will be automatically turned into \code{nimPrint}. This
is a function that prints its arguments int order using
\code{cat} in R, or using \code{std::cout} in C++ code
generated from compiling nimbleFunctions. Non-scalar numeric
objects can be included, although their output will be
formatted slightly different in uncompiled and compiled
nimbleFunctions.
>>>>>>> added man files from devel to ease rebase
}
\examples{
ans <- matrix(1:4, nrow = 2) ## R code, not NIMBLE code
nimPrint('Answer is ', ans, '\\n') ## would work in R or NIMBLE
}

