% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in RtmpLwENQY/file6161636d898a
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d64b7a1f83
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497657d3b1a5
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871776a8110
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c347686f848
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b7dc6e865
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba69801dd5
>>>>>>> fixed up some roxygen stuff and
\name{nimStop}
\alias{nimStop}
\title{Halt execution of a nimbleFunction function method.  Part of the NIMBLE language}
\usage{
nimStop(msg)
}
\arguments{
\item{msg}{Character object to be output as an error message}}\description{
Halt execution of a nimbleFunction function method.  Part of the NIMBLE language
}
\details{
The NIMBLE stop is similar to the native R stop, but it takes only one argument, the error message to be output.  During uncompiled NIMBLE execution, nimStop simply calls R's stop funtion. During compiled execution it calls the error function from the R headers.  stop is an alias for nimStop in the NIMBLE language
}
\author{
Perry de Valpine
}

