% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUGS_nodeInfo.R
\docType{class}
\name{nodeInfoClass-class}
\alias{nodeInfoClass}
\alias{nodeInfoClass-class}
\title{nodeInfoClass contains information for a node}
\description{
nodeInfoClass contains information for a node
}

