% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_implementations.R
\name{Normal-Inverse-Gamma}
\alias{Normal-Inverse-Gamma}
\alias{dnorm_invgamma}
\alias{rnorm_invgamma}
\title{The Joint Normal Inverse Gamma Distribution}
\usage{
dnorm_invgamma(x, mean_location, mean_scale, var_shape, var_scale,
  log = FALSE)

rnorm_invgamma(n = 1, mean_location, mean_scale, var_shape, var_scale)
}
\arguments{
\item{x}{vector of two values, the representing mean and variance of another random variable.}

\item{mean_location}{location parameter for first element of the random vector.}

\item{mean_scale}{scale of the first element of the random vector, must be positive.}

\item{var_shape}{shape of second element of random vector, must be positive.}

\item{var_scale}{scale of second element of random vector, must be positive.}

\item{log}{logical; if TRUE, probability density is returned on the log scale.}

\item{n}{number of observations (only \code{n=1} is handled currently).}
}
\value{
\code{dnorm_invgamma} gives the density and \code{rnorm_invgamma}
generates random deviates.
}
\description{
Density, distribution function, quantile function and random
  generation for the normal-inverse gamma distribution
}
\details{
The normal-inverse gamma distribution generally arises in the context of the conjugate
prior distribution for the two-parameter univariate normal sampling model.
The distribution can be seen as an inverse gamma marginal distribution for the second
element of the bivariate random vector (i.e., the 'variance' element) and (conditionally)
a normal distribution for the first element of the random vector (i.e., the 'mean' element).
The distribution with parameters \code{mean_location} \eqn{=\mu_0},
\code{mean_scale} \eqn{=\kappa_0}, \code{var_shape} \eqn{=\alpha},
\code{var_scale} \eqn{=\beta} has density
\deqn{
  f(\mu, \sigma^2)= \frac{\kappa_0^{1/2}}{(2\pi)^{1/2} (\sigma^2)^{\alpha+3/2}}
\frac{\beta^\alpha}{\Gamma(\alpha)} e^{-\frac{1}{\sigma^2}(\beta + \frac{\kappa_0(\mu-\mu_0)^2}{2})}}
for \eqn{\sigma^2 > 0}, \eqn{\kappa_0 > 0}, \eqn{\alpha > 0} and \eqn{\beta > 0}.
(Here \eqn{\Gamma(\alpha)}{Gamma(a)} is the function implemented by \R's
 \code{\link{gamma}()} and defined in its help.

The distribution is the same as a normal-scaled inverse chi-square with degrees
of freedom equal to \eqn{2\alpha} and scale equal to \eqn{\beta / \alpha}.

See Gelman et al., Section 3.3.
}
\examples{
x <- rnorm_invgamma(1, mean_location = 0, mean_scale = 3, var_shape = 1, var_scale = 1)
dnorm_invgamma(x, mean_location = 0, mean_scale = 3, var_shape = 1, var_scale = 1)
}
\author{
Christopher Paciorek
}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A., and
Rubin, D.B. (2013) \emph{Bayesian Data Analysis}, 3rd ed. Chapman and Hall/CRC.
}
\seealso{
\link{Distributions} for other standard distributions
}

