% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{optimControlNimbleList}
\alias{optimControlNimbleList}
\title{EXPERIMENTAL Data type for the \code{control} parameter of \code{\link{nimOptim}}}
\format{\preformatted{List of 1
 $ new:function (...)  
}}
\usage{
optimControlNimbleList
}
\description{
\code{\link{nimbleList}} definition for the type of \code{\link{nimbleList}} input as the \code{control} parameter
to \code{\link{nimOptim}}. See \code{\link{optim}} for details.
}
\seealso{
\code{\link{optim}}, \code{\link{nimOptim}}
}
\keyword{datasets}

