% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{optimDefaultControl}
\alias{optimDefaultControl}
\title{Creates a deafult \code{control} argument for \code{\link{optim}} (just an empty list).}
\usage{
optimDefaultControl()
}
\value{
an empty \code{list}.
}
\description{
Creates a deafult \code{control} argument for \code{\link{optim}} (just an empty list).
}
\seealso{
\code{\link{nimOptim}}, \code{\link{optim}}
}

