% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTransform.R
\name{parameterTransform}
\alias{parameterTransform}
\title{Automated transformations of model nodes to unconstrained scales}
\usage{
parameterTransform(model, nodes)
}
\description{
ADD DETAILS
}
\author{
Daniel Turek
}
