% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{pow_int}
\alias{pow_int}
\title{Power function for integer-valued exponent}
\usage{
pow_int(a, b)
}
\arguments{
\item{a}{Base}

\item{b}{Exponent}
}
\value{
a^b
}
\description{
pow function with exponent required to be integer
}
\details{
This is required in nimble models and nimbleFunctions if derivatives will be tracked
but tracked only with respect to a, such that b might be any (positive, 0, or negative) integer.
This contrasts with pow(a, b) (equivalent to a^b), which requires b > 0 if derivatives will be
tracked, even if they will only be requested with respect to a.
}
