% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{printErrors}
\alias{printErrors}
\title{Print error messages after failed compilation}
\usage{
printErrors(excludeWarnings = TRUE)
}
\arguments{
\item{excludeWarnings}{logical indicating whether compiler warnings should be printed; generally such warnings can be ignored.}
}
\description{
Retrieves the error file from R's tempdir and prints to the screen.
}
\author{
Christopher Paciorek
}
