% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
=======
% Please edit documentation in RtmpLwENQY/file616162260d5d
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3359ae819
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0b253e8a
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6320f3e98
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497656247499
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78712a8c67
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c3443f77500
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b48daab2c
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba30b2868d
>>>>>>> fixed up some roxygen stuff and
\name{resize}
\alias{resize}
\title{Resizes a modelValues object}
\usage{
resize(container, k)
}
\arguments{
\item{container}{modelValues object}\item{k}{number of rows that modelValues is set to}}\description{
Adds or removes rows to a modelValues object. If rows are added to a modelValues object, the default values are \code{NA}. Works in both R and NIMBLE.
}
\details{
See the User Manual or \code{help(modelValuesBaseClass)} for infomation about modelValues objects
}
\examples{
mvSpec <- modelValuesSpec(vars = c('a', 'b'),
             types = c('double', 'double'),
             sizes = list(a = 1, b = c(2,2) ) )
mv <- modelValues(mvSpec)
as.matrix(mv)
resize(mv, 3)
as.matrix(mv)
}
\author{
Clifford Anderson-Bergman
}

