% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_util.R
\name{Rmatrix2mvOneVar}
\alias{Rmatrix2mvOneVar}
\title{Set values of one variable of a modelValues object from an R matrix}
\usage{
Rmatrix2mvOneVar(mat, mv, varName, k)
}
\arguments{
\item{mat}{Input matrix}

\item{mv}{modelValues object to be modified.}

\item{varName}{Character string giving the name of the variable on \code{mv} to be set}

\item{k}{Number of rows to use}
}
\description{
Normally a modelValues object is accessed one "row" at a time.  This function allows all rows for one variable to set from a matrix with one dimension more than the variable to be set.
}
\details{
This function may be deprecated in the future when a more natural system for interacting with modelValues objects is developed.
}

