% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp8XMcGu/file9e315d93b1d5
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\alias{RmodelBaseClass-class}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

