% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{runLaplace}
\alias{runLaplace}
\title{Placeholder for runLaplace}
\usage{
runLaplace(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleQuad` package.
}
