% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_RJ.R
\name{sampler_RJ_indicator}
\alias{sampler_RJ_indicator}
\title{Reversible Jump sampler}
\usage{
sampler_RJ_indicator(model, mvSaved, target, control)
}
\description{
This sampler perform a Reversible Jump MCMC step for the node to which is assigned, using an univariate normal proposal distribution. This is a specialized sampler used by \code{\link{configureRJ}} function, when the model code is written using indicator variables associated with the nodes of interest. See Details in \code{\link{configureRJ}}.
}
\details{
The RJ sampler accepts the following control list elements:
\itemize{
\item fixedValue. The value taken from the variable when out of the model. (default = 0)
\item mean. The mean of the normal proposal distribution. (default = 0)
\item scale. The standard deviation of the normal proposal distribution. (default  = 1)
\item positive. A logical argument specifying whether the proposal is strictly positive. (default = FALSE)
}
}
\references{
Peter J. Green. (1995). Reversible jump markov chain monte carlo computation and bayesian model determination. \emph{Biometrika}, 82(4), 711-732.
}
\seealso{
\code{\link{sampler_BASE}} \code{\link{sampler_RJ}} \code{\link{configureRJ}}
}
