% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee6563fece
=======
% Please edit documentation in RtmpLwENQY/file616156fdd32a
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3144ee279
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b016902b0f
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d691dbf28
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497627b49973
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78717c9fa53a
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c343148fe6d
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0bd18b92
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba21c84593
>>>>>>> fixed up some roxygen stuff and
\name{setAndCalculate}
\alias{setAndCalculate}
\title{Creates a nimbleFunction for setting the values of one or more model nodes,
calculating the associated deterministic dependents and logProb values,
and returning the total sum log-probability.}
\usage{
setAndCalculate(model, targetNodes)
}
\arguments{
\item{model}{An uncompiled or compiled NIMBLE model.  This argument is required.}\item{targetNodes}{A character vector containing the names of one or more nodes or variables in the model.  This argument is required.}}\description{
This nimbleFunction generator must be specialized to any model object and one or more model nodes.
A specialized instance of this nimbleFunction will set the values of the target nodes in the specified model,
calculate the associated logProbs, calculate the values of any deterministic dependents,
calculate the logProbs of any stochastic dependents,
and return the sum log-probability associated with the target nodes and all stochastic dependent nodes.
}
\details{
Calling setAndCalculate(model, targetNodes) will return a function with a single, required argument:

targetValues: A vector of numeric values which will be put into the target nodes in the specified model object.  The length of this numeric vector much exactly match the number of target nodes.
}
\examples{
my_setAndCalc <- setAndCalculate(Rmodel, c('x[1]', 'x[2]', 'y[1]', 'y[2]'))
lp <- my_setAndCalc(c(1.2, 1.4, 7.6, 8.9))
}
\author{
Daniel Turek
}

