% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee55370303
=======
% Please edit documentation in RtmpLwENQY/file61611b30a085
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb32acd865c
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0439b7288
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d66c1c0846
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976b57e192
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78714dde7fd3
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c343c3f05
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b526e705b
>>>>>>> removed top level tests dir
\name{setSize}
\alias{setSize}
\title{set the size of a numeric variable in NIMBLE}
\usage{
setSize(numObj, ..., row)
}
\arguments{
  \item{numObj}{This is the object to be resized} \item{...}{sizes, provided as scalars, in order, with exactly
  as many as needed for the object} \item{row}{Optional
  argument that is not currently used}
}
\description{
set the size of a numeric variable in NIMBLE. This works in
R and NIMBLE, but in R it usually has no effect.
}
\details{
This function is part of the NIMBLE language. Its purpose
is to explicitly resize a multivariate object (vector,
matrix or array), currently up to 4 dimensions. Explicit
resizing is not needed when an entire object is assigned
to. For example, in \code{Y <- A \%*\% B}, where A and
B are matrices, \code{Y} will be resized automatically.
Explicit resizing is necessary when assignment will be by
indexed elements or blocks, if the object is not already
an appropriate size for the assignment. This does work
in uncompiled (R) and well as compiled execution, but it
is really only necessary for compiled execution. During
uncompiled execution, it may not catch bugs due to resizing
because R objects will be dynamically resized during
assignments anyway.
}
\author{
NIMBLE development team
}

