% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setValues}
\alias{setValues}
\title{Fill a set of nodes in a model from a vector of values}
\usage{
setValues(input, model, nodes)
}
\arguments{
  \item{input}{a vector of values to be put in the model's
  nodes}

  \item{model}{a NIMBLE model object, either compiled or
  uncompiled}

  \item{nodes}{a vector of node names, allowing index
  blocks that will be expanded}
}
\value{
NULL, but this function works by the side-effect of
modifying the model.
}
\description{
Take a vector of node names for a model and fill them
sequentially from the values in a vector.  Works in R and
NIMBLE.
}
\details{
Calling setValues(P, model, nodes) will place values from
P, in order, into the nodes provided for the model.  This
is being deprecated by values(model, nodes) <- P, but the
development syntax of setValues(P, model, nodes) is still
supported.

When provided nodes are from matrices or arrays, the values
will filled following column-wise order.

The reverse of setValues(P, model, nodes) is getValues(P,
model, nodes), which is being deprecated by P <-
values(model, nodes)

These functions work in R and in NIMBLE run-time code that
can be compiled.
}
\author{
NIMBLE development team
}

