% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNP_distributions.R
\name{StickBreakingFunction}
\alias{StickBreakingFunction}
\alias{stick_breaking}
\title{The Stick Breaking Function}
\usage{
stick_breaking(z, log = 0)
}
\arguments{
\item{z}{vector argument.}

\item{log}{logical; if TRUE, weights are returned on the log scale.}
}
\description{
Computes probabilities based on stick breaking construction
}
\details{
The stick breaking function produces a vector of probabilities that adds up to one,
based on a series of individual probabilities in \code{z}, that should be between 
\deqn{(0,1)]}, and define the breaking length of the stick.
The returned vector has length equal to the length of z plus 1. 
If \code{z[k]} is equal to 1, then the returned vector has length smaller than \code{z}. 
If one of the components is smaller than 0 or greater than 1, NaNs are returned.
}
\examples{
z <- rbeta(5, 1, 1)
stick_breaking(z)

cstick_breaking <- compileNimble(stick_breaking)
cstick_breaking(z)
}
\author{
Claudia Wehrhahn
}
\references{
Sethuraman, J. (1994). A constructive definition of Dirichlet priors.
 \emph{Statistica Sinica}, 639-650.
}

