% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{summaryAGHQ}
\alias{summaryAGHQ}
\title{Placeholder for summaryAGHQ}
\usage{
summaryAGHQ(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleQuad` package.
}
