% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{summaryLaplace}
\alias{summaryLaplace}
\title{Placeholder for summaryLaplace}
\usage{
summaryLaplace(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleQuad` package.
}
