% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpPe12xI/file655137583266
\docType{data}
\name{svdNimbleList}
\alias{svdNimbleList}
\title{svdNimbleList definition}
\format{\preformatted{List of 1
 $ new:function (...)  
}}
\usage{
svdNimbleList
}
\description{
\code{nimbleList} definition for the type of \code{nimbleList} returned by \code{\link{nimSvd}}.
}
\author{
NIMBLE development team
}
\seealso{
\code{\link{nimSvd}}
}
\keyword{datasets}

