% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCpp_toEigenize.R
\docType{data}
\name{toEigenCalls}
\alias{toEigenCalls}
\title{[[' = 'sizeDoubleBracket', ## for nimbleFunctionList, this will always  go through chainedCall(nfList[[i]], 'foo')(arg1, arg2) ## NEVER GETS MARKED WITH TOEIGENIZE}
\format{An object of class \code{list} of length 149.}
\usage{
toEigenCalls
}
\description{
[[' = 'sizeDoubleBracket', ## for nimbleFunctionList, this will always  go through chainedCall(nfList[[i]], 'foo')(arg1, arg2) ## NEVER GETS MARKED WITH TOEIGENIZE
}
\keyword{datasets}

