% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in RtmpLwENQY/file6161636d898a
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d64b7a1f83
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497657d3b1a5
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871776a8110
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c347686f848
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b7dc6e865
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba69801dd5
>>>>>>> fixed up some roxygen stuff and
\name{values}
\alias{values}
\alias{values<-}
\title{Access values for a set of nodes in a model}
\usage{
values(model, nodes)
}
\arguments{
\item{model}{a NIMBLE model object, either compiled or uncompiled}\item{nodes}{a vector of node names, allowing index blocks that will be expanded}}\value{
A vector of values concatenated from the provided nodes in the model
}
\description{
Get or set values for a set of nodes in a model
}
\details{
Calling \code{values(model, nodes)} returns a vector of the concatenation of values from the nodes requested
\code{P <- values(model, nodes)} is a newer syntax for \code{getValues(P, model, values)}, which still works and modifies P in the calling environment.

Calling \code{values(model, nodes) <- P} sets the value of the nodes in the model, in sequential order from the vector P.

In both uses, when requested nodes are from matrices or arrays, the values will be handled following column-wise order.

The older function \code{getValues(P, model, nodes)} is equivalent to \code{P <- values(model, nodes)}, and the older function \code{setValues(P, model, nodes)} is equivalent to \code{values(model, nodes) <- P}


These functions work in R and in NIMBLE run-time code that can be compiled.
}
\author{
NIMBLE development team
}

