% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/types_modelValuesAccessor.R
\docType{class}
\name{valuesAccessorVector-class}
\alias{valuesAccessorVector}
\alias{valuesAccessorVector-class}
\title{##    #               var         = 'ANY',character',
##    #               first       = 'ANY',numeric',
##    #               last        = 'ANY',numeric',
##    #               length	  = 'ANY'numeric'}
\description{
##    #               var         = 'ANY',character',
##    #               first       = 'ANY',numeric',
##    #               last        = 'ANY',numeric',
##    #               length	  = 'ANY'numeric'
}

