% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{waicDetailsList}
\alias{waicDetailsList}
\title{waicDetailsList definition}
\format{
An object of class \code{list} of length 1.
}
\usage{
waicDetailsList
}
\description{
\code{waicDetailsList} definition for the \code{nimbleList} type returned by WAIC
computation.
}
\details{
See \code{help(waic)} for details on the elements of the list.
}
\author{
NIMBLE development team
}
\keyword{datasets}
