% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_WAIC.R
\docType{data}
\name{waicList}
\alias{waicList}
\title{waicList definition}
\format{
An object of class \code{list} of length 1.
}
\usage{
waicList
}
\description{
\code{waicList} definition for the \code{nimbleList} type returned by WAIC
computation.
}
\author{
NIMBLE development team
}
\keyword{datasets}
