% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{withNimbleOptions}
\alias{withNimbleOptions}
\title{Temporarily set some NIMBLE options.}
\usage{
withNimbleOptions(options, expr)
}
\arguments{
\item{options}{a list of options suitable for \code{nimbleOptions}.}

\item{expr}{an expression or statement to evaluate.}
}
\value{
expr as evaluated with given options.
}
\description{
Temporarily set some NIMBLE options.
}
\examples{
\dontrun{
if (!(getNimbleOption('showCompilerOutput') == FALSE)) stop()
nf <- nimbleFunction(run = function(){ return(0); returnType(double()) })
cnf <- withNimbleOptions(list(showCompilerOutput = TRUE), {
    if (!(getNimbleOption('showCompilerOutput') == TRUE)) stop()
    compileNimble(nf)
})
if (!(getNimbleOption('showCompilerOutput') == FALSE)) stop()
}
}

