% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppDefs_ADtools.R
\docType{class}
\name{ADproxyModelClass-class}
\alias{ADproxyModelClass-class}
\alias{ADproxyModelClass}
\title{create an ADproxyModelClass object}
\description{
create an ADproxyModelClass object. For internal use.
}
\details{
This is a proxy model for model_AD. The class needs just enough pieces to be used like a model for purposes of nodeFunction compilation. The model will contain an ADproxyModel and then the nodeFunction setup code will extract it. The model interface will population the proxy model's CobjectInterface
}

\author{
NIMBLE development team
}
