% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR.R
\name{as.carAdjacency}
\alias{as.carAdjacency}
\title{Convert CAR structural parameters to adjecency, weights, num format}
\usage{
as.carAdjacency(...)
}
\description{
This will convert alternate representations of CAR process structure into
(adj, weights, num) form required by dcar_normal.  Two alternate
representations are handled:
}
\details{
A single matrix argument will be interpreted as a matrix of symmetric un-normalized weights;

Two lists will be interpreted as (the first) a list of numeric vectors
specifying the adjacency (neighboring) indices of each CAR process component,
and (the second) a list of numeric vectors giving the un-normalized weights
for each of these neighboring relationships.
}
\author{
Daniel Turek
}

