<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoBlock.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb866a88ef1b
>>>>>>> upstream/devel
\name{autoBlock}
\alias{autoBlock}
\title{Automated parameter blocking for efficient MCMC sampling}
\usage{
autoBlock(Rmodel, autoIt = 20000, run = list("all", "default"),
<<<<<<< HEAD
  verbose = FALSE, round = TRUE)
}
\arguments{
\item{Rmodel}{}
=======
  setSeed = TRUE, verbose = FALSE, round = TRUE)
}
\arguments{
  \item{Rmodel}{}
>>>>>>> upstream/devel
}
\value{
info
}
\description{
Automated blocking description
}
\details{
Automated blocking details
}
\examples{
ab <- autoBlock(Rmodel)
}
\author{
Daniel Turek
}

