% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpEe5sYL/file3a481707c571
\name{autoBlock}
\alias{autoBlock}
\title{Automated parameter blocking for efficient MCMC sampling}
\usage{
autoBlock(Rmodel, autoIt = 20000, run = list("all", "default"),
  setSeed = TRUE, verbose = FALSE, makePlots = FALSE, round = TRUE)
}
\arguments{
  \item{Rmodel}{}
}
\value{
info
}
\description{
Automated blocking description
}
\details{
Automated blocking details
}
\examples{
ab <- autoBlock(Rmodel)
}
\author{
Daniel Turek
}

