% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmprA2zEg/file70ee1de39f0
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

