<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUGS_BUGSdecl.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb86105b10c6
>>>>>>> upstream/devel
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
<<<<<<< HEAD
BUGSdeclClass contains the information extracted from one BUGS declaration
=======
BUGSdeclClass contains the information extracted from one
BUGS declaration
>>>>>>> upstream/devel
}

