% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpPe12xI/file655137ba21ba
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

