% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpG2U9tF/file426f45b0f34c
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

