% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp8XMcGu/file9e3117b1ffe4
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

