% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpVY8pk7/file3afa7d2edf70
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

