% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpGOE6Ga/file40312514b41c
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

