% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpkX8ia8/file3b6f3a19c02e
\docType{class}
\name{BUGSdeclClass-class}
\alias{BUGSdeclClass}
\alias{BUGSdeclClass-class}
\title{BUGSdeclClass contains the information extracted from one BUGS declaration}
\description{
BUGSdeclClass contains the information extracted from one BUGS declaration
}

