% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxF_build.R
\name{buildAuxF}
\alias{buildAuxF}
\title{Creates an auxiliary particle filter algorithm to estimate log-likelihood.}
\usage{
buildAuxF(model, nodes, control = list())
}
\arguments{
\item{model}{A nimble model object, typically representing a state space model or a hidden Markov model}

\item{nodes}{A character vector specifying the latent model nodes over which the particle filter will stochastically integrate over to estimate the log-likelihood function}

\item{control}{A list specifying different control options for the particle filter.  Options are described in the details section below.}
}
\description{
Creates an auxiliary particle filter algorithm to estimate log-likelihood.
}
\details{
\describe{
 \item{"thresh"}{ A number between 0 and 1 specifying when to resample: the resampling step will occur when the
  effective sample size is less than thresh*(number of particles).  Defaults to 0.5.}
 \item{"saveAll"}{Indicates whether to save state samples for all time points (T), or only for the most recent time point (F)}
}

     The auxiliary particle filter modifies the bootstrap filter (\code{\link{buildBootF}})
     by adding a lookahead step to
     the algorithm: before propagating particles from one time point to the next via the transition equation,
     the auxiliary filter calculates a weight for each pre-propogated particle by predicting how well
     the particle will agree with the next data point.  These pre-weights are used to conduct an initial
     resampling step before propagation.

 The resulting specialized particle filter algorthm will accept a
 single integer argument (m, default 10,000), which specifies the number
 of random \'particles\' to use for estimating the log-likelihood.  The algorithm
 returns the estimated log-likelihood value, and saves
 unequally weighted samples from the posterior distribution of the latent
 states in the mvWSamp model values object, with corresponding logged weights in mvWSamp['wts',].
 An equally weighted sample from the posterior can be found in mvEWsamp.

  The auxiliary particle filter uses a lookeahead function to select promising particles before propogation.  Currently, the lookahead
  funciton uses the expected valu of the latent state at the next time point given the current particle, e E(x[t+1]|x[t]).
  The auxiliary particle filter currently only works for models with univariate normal transition densities.
}
\examples{
model <- nimbleModel(code = ...)
my_AuxF <- buildAuxF(model, 'x[1:100]', control = list(thresh = 0.9))
Cmodel <- compileNimble(model)
Cmy_AuxF <- compileNimble(my_AuxF, project = model)
logLike <- Cmy_AuxF(m = 100000)
hist(as.matrix(Cmy_Auxf$mvEWSamp, 'x'))
}
\author{
Nicholas Michaud
}
\references{
Pitt, Michael K., and Neil Shephard. "Filtering via simulation: Auxiliary particle filters."
   Journal of the American statistical association 94.446 (1999): 590-599.
}
\seealso{
Other particle filtering methods: \code{\link{buildBootF}};
  \code{\link{buildENKF}}; \code{\link{buildLWF}}
}

