% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{buildAuxiliaryFilter}
\alias{buildAuxiliaryFilter}
\title{Placeholder for buildAuxiliaryFilter}
\usage{
buildAuxiliaryFilter(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleSMC` package
}
