% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{buildBootstrapFilter}
\alias{buildBootstrapFilter}
\title{Placeholder for buildBootstrapFilter}
\usage{
buildBootstrapFilter(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleSMC` package
}
