% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpXPF8d9/file34d06acbde05
\name{buildConjugateSamplerFunctions}
\alias{buildConjugateSamplerFunctions}
\title{Rebuild conjugate sampler functions}
\usage{
buildConjugateSamplerFunctions(writeToFile = NULL)
}
\description{
Rebuild conjugate sampler functions
}

