% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MCMC_conjugacy.R
\name{buildConjugateSamplerFunctions}
\alias{buildConjugateSamplerFunctions}
\title{Rebuild conjugate sampler functions}
\usage{
buildConjugateSamplerFunctions(writeToFile = NULL)
}
\description{
Rebuild conjugate sampler functions
}

