% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{buildIteratedFilter2}
\alias{buildIteratedFilter2}
\title{Placeholder for buildIteratedFilter2}
\usage{
buildIteratedFilter2(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to the `nimbleSMC` package.
}
