% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
\name{checkInterrupt}
\alias{checkInterrupt}
\title{Check for interrupt (e.g. Ctrl-C) during nimbleFunction execution. Part of the NIMBLE language.}
\usage{
checkInterrupt()
}
\description{
Check for interrupt (e.g. Ctrl-C) during nimbleFunction
execution. Part of the NIMBLE language.
}
\details{
During execution of nimbleFunctions that take a long time,
it is nice to occassionally check if the user has entered
an interrupt and bail out of execution if so. This function
does that. During uncompiled nimbleFunction execution,
it does nothing. During compiled execution, it calls
R_checkUserInterrupt() of the R headers.
}
\author{
Perry de Valpine
}

