% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{checkInterrupt}
\alias{checkInterrupt}
\title{Check for interrupt (e.g. Ctrl-C) during nimbleFunction execution. Part of the NIMBLE language.}
\usage{
checkInterrupt()
}
\description{
Check for interrupt (e.g. Ctrl-C) during nimbleFunction execution. Part of the NIMBLE language.
}
\details{
During execution of nimbleFunctions that take a long time, it is nice to occassionally check if the user has entered an interrupt and bail out of execution if so.  This function does that.  During uncompiled nimbleFunction execution, it does nothing.  During compiled execution, it calls R_checkUserInterrupt() of the R headers.
}
\author{
Perry de Valpine
}

