% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp6hGOZh/file12a1f1a1bdbb9
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

