% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp9zTnuX/file4e2374d7c4a
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

