% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d5fd5d3d0
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

