% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpNY1hDh/file4cf24d711f98
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

