% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpK6dyXN/file1e095fed3b47
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

