% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpyIO6m7/file1cfe2d520c42
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

