% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpowSbmr/filed7685f5d6c66
\docType{class}
\name{CmodelBaseClass-class}
\alias{CmodelBaseClass}
\alias{CmodelBaseClass-class}
\title{Class \code{CmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

