% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp8XMcGu/file9e312bd6d276
\docType{class}
\name{CnimbleFunctionBase-class}
\alias{CnimbleFunctionBase}
\alias{CnimbleFunctionBase-class}
\title{Class \code{CnimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

