% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpOWZhCV/file24f82962196c
\docType{class}
\name{CnimbleFunctionBase-class}
\alias{CnimbleFunctionBase}
\alias{CnimbleFunctionBase-class}
\title{Class \code{CnimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

