% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpG2U9tF/file426f4a09d0c0
\docType{class}
\name{CnimbleFunctionBase-class}
\alias{CnimbleFunctionBase}
\alias{CnimbleFunctionBase-class}
\title{Class \code{CnimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

