% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmph6b7yW/file2612128c25a3
\docType{class}
\name{codeBlockClass-class}
\alias{codeBlockClass}
\alias{codeBlockClass-class}
\title{Class \code{codeBlockClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

