% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_utils.R
\docType{class}
\name{codeBlockClass-class}
\alias{codeBlockClass-class}
\alias{codeBlockClass}
\title{Class \code{codeBlockClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

