% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_comparisons.R
\name{combine_MCMC_comparison_results}
\alias{combine_MCMC_comparison_results}
\title{Combine multiple objects returned by compareMCMCs}
\usage{
combine_MCMC_comparison_results(..., name = "MCMCresults")
}
\arguments{
\item{...}{MCMC comparison results for an individual model.  These results can be obtained from the objects returned by \code{\link{compareMCMCs}}.}

\item{name}{(default "MCMCresults") name to be given to the resulting set of comparisons.  This will be used in the html generated by \code{\link{make_MCMC_comparison_pages}}.  It is simply a list name and thus can easily be modified later.}
}
\value{
An object in the same format as returned by \code{\link{compareMCMCs}} with \code{summary} = TRUE.
}
\description{
Useful for running different cases separately and combining them.
}
\examples{
\dontrun{
nimbleMCMCresults <- compareMCMCs(
 modelInfo = list(
   code = code,
   data = data,
   constants = constants,
   inits = inits),
 MCMCs = c('nimble', 'nimble_slice'),
 summary = TRUE
)

jagsMCMCresults <- compareMCMCs(
 modelInfo = list(
   code = code,
   data = data,
   constants = constants,
   inits = inits),
 MCMCs = c('jags'),
 summary = TRUE
)

combinedMCMCresults <- combine_MCMC_comparison_results(nimbleMCMCresults[['model1']],
                                                       jagsMCMCresults[['model1']],
                                                       name = 'combinedMCMCresults')
}

}
\seealso{
\code{\link{compareMCMCs}}, \code{\link{rename_MCMC_comparison_method}}, \code{\link{make_MCMC_comparison_pages}}, \code{\link{reshape_comparison_results}}.
}
