% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{compareMCMCs}
\alias{compareMCMCs}
\title{Placeholder for compareMCMCs}
\usage{
compareMCMCs(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function has been moved to a separate package
}
