% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_RJ.R
\name{configureRJ}
\alias{configureRJ}
\title{Configure reversible jump sampler}
\usage{
configureRJ(mcmcConf, nodes, indicator = NULL, prior = NULL,
  control_RJ = list(fixedValue = NULL, mean = NULL, scale = NULL,
  positive = NULL))
}
\arguments{
\item{mcmcConf}{a \code{MCMCconf} object.}

\item{nodes}{a character vector, containing the names of the nodes for which the variable selection is performed.}

\item{indicator}{a character vector, containing the names of the indicator variables associated with \code{nodes} that are involved in the reversible jump step. (see details?)}

\item{prior}{a vector of prior probabilities for each node to be equal to 0 (or another fixed value) or not. (see details?)}

\item{control_RJ}{named list with arguments
\itemize{
\item fixedValue. The value taken from the variable when out of the model. (default = 0)
\item mean. The mean of the normal proposal distribution. (default = 0)
\item scale. The standard deviation of the normal proposal distribution. (default  = 1)
\item positive. A logical argument specifying whether the proposal is strictly positive. (default = FALSE)
}}
}
\description{
Modifies the \code{MCMCconf} object of a specific model, to include a Reversible Jump MCMC sampler for variable selection using an univariate normal proposal distribution. User can control which value the variable takes when not in the model (typically 0), the mean and scale of the proposal, and whether or not the proposal is strictly positive. This function supports two different ways of writing the model.
}
\details{
This functions modifies the sampler in \code{MCMCconf} for each of the nodes provided in the \code{nodes} argument. To these elements two samplers are assigned: a specialized Reversible Jump sampler and a modified version of the build sampler that is used only when the target node is in the model. 

The specialized RJ sampler depends on whether the \code{prior} or \code{indicator} arguments are provided, according on how the model is written. When \code{prior} is provided, the specialized sampler \code{sampler_RJ} is assigned directly to the \code{nodes}. When \code{indicator} is provided, the specialized sampler \code{sampler_RJ_indicator} is assigned to nodes in \code{indicator}.
}
\examples{


}
\references{
Peter J. Green. (1995). Reversible jump markov chain monte carlo computation and bayesian model determination. \emph{Biometrika}, 82(4), 711-732.
}
\seealso{
\code{\link{sampler_BASE}} \code{\link{sampler_RJ}} \code{\link{configureRJ}}
}
