% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpTGWyjf/file1d683c0104d5
\name{copierVector}
\alias{copierVector}
\title{##    #               var         = 'ANY',character',
##    #               first       = 'ANY',numeric',
##    #               last        = 'ANY',numeric',
##    #               length	  = 'ANY'numeric'}
\usage{
copierVector(accessFrom_name, accessTo_name, isFromMV, isToMV)
}
\description{
## # var = 'ANY',character', ## # first = 'ANY',numeric', ##
# last = 'ANY',numeric', ## # length = 'ANY'numeric'
}

