% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmp3xd4YE/filefb26f27dd62
\docType{class}
\name{cppVar-class}
\alias{cppVar}
\alias{cppVar-class}
\title{##                           templateArgs = 'ANY',list',
##                           baseScope = 'ANY',list',
##                           baseConst = 'ANY',logical',
##                           baseConstPtr = 'ANY',numeric',
##                           static = 'ANY',logical',
##                           arraySizes = 'ANY',integer',
##                           constructor = 'ANY',character',
##                           selfDereference = 'ANY'logical'}
\description{
##                           templateArgs = 'ANY',list',
##                           baseScope = 'ANY',list',
##                           baseConst = 'ANY',logical',
##                           baseConstPtr = 'ANY',numeric',
##                           static = 'ANY',logical',
##                           arraySizes = 'ANY',integer',
##                           constructor = 'ANY',character',
##                           selfDereference = 'ANY'logical'
}

