% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpK6dyXN/file1e0968b466c9
\name{deregisterDistributions}
\alias{deregisterDistributions}
\title{Remove user-supplied distributions from use in NIMBLE BUGS models}
\usage{
deregisterDistributions(distributionsNames)
}
\arguments{
\item{distributionsNames}{a character vector giving the names of the distributions to be dergistered}
}
\description{
Deregister distributional information originally supplied by the user
for use in BUGS model code
}
\author{
Christopher Paciorek
}

