<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions_processInputList.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb86224967af
>>>>>>> upstream/devel
\name{deregisterDistributions}
\alias{deregisterDistributions}
\title{Remove user-supplied distributions from use in NIMBLE BUGS models}
\usage{
deregisterDistributions(distributionsNames)
}
\arguments{
<<<<<<< HEAD
\item{distributionsNames}{a character vector giving the names of the distributions to be dergistered}
}
\description{
Deregister distributional information originally supplied by the user
for use in BUGS model code
=======
  \item{distributionsNames}{a character vector giving the
  names of the distributions to be dergistered}
}
\description{
Deregister distributional information originally supplied
by the user for use in BUGS model code
>>>>>>> upstream/devel
}
\author{
Christopher Paciorek
}

