% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_derived.R
\name{derived_BASE}
\alias{derived_BASE}
\alias{derived_mean}
\alias{derived_variance}
\alias{derived_logProb}
\alias{derived_predictive}
\alias{derived}
\title{MCMC Derived Quantities}
\usage{
derived_BASE()

derived_mean(model, mcmc, interval, control)

derived_variance(model, mcmc, interval, control)

derived_logProb(model, mcmc, interval, control)

derived_predictive(model, mcmc, interval, control)
}
\description{
Details of the NIMBLE MCMC engine handles derived quantities, which are deterministic functions that can be calaculated and recorded after each MCMC sampling iteration.
}
\examples{
conf$addDerivedQuantity("mean", nodes = c("a", "b"))

conf$addDerivedQuantity("mean", nodes = "theta", interval = 5)

conf$addDerivedQuantity("variance", nodes = "x[1:4]", control = list(recordingFrequency = 10))

conf$addDerivedQuantity("logProb", nodes = c('alpha', 'beta'))

conf <- configureMCMC(model, mean = 'a', variance = 'b', logProb = TRUE)

}
\seealso{
\code{\link{configureMCMC}} \code{\link{addDerivedQuantity}} \code{\link{buildMCMC}} \code{\link{runMCMC}} \code{\link{nimbleMCMC}}
}
\author{
Daniel Turek
}
