% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpPe12xI/file655137583266
\docType{data}
\name{eigenNimbleList}
\alias{eigenNimbleList}
\title{eigenNimbleList definition}
\format{\preformatted{List of 1
 $ new:function (...)  
}}
\usage{
eigenNimbleList
}
\description{
\code{nimbleList} definition for the type of \code{nimbleList} returned by \code{\link{nimEigen}}.
}
\author{
NIMBLE development team
}
\seealso{
\code{\link{nimEigen}}
}
\keyword{datasets}

