% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{eigenNimbleList}
\alias{eigenNimbleList}
\title{eigenNimbleList definition}
\format{An object of class \code{list} of length 1.}
\usage{
eigenNimbleList
}
\description{
\code{nimbleList} definition for the type of \code{nimbleList} returned by \code{\link{nimEigen}}.
}
\seealso{
\code{\link{nimEigen}}
}
\author{
NIMBLE development team
}
\keyword{datasets}
