% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGS_readBUGS.R
\name{getBUGSexampleDir}
\alias{getBUGSexampleDir}
\title{Get the directory path to one of the classic BUGS examples installed with NIMBLE package}
\usage{
getBUGSexampleDir(example)
}
\arguments{
\item{example}{The name of the classic BUGS example.}
}
\value{
Character string of the fully pathed directory of the BUGS example.
}
\description{
NIMBLE comes with some of the classic BUGS examples.  \code{getBUGSexampleDir} looks up the location of an example from its name.
}
\author{
Christopher Paciorek
}
\seealso{
\code{\link{readBUGSmodel}} for usage in creating a model from a classic BUGS example
}

