<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb86f7ff86e
>>>>>>> upstream/devel
\name{getNimbleOption}
\alias{getNimbleOption}
\title{Get NIMBLE Option}
\usage{
getNimbleOption(x)
}
\arguments{
<<<<<<< HEAD
\item{x}{a character string holding an option name}
=======
  \item{x}{a character string holding an option name}
>>>>>>> upstream/devel
}
\value{
The value of the option.
}
\description{
<<<<<<< HEAD
Allow the user to get the value of a global _option_
that affects the way in which NIMBLE operates
=======
Allow the user to get the value of a global _option_ that
affects the way in which NIMBLE operates
>>>>>>> upstream/devel
}
\examples{
getNimbleOption('verifyConjugatePosteriors')
}
\author{
Christopher Paciorek
}

