% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee1e8d0966
=======
% Please edit documentation in RtmpLwENQY/file61612b823b49
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb35b4fa7ed
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b079b14439
>>>>>>> took Makefiles out of
\name{getNimbleOption}
\alias{getNimbleOption}
\title{Get NIMBLE Option}
\usage{
getNimbleOption(x)
}
\arguments{
  \item{x}{a character string holding an option name}}\value{
The value of the option.
}
\description{
Allow the user to get the value of a global _option_ that
affects the way in which NIMBLE operates
}
\examples{
getNimbleOption('verifyConjugatePosteriors')
}
\author{
Christopher Paciorek
}

