% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee1e8d0966
=======
% Please edit documentation in RtmpLwENQY/file61612b823b49
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb35b4fa7ed
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b079b14439
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6518728f4
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file49762eb6cc9
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78717038320b
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c344ed94e0b
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b6904a114
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba166dca49
>>>>>>> fixed up some roxygen stuff and
\name{getNimbleOption}
\alias{getNimbleOption}
\title{Get NIMBLE Option}
\usage{
getNimbleOption(x)
}
\arguments{
\item{x}{a character string holding an option name}}\value{
The value of the option.
}
\description{
Allow the user to get the value of a global _option_
that affects the way in which NIMBLE operates
}
\examples{
getNimbleOption('verifyConjugatePosteriors')
}
\author{
Christopher Paciorek
}

