% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getNimbleOption}
\alias{getNimbleOption}
\title{Get NIMBLE Option}
\usage{
getNimbleOption(x)
}
\arguments{
\item{x}{a character string holding an option name}
}
\value{
The value of the option.
}
\description{
Allow the user to get the value of a global _option_
that affects the way in which NIMBLE operates
}
\examples{
getNimbleOption('verifyConjugatePosteriors')
}
\author{
Christopher Paciorek
}

