% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmp60KOwh/file78717f47405e
\name{getNimbleProject}
\alias{getNimbleProject}
\title{return the nimble project, if any, associated with a model or nimbleFunction object.}
\usage{
getNimbleProject(project, stopOnNull = FALSE)
}
\description{
return the nimble project, if any, associated with a model
or nimbleFunction object.
}

