% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee38fefa33
=======
% Please edit documentation in RtmpLwENQY/file61611adbe820
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb319d6ed7e
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b070e06ec3
>>>>>>> took Makefiles out of
\name{getNimbleProject}
\alias{getNimbleProject}
\title{return the nimble project, if any, associated with a model or nimbleFunction object.}
\usage{
getNimbleProject(project, stopOnNull = FALSE)
}
\description{
return the nimble project, if any, associated with a model
or nimbleFunction object.
}

