% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee38fefa33
=======
% Please edit documentation in RtmpLwENQY/file61611adbe820
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb319d6ed7e
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b070e06ec3
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d611be8f
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file49766c690e7
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78717f47405e
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c344f811986
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in Rtmpzdod9z/file79ee38fefa33
>>>>>>> added man files from devel to ease rebase
=======
% Please edit documentation in Rtmpzdod9z/file79ee38fefa33
>>>>>>> 6f0bc2e7a0885e03527d78f21be97fad5c54074a
\name{getNimbleProject}
\alias{getNimbleProject}
\title{return the nimble project, if any, associated with a model or nimbleFunction object.}
\usage{
getNimbleProject(project, stopOnNull = FALSE)
}
\description{
return the nimble project, if any, associated with a model
or nimbleFunction object.
}

