% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb867b867383
\name{getNimbleProject}
\alias{getNimbleProject}
\title{return the nimble project, if any, associated with a model or nimbleFunction object.}
\usage{
getNimbleProject(project, stopOnNull = FALSE)
}
\description{
return the nimble project, if any, associated with a model
or nimbleFunction object.
}

