% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/all_utils.R
=======
% Please edit documentation in RtmpgVE0pU/file7ba03a6f273b
>>>>>>> upstream/devel
\name{getsize}
\alias{getsize}
\title{Returns number of rows of modelValues}
\usage{
getsize(container)
}
\arguments{
\item{container}{modelValues object}
}
\description{
Returns the number of rows of NIMBLE modelValues object. Works in R and NIMBLE.
}
\details{
<<<<<<< HEAD
See the User Manual or \code{help(modelValuesBaseClass)} for information about modelValues objects
=======
See the User Manual or \code{help(modelValuesBaseClass)} for
information about modelValues objects
>>>>>>> upstream/devel
}
\examples{
mvSpec <- modelValuesSpec(vars = 'a', types = 'double', sizes = list(a = 1) )
	mv <- modelValues(mvSpec)
 resize(mv, 10)
	getsize(mv)
}
\author{
Clifford Anderson-Bergman
}

