% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
=======
% Please edit documentation in RtmpLwENQY/file616162260d5d
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3359ae819
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0b253e8a
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6320f3e98
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497656247499
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78712a8c67
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c3443f77500
>>>>>>> various cleanups of roxygen stuff
\name{getsize}
\alias{getsize}
\title{Returns number of rows of modelValues}
\usage{
getsize(container)
}
\arguments{
  \item{container}{modelValues object}}\description{
Returns the number of rows of NIMBLE modelValues object.
Works in R and NIMBLE.
}
\details{
See the User Manual or \code{help(modelValuesBaseClass)} for
information about modelValues objects
}
\examples{
mvSpec <- modelValuesSpec(vars = 'a', types = 'double', sizes = list(a = 1) )
	mv <- modelValues(mvSpec)
 resize(mv, 10)
	getsize(mv)
}
\author{
Clifford Anderson-Bergman
}

