% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{getsize}
\alias{getsize}
\title{Returns number of rows of modelValues}
\usage{
getsize(container)
}
\arguments{
\item{container}{modelValues object}
}
\description{
Returns the number of rows of NIMBLE modelValues object. Works in R and NIMBLE.
}
\details{
See the \href{https://r-nimble.org/html_manual/cha-welcome-nimble.html}{User Manual} or \code{help(modelValuesBaseClass)} for information about modelValues objects
}
\examples{
mvConf <- modelValuesConf(vars = 'a', types = 'double', sizes = list(a = 1) )
mv <- modelValues(mvConf)
 resize(mv, 10)
getsize(mv)
}
\author{
Clifford Anderson-Bergman
}
