% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee3ffd4de0
=======
% Please edit documentation in RtmpLwENQY/file616175a67bec
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb36c4eadde
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0651fe692
>>>>>>> took Makefiles out of
\name{hasBracket}
\alias{hasBracket}
\title{Determine if a piece of code contains a '['}
\usage{
hasBracket(code)
}
\description{
Determine if a piece of code contains a '['
}

