<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUGS_utils.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb867af80002
>>>>>>> upstream/devel
\name{hasBracket}
\alias{hasBracket}
\title{Determine if a piece of code contains a '['}
\usage{
hasBracket(code)
}
\description{
Determine if a piece of code contains a '['
}

