% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_utils.R
\name{identityMatrix}
\alias{identityMatrix}
\title{Create an Identity matrix (Deprecated)}
\usage{
identityMatrix(d)
}
\arguments{
\item{d}{The size of the identity matrix to return, will return a d-by-d matrix}
}
\value{
A d-by-d identity matrix
}
\description{
Returns a d-by-d identity matrix (square matrix of 0's, with 1's on the main diagnol).
}
\details{
This function can be used in NIMBLE run code.  It is deprecated because now one can use diag(d) instead.
}
\examples{
Id <- identityMatrix(d = 3)

}
\author{
Daniel Turek
}
