% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpiobOoV/file1055162300f0
\name{is.nf}
\alias{is.nf}
\title{check if a nimbleFunction}
\usage{
is.nf(f)
}
\arguments{
\item{f}{object to be tested}
}
\description{
Checks an object to determine if it is a nimbleFunction (i.e., a function created by \code{nimbleFunction} using setup code).
}
\seealso{
\link{nimbleFunction} for how to create a nimbleFunction
}

