% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpQKPuOQ/file530065734105
\name{is.nf}
\alias{is.nf}
\title{check if a nimbleFunction}
\usage{
is.nf(f, inputIsName = FALSE)
}
\arguments{
\item{f}{object to be tested}

\item{inputIsName}{logical indicating whether the function is provided as the character name of the function or the function object itself}
}
\description{
Checks an object to determine if it is a nimbleFunction (i.e., a function created by \code{nimbleFunction} using setup code).
}
\seealso{
\link{nimbleFunction} for how to create a nimbleFunction
}

