% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_util.R
\name{is.nf}
\alias{is.nf}
\title{check if a nimbleFunction}
\usage{
is.nf(f, inputIsName = FALSE)
}
\arguments{
\item{f}{object to be tested}

\item{inputIsName}{logical indicating whether the function is provided as the character name of the function or the function object itself}
}
\description{
Checks an object to determine if it is a nimbleFunction (i.e., a function created by \code{nimbleFunction} using setup code).
}
\seealso{
\code{\link{nimbleFunction}} for how to create a nimbleFunction
}

