% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpiobOoV/file10552850a4cf
\name{makeParamInfo}
\alias{makeParamInfo}
\title{Make an object of information about a model-parameter pairing for getParam.  Used internally}
\usage{
makeParamInfo(model, nodes, param)
}
\arguments{
\item{model}{A model such as returned by \code{\link{nimbleModel}}.}

\item{nodes}{A character string naming one or more stochastic nodes, such as "mu", "c('mu', 'beta[2]')", or "eta[1:3, 2]"}

\item{param}{A character string naming a parameter of the distribution followed by node, such as "mean", "rate", "lambda", or whatever parameter names are relevant for the distribution of the node.}
}
\description{
Creates a simple getParam_info object, which has a list with a paramID and a type
}
\details{
This is used internally by \code{\link{getParam}}.  It is not intended for direct use by a user or even a nimbleFunction programmer.
}

