<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUGS_modelDef.R
=======
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmpCszhwx/fileb8663a68294
>>>>>>> upstream/devel
\docType{class}
\name{modelDefClass-class}
\alias{modelDefClass}
\alias{modelDefClass-class}
\title{Class for NIMBLE model definition}
\description{
<<<<<<< HEAD
Class for NIMBLE model definition that is not usually needed directly by a user.
}
\details{
See \code{?modelBaseClass} for information about creating NIMBLE BUGS models.
=======
Class for NIMBLE model definition that is not usually
needed directly by a user.
}
\details{
See \code{?modelBaseClass} for information about creating
NIMBLE BUGS models.
>>>>>>> upstream/devel
}

