% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee3c7d894b
=======
% Please edit documentation in RtmpLwENQY/file6161313dcb78
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb31e5ccde2
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b015ffe346
>>>>>>> took Makefiles out of
\docType{class}
\name{modelDefClass-class}
\alias{modelDefClass}
\alias{modelDefClass-class}
\title{Class for NIMBLE model definition}
\description{
Class for NIMBLE model definition that is not usually needed
directly by a user.
}
\details{
See \code{?modelBaseClass} for information about creating
NIMBLE BUGS models.
}

