% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee3c7d894b
=======
% Please edit documentation in RtmpLwENQY/file6161313dcb78
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb31e5ccde2
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b015ffe346
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6267e0090
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file49767285222b
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871445dfafa
>>>>>>> added line to remove .o files after create
\docType{class}
\name{modelDefClass-class}
\alias{modelDefClass}
\alias{modelDefClass-class}
\title{Class for NIMBLE model definition}
\description{
Class for NIMBLE model definition that is not usually needed
directly by a user.
}
\details{
See \code{?modelBaseClass} for information about creating
NIMBLE BUGS models.
}

