% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpSlIkmN/file23692d9d3bcd
\docType{class}
\name{modelDefClass-class}
\alias{modelDefClass}
\alias{modelDefClass-class}
\title{Class for NIMBLE model definition}
\description{
Class for NIMBLE model definition that is not usually needed
directly by a user.
}
\details{
See \code{?modelBaseClass} for information about creating
NIMBLE BUGS models.
}

