% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGS_modelDef.R
\docType{class}
\name{modelDefClass-class}
\alias{modelDefClass-class}
\alias{modelDefClass}
\title{Class for NIMBLE model definition}
\description{
Class for NIMBLE model definition that is not usually needed directly by a user.
}
\details{
See \code{\link{modelBaseClass}} for information about creating NIMBLE BUGS models.
}

