% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee18dea6c8
=======
% Please edit documentation in RtmpLwENQY/file61613d8c344d
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3b738736
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b03a3987b5
>>>>>>> took Makefiles out of
\name{modelValues}
\alias{modelValues}
\title{Create a NIMBLE modelValues Object}
\usage{
modelValues(spec, m = 1)
}
\arguments{
  \item{spec}{An object which includes information for
  building modelValues. Can either be a NIMBLE model (see
  \code{help(modelBaseClass)}) or the object returned from
  \code{modelValuesSpec}} \item{m}{The number of rows to
  create in the modelValues object. Can later be changed
  with \code{resize}}
}
\description{
Builds modelValues object from a model values specification
object, which can include a NIMBLE model
}
\details{
See the User Manual or \code{help(modelValuesBaseClass)}
for information about manipulating NIMBLE modelValues object
returned by this function
}
\examples{
#From model object:
code <- nimbleCode({
 a ~ dnorm(0,1)
 for(i in 1:3){
		for(j in 1:3)
			b[i,j] ~ dnorm(0,1)
		}
})
Rmodel <- nimbleModel(code)
Rmodel_mv <- modelValues(Rmodel, m = 2)
	#Custom modelValues object:
mvSpec <- modelValuesSpec(vars = c('x', 'y'),
             types = c('double', 'int'),
             sizes = list(x = 3, y = c(2,2)))
custom_mv <- modelValues(mvSpec, m = 2)
custom_mv['y',]
}
\author{
NIMBLE development team
}

