% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpKDzmR2/file706c710b7e3e
\name{nfMethod}
\alias{nfMethod}
\title{access a member function (or method) of a nimbleFunction}
\usage{
nfMethod(nf, methodName)
}
\arguments{
  \item{nf}{a specialized nimbleFunction, i.e. one that has
  already had setup parameters processed} \item{methodName}
  {a character string giving the name of the member function
  to call}
}
\value{
a function that can be called.
}
\description{
access (call) a member function of a nimbleFunction other
than \code{run}, which calling the nimbleFunction directly
uses. Works in R and NIMBLE.
}
\details{
nimbleFunctions have a default member function called
\code{run}, and may have other member functions provided
via the \code{methods} argument to \code{nimbleFunction}.
\code{nfMethod} is primarily used to call the latter,
although it could be used with \code{methodName = 'run'}.
Normally arguments will be provided after \code{nfMethod},
e.g. \code{nfMethod(myNimbleFunction, 'foo')(x)}.
}
\author{
NIMBLE development team
}

