% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee4d5b9a12
=======
% Please edit documentation in RtmpLwENQY/file6161636d898a
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb367cbd951
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f515226
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d64b7a1f83
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497657d3b1a5
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file7871776a8110
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c347686f848
>>>>>>> various cleanups of roxygen stuff
\name{nfMethod}
\alias{nfMethod}
\title{Internal function for accessing a member function (method) of a nimbleFunction.  Normally a user will write \code{nf$method(x)} instead of \code{nfMethod(nf, method)(x)}.}
\usage{
nfMethod(nf, methodName)
}
\arguments{
  \item{nf}{a specialized nimbleFunction, i.e. one that has
  already had setup parameters processed} \item{methodName}{a character string giving the name of the member function
  to call}
}
\value{
a function that can be called.
}
\description{
access (call) a member function of a nimbleFunction,
including \code{run}.
}
\details{
nimbleFunctions have a default member function called
\code{run}, and may have other member functions provided via
the \code{methods} argument to \code{nimbleFunction}. As an
internal step, the NIMBLE compiler turns \code{nf$method(x)}
into \code{nfMethod(nf, method)(x)}, but a NIMBLE user or
programmer would not normally need to use \code{nfMethod}
directly.
}
\author{
NIMBLE development team
}

