% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmphInSZr/filecb5b764d5a4c
\name{nfMethod}
\alias{nfMethod}
\title{Internal function for accessing a member function (method) of a nimbleFunction.  Normally a user will write \code{nf$method(x)} instead of \code{nfMethod(nf, method)(x)}.}
\usage{
nfMethod(nf, methodName)
}
\arguments{
\item{nf}{a specialized nimbleFunction, i.e. one that has already had setup parameters processed}

\item{methodName}{a character string giving the name of the member function to call}
}
\value{
a function that can be called.
}
\description{
access (call) a member function of a nimbleFunction, including \code{run}.
}
\details{
nimbleFunctions have a default member function called \code{run}, and may have other member functions provided via the \code{methods} argument to \code{nimbleFunction}.
As an internal step, the NIMBLE compiler turns \code{nf$method(x)} into \code{nfMethod(nf, method)(x)}, but a NIMBLE user or programmer would not normally need to use \code{nfMethod} directly.
}
\author{
NIMBLE development team
}

