% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d45fb9821
\name{nimArray}
\alias{array}
\alias{nimArray}
\title{creates an array}
\usage{
nimArray(value = 0, dim = c(1, 1), init = TRUE, type = "double")
}
\arguments{
\item{value}{the initial value for each element of the array (default = 0)}

\item{dim}{a vector specifying the dimensionality and sizes of the array, provided as c(size1, ...) (default = c(1, 1))}

\item{init}{logical, whether to initialize elements of the matrix (default = TRUE)}

\item{type}{character representing the data type, i.e. 'double' or 'integer' (default = 'double')}
}
\description{
creates an array
}
\details{
When used in a \code{nimbleFunction} (in \code{run} or other member function), \code{array} is a synonym for \code{nimArray}.  When used with only the first two arguments, this behaves similarly to R's \code{array} function.  NIMBLE provides additional arguments to control the initialization value, whether or not initialization will occur, and the type of scalar elements.  Using \code{init=FALSE} when initialization is not necessary can make compiled nimbleFunctions a bit faster.
}
\author{
Daniel Turek
}
\seealso{
\link{numeric} \link{integer} \link{matrix}
}

