% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpowSbmr/filed7686395a88c
\name{nimArray}
\alias{array}
\alias{nimArray}
\title{creates an array}
\usage{
nimArray(value = 0, dim = c(1, 1), init = TRUE, type = "double")
}
\arguments{
\item{value}{the initial value for each element of the array (default = 0)}

\item{dim}{a vector specifying the dimensionality and sizes of the array, provided as c(size1, ...) (default = c(1, 1))}

\item{init}{logical, whether to initialize elements of the matrix (default = TRUE)}

\item{type}{character representing the data type, i.e. 'double' or 'integer' (default = 'double')}
}
\description{
creates an array
}
\author{
Daniel Turek
}
\seealso{
\link{numeric} \link{integer} \link{matrix}
}

