\name{nimble-internal}
\title{Functions and Classes Internal to NIMBLE}
\alias{-Class}
\alias{[,numberedModelValuesAccessors-method}
\alias{[<-,numberedModelValuesAccessors-method}
\alias{[,numberedObjects-method}
\alias{[<-,numberedObjects-method}
\alias{[[,CNumericList-method}
\alias{[[<-,CNumericList-method}
\alias{[[,RNumericList-method}
\alias{[[<-,RNumericList-method}
\alias{[[,nimPointerList-method}
\alias{[[<-,nimPointerList-method}
\alias{[[<-,nimbleFunctionList-method}
\alias{[,distributionsClass-method}
\alias{[[,distributionsClass-method}
\alias{[[,conjugacyRelationshipsClass-method}
\alias{dsqrtinvgamma}
\alias{rsqrtinvgamma}
\alias{calc_dmnormConjugacyContributions}
\alias{calc_dmnormAltParams}
\alias{calc_dwishAltParams}
\alias{getNodeFunctionIndexedInfo}
\alias{singleModelValuesAccess}
\alias{getNimbleProject}
\alias{nimbleInternalFunctions}
\alias{nimbleUserNamespace}
\description{
Functions and classes used internally in NIMBLE and not expected to be called directly by users. Some functions and classes not intended for direct use are documented and/or exported because they are used within Reference Class methods for classes programmatically generated by NIMBLE.
}
\author{NIMBLE Development Team}