\name{nimble-internal}
\title{Functions and Classes Internal to NIMBLE}
\alias{as.matrix.modelValuesBaseClass-Class}
\alias{autoBlockClass-Class}
\alias{BUGScontextClass-Class}
\alias{BUGSsingleContextClass-Class}
\alias{CmultiNimbleFunctionClass-Class}
\alias{conjugacyClass-Class}
\alias{conjugacyRelationshipsClass-Class}
\alias{cppBUGSmodelClass-Class}
\alias{cppCodeFileClass-Class}
\alias{cppCPPfileClass-Class}
\alias{cppHfileClass-Class}
\alias{cppModelValuesClass-Class}
\alias{cppNamedObjectsClass-Class}
\alias{cppNimbleFunctionClass-Class}
\alias{cppProjectClass-Class}
\alias{cppVirtualNimbleFunctionClass-Class}
\alias{dependentClass-Class}
\alias{distClass-Class}
\alias{distributionsClass-Class}
\alias{exprClass-Class}
\alias{exprTypeInfoClass-Class}
\alias{indexedNodeInfoTableClass-Class}
\alias{keywordInfoClass-Class}
\alias{MakeCustomModelClass-Class}
\alias{makeCustomModelValuesClass-Class}
\alias{mapsClass-Class}
\alias{modelDefInfoClass-Class}
\alias{mvInfoClass-Class}
\alias{nfCompilationInfoClass-Class}
\alias{nimbleGraphClass-Class}
\alias{nimbleProjectClass-Class}
\alias{posteriorClass-Class}
\alias{RCfunctionCompileClass-Class}
\alias{RCfunInfoClass-Class}
\alias{RMakeCustomModelClass-Class}
\alias{setupCodeTemplateClass-Class}
\alias{singleModelValuesAccessClass-Class}
\alias{varInfoClass-Class}
\alias{[,numberedModelValuesAccessors-method}
\alias{[<-,numberedModelValuesAccessors-method}
\alias{[,numberedObjects-method}
\alias{[<-,numberedObjects-method}
\alias{[[,CNumericList-method}
\alias{[[<-,CNumericList-method}
\alias{[[,RNumericList-method}
\alias{[[<-,RNumericList-method}
\alias{[[,nimPointerList-method}
\alias{[[<-,nimPointerList-method}
\alias{[[<-,nimbleFunctionList-method}
\alias{[,distributionsClass-method}
\alias{[[,distributionsClass-method}
\alias{[[,conjugacyRelationshipsClass-method}
\alias{calc_dmnormConjugacyContributions}
\alias{calc_dmnormAltParams}
\alias{getNodeFunctionIndexedInfo}
\alias{singleModelValuesAccess}
\alias{getNimbleProject}
\alias{nimbleInternalFunctions}
\description{
Functions and classes used internally in NIMBLE and not expected to be called directly by users. Some functions and classes not intended for direct use are documented and/or exported because they are used within Reference Class methods for classes programmatically generated by NIMBLE.
}
\author{NIMBLE Development Team}