\name{nimble-internal}
\title{Functions and Classes Internal to NIMBLE}
\alias{activeBindingTemplate}
\alias{addBlankModelValueRows}
\alias{addIndentToList}
\alias{addMissingIndexingRecurse}
\alias{addModelDollarSign}
\alias{addNecessaryIndexingToNewNode}
\alias{addNodeFxn}
\alias{addSingleModelValuesAccess}
\alias{addSingleVarAccess}
\alias{allNodeNames}
\alias{anyNonScalar}
\alias{areMVSymTabsEqual}
\alias{argType2symbol}
\alias{argTypeList2symbolTable}
\alias{arithmeticOutputType}
\alias{as.matrix.CmodelValues}
\alias{as.matrix.modelValuesBaseClass}
\alias{assignmentAsFirstArgFuns}
\alias{assignmentOperators}
\alias{assignmentTypeWarn}
\alias{binaryLeftOperators}
\alias{binaryMidLogicalOperators}
\alias{binaryMidOperators}
\alias{binaryOperators}
\alias{binaryOrUnaryOperators}
\alias{brackOperators}
\alias{BUGScode}
\alias{BUGScontextClass}
\alias{BUGSmodel}
\alias{BUGSsingleContextClass}
\alias{buildCopyLineFromSEXP}
\alias{buildCopyLineToSEXP}
\alias{buildIntermCalls}
\alias{buildIntermNone}
\alias{buildModelInterface}
\alias{buildNimbleFxnInterface}
\alias{buildSimpleIntermCall}
\alias{buildSymbolTable}
\alias{cAddBlank}
\alias{calcAdaptationFactor}
\alias{calcCoeffAndOffset}
\alias{calc_E_llk_gen}
\alias{callsFromExternalUnaries}
\alias{callToSkipInEigenization}
\alias{cc_checkLinearity}
\alias{cc_combineExprsAddition}
\alias{cc_combineExprsDivision}
\alias{cc_combineExprsMultiplication}
\alias{cc_combineExprsSubtraction}
\alias{cc_expandDetermNodesInExpr}
\alias{cc_getNodeDistributionText}
\alias{cc_getNodeParamExpr}
\alias{cc_getNodesInExpr}
\alias{cc_getNodeValueExpr}
\alias{cc_linkCheck}
\alias{cc_makeConjugateSamplerName}
\alias{cc_makeDDistributionName}
\alias{cc_makeRDistributionName}
\alias{cc_makeSamplerTypeName}
\alias{cc_negateExpr}
\alias{cc_nodeInExpr}
\alias{cCopyVariableRows}
\alias{cc_otherParamsCheck}
\alias{C_dcat}
\alias{C_ddirch}
\alias{C_dmnorm_chol}
\alias{C_dmulti}
\alias{C_dt_nonstandard}
\alias{C_dwish_chol}
\alias{cGetNRow}
\alias{checkForDuplicateNodeDeclaration}
\alias{checkNimbleFunctionListCpp}
\alias{cleanWhite}
\alias{cloglog}
\alias{CmodelBaseClass}
\alias{CmodelValues}
\alias{CnimbleFunctionBase}
\alias{CNumericList}
\alias{codeBlockClass}
\alias{codeSubstitute}
\alias{compileBUGSmodel}
\alias{compileNimbleFunction}
\alias{compileRCfunction}
\alias{conjugacyClass}
\alias{conjugacyRelationshipsClass}
\alias{conjugacyRelationshipsInputList}
\alias{conjugacyRelationshipsObject}
\alias{conjugateSamplerDefinitions}
\alias{controlDefaultList}
\alias{copyModelValuesElements}
\alias{CppBinaryOperators2}
\alias{cppBUGSmodelClass}
\alias{cppCasts}
\alias{cppClassDef}
\alias{cppCodeBlock}
\alias{cppCodeFileClass}
\alias{cppCPPfileClass}
\alias{cppDefinition}
\alias{cppDouble}
\alias{cppEigenMap}
\alias{cppFunctionDef}
\alias{cppHfileClass}
\alias{cppInt}
\alias{cppKeywordsThatFillSemicolons}
\alias{cppLiteral}
\alias{cppMidOperators}
\alias{cppMinusOne}
\alias{cppModelValuesAccessorVector}
\alias{cppModelValuesClass}
\alias{cppModelVariableAccessorVector}
\alias{cppNamedObjectsClass}
\alias{cppNamespace}
\alias{cppNimArr}
\alias{cppNimArrPtr}
\alias{cppNimbleFunctionClass}
\alias{cppNodeFunctionVector}
\alias{cppOutputBinaryOrUnary}
\alias{cppOutputBlank}
\alias{cppOutputBracket}
\alias{cppOutputCall}
\alias{cppOutputCallAsIs}
\alias{cppOutputCalls}
\alias{cppOutputCallWithName}
\alias{cppOutputCast}
\alias{cppOutputChainedCall}
\alias{cppOutputCout}
\alias{cppOutputDereference}
\alias{cppOutputDoubleBracket}
\alias{cppOutputEigBlank}
\alias{cppOutputEigenMapAssign}
\alias{cppOutputEigExternalUnaryFunction}
\alias{cppOutputEigMemberFunction}
\alias{cppOutputEigMemberFunctionNoTranslate}
\alias{cppOutputFor}
\alias{cppOutputIfWhile}
\alias{cppOutputMemberFunction}
\alias{cppOutputMemberFunctionDeref}
\alias{cppOutputMidOperator}
\alias{cppOutputNFmethod}
\alias{cppOutputNFvar}
\alias{cppOutputNimFunListAccess}
\alias{cppOutputNonNativeEigen}
\alias{cppOutputNumList}
\alias{cppOutputParen}
\alias{cppOutputPow}
\alias{cppOutputPtrType}
\alias{cppOutputReturn}
\alias{cppOutputTemplate}
\alias{cppProjectClass}
\alias{cppSEXP}
\alias{cppStrideType}
\alias{cppVar}
\alias{cppVarFull}
\alias{cppVecNimArr}
\alias{cppVecNimArrPtr}
\alias{cppVirtualNimbleFunctionClass}
\alias{cppVoid}
\alias{C_rcat}
\alias{C_rdirch}
\alias{createDefault_isDataObj}
\alias{createMakevars}
\alias{createNamedObjectsFromList}
\alias{C_rmnorm_chol}
\alias{C_rmulti}
\alias{C_rt_nonstandard}
\alias{C_rwish_chol}
\alias{cube}
\alias{dcat}
\alias{ddirch}
\alias{declareHandler}
\alias{dependentClass}
\alias{determineNdimFromOneCase}
\alias{dimOrLength}
\alias{distClass}
\alias{distribution_dFuns}
\alias{distributionFuns}
\alias{distribution_rFuns}
\alias{distributions}
\alias{distributionsClass}
\alias{distributionsInputList}
\alias{dmFunHandler}
\alias{dmnorm_chol}
\alias{dmulti}
\alias{dropSingleSizes}
\alias{dt_nonstandard}
\alias{dwish_chol}
\alias{eigCalls}
\alias{eigenizeArrayize}
\alias{eigenize_asRowOrCol}
\alias{eigenize_assign_before_recurse}
\alias{eigenizeCalls}
\alias{eigenizeCallsBeforeRecursing}
\alias{eigenize_chainedCall}
\alias{eigenize_chol}
\alias{eigenize_cWiseAddSub}
\alias{eigenize_cWiseBinaryArray}
\alias{eigenize_cWiseBinaryEitherMatch}
\alias{eigenize_cWiseBinaryMatrix}
\alias{eigenize_cWiseMultDiv}
\alias{eigenize_cWiseUnaryArray}
\alias{eigenize_cWiseUnaryEither}
\alias{eigenize_cWiseUnaryMatrix}
\alias{eigenize_doNotRecurse}
\alias{eigenizeMatricize}
\alias{eigenizeName}
\alias{eigenizeNameStrided}
\alias{eigenize_nfVar}
\alias{eigenize_reductionArray}
\alias{eigenize_reductionBinaryEither}
\alias{eigenize_reductionEither}
\alias{eigenizeTranslate}
\alias{EigenNewExpr}
\alias{eigOtherMemberFunctionCalls}
\alias{eigProxyCalls}
\alias{eigProxyCallsExternalUnary}
\alias{eigProxyTranslate}
\alias{eigProxyTranslateExternalUnary}
\alias{embedInRbracket}
\alias{embedListInRbracket}
\alias{emptyTypeInfo}
\alias{ETentriesFromExternalUnaries}
\alias{evalBracketArgs}
\alias{evalBracketArgsKnownBracket}
\alias{example_fillInMissingIndices}
\alias{example_getMissingDimensions}
\alias{existsFunctionEnvVar}
\alias{expandMVNames}
\alias{expit}
\alias{exprAsListDrop2}
\alias{exprClass}
\alias{exprClasses_addDebugMarks}
\alias{exprClasses_buildInterms}
\alias{exprClasses_eigenize}
\alias{exprClasses_initSizes}
\alias{exprClasses_insertAssertions}
\alias{exprClasses_labelForEigenization}
\alias{exprClasses_liftMaps}
\alias{exprClasses_processSpecificCalls}
\alias{exprClasses_setSizes}
\alias{exprName2Cpp}
\alias{exprTypeInfoClass}
\alias{extractAnyVectorizedIndexExprs}
\alias{gd_allNeighbors}
\alias{gd_getDependencies_IDs}
\alias{generalFunSizeHandler}
\alias{generateAll}
\alias{generateClassHeader}
\alias{generateFunctionHeader}
\alias{generateObjectDefs}
\alias{genIndexVarValues_recurse}
\alias{genLogProbNodeExprAndReplacements}
\alias{genName2}
\alias{genNodeInfo_singleDeclaration}
\alias{genReplacementsAndCodeRecurse}
\alias{getAvailableNames}
\alias{getBoolValue}
\alias{getCallText}
\alias{getCNumericListValue}
\alias{getCRows}
\alias{getDoubleValue}
\alias{getFunctionEnvVar}
\alias{getIntValue}
\alias{getLoadingNamespace}
\alias{getModelAccessorValues}
\alias{getModelObjectPtr}
\alias{getModelValuesAccessorValues}
\alias{getModelValuesPtrFromModel}
\alias{getModelValuesSpec}
\alias{getMVElement}
\alias{getMVName}
\alias{getMVptr}
\alias{getNimbleTypes}
\alias{getNimValues}
\alias{getNodeFxnPtrs}
\alias{getNRow}
\alias{getPosteriorDensityFromConjSampler}
\alias{getPosteriorDensityFromConjSampler_virtual}
\alias{getSymbolicParentNodes}
\alias{getSymbolicParentNodesRecurse}
\alias{getValuesAccess}
\alias{getVarAndIndices}
\alias{graphNode}
\alias{icloglog}
\alias{identityAssert}
\alias{ifOrWhile}
\alias{ilogit}
\alias{IncludeCodeDir}
\alias{inprod}
\alias{insertExprClassLayer}
\alias{insertIndexingBracket}
\alias{IntermLabelMaker}
\alias{inverse}
\alias{iprobit}
\alias{is.blank}
\alias{isCodeScalar}
\alias{isEigScalar}
\alias{isExprLiftable}
\alias{is.indexed}
\alias{isModelValuesSpec}
\alias{is.nf}
\alias{is.nfGenerator}
\alias{isRbracket}
\alias{is.rcf}
\alias{is.vectorized}
\alias{labelFunctionCreator}
\alias{length.modelValuesAccessorVector}
\alias{length.modelVariableAccessorVector}
\alias{length.nimPointerList}
\alias{linkInverses}
\alias{logdet}
\alias{logfact}
\alias{loggam}
\alias{logit}
\alias{makeBUGSactiveBindingDef}
\alias{makeBUGSclassFields}
\alias{makeCallList}
\alias{makeCNumericListPtr}
\alias{makeCppNIMBLEfunction}
\alias{makeCSingleModelValuesAccessor}
\alias{makeCSingleVariableAccessor}
\alias{MakeCustomModelClass}
\alias{makeCustomModelValuesClass}
\alias{makeDefaultDirName}
\alias{makeEigenArgsMatch}
\alias{makeEigenName}
\alias{makeEigenTypeLabel}
\alias{makeEnvName}
\alias{makeLogProbEnvName}
\alias{makeLogProbName}
\alias{makeMapAccessExpr}
\alias{makeMapExprFromBrackets}
\alias{makeMapSetupCodeExprs}
\alias{makeMapSetupCodeNames}
\alias{makeModelBindingFields}
\alias{makeModelCppCopyTypes}
\alias{makeModelValuesClassFields}
\alias{makeNameFromMatchedCall}
\alias{makeNameName}
\alias{makeNFBindingFields}
\alias{makeNimbleFxnCppCopyTypes}
\alias{makeNimbleFxnInterfaceCallMethodCode}
\alias{makeNumericList}
\alias{makeOffsetRexpr}
\alias{makeRowName}
\alias{makeSingleIndexAccessCodeNames}
\alias{makeSingleIndexAccessExpr}
\alias{makeSingleIndexAccessorSetupCode}
\alias{makeSizeExpressions}
\alias{makeSpaces}
\alias{makeStrideRexprs}
\alias{makeVecName}
\alias{mapsClass}
\alias{matrixEigenOperators}
\alias{matrixFlipOperators}
\alias{matrixMultOperators}
\alias{matrixSolveOperators}
\alias{matrixSquareOperators}
\alias{matrixSquareReductionOperators}
\alias{mcmc_findControlListNamesInCode}
\alias{mcmc_listContentsToStr}
\alias{mcmcNodeInit}
\alias{mcmcNodeInit_virtual}
\alias{midOperators}
\alias{minMaxHandler}
\alias{modelDefInfoClass}
\alias{modelValuesAccessor}
\alias{modelValuesAccessorVector}
\alias{modelValuesElement2Matrix}
\alias{modelVariableAccessor}
\alias{modelVariableAccessorVector}
\alias{ModifiedRmmParseKeywords2}
\alias{mvAccessHandler}
\alias{mvFirstArgCheckLists}
\alias{mvInfoClass}
\alias{nameMashupFromExpr}
\alias{ndf_addModelDollarSignsToMethods}
\alias{ndf_createContains}
\alias{ndf_createMethodList}
\alias{ndf_createSetupFunction}
\alias{ndf_createSingleMethod}
\alias{ndf_createStochCalculate}
\alias{ndf_createStochSimulate}
\alias{ndf_createVirtualNodeFunctionDefinition}
\alias{ndf_createVirtualNodeFunctionDefinitionsList}
\alias{ndf_generateGetParamFunction}
\alias{NeedMakevarsFile}
\alias{neverReplaceable}
\alias{newAssignmentExpression}
\alias{newBracketExpr}
\alias{newEPT}
\alias{newManyModelValuesAccess}
\alias{newManyVarAccess}
\alias{newNodeFxnVec}
\alias{newObjElementPtr}
\alias{nf_assignmentLHSvars}
\alias{nfBuildCInterface}
\alias{nf_changeNimKeywords}
\alias{nf_changeNimKeywordsOne}
\alias{nfCompilationInfoClass}
\alias{nf_createAList}
\alias{nf_createAllNamesFromMethodList}
\alias{nf_createGeneratorFunctionArgs}
\alias{nf_createGeneratorFunctionDef}
\alias{nf_createRefClassDef}
\alias{nf_createRefClassDef_fields}
\alias{nf_createSetupOutputNames}
\alias{nfGetDefVar}
\alias{nf_getGeneratorFunction}
\alias{nf_getInstances}
\alias{nf_getMethodList}
\alias{nf_getNamesFromSetupOutputDeclaration}
\alias{nf_getRefClassObject}
\alias{nf_getSetupOutputNames}
\alias{nf_getVarFromAssignmentLHScode}
\alias{nfMethodErrorHandler}
\alias{nfMethodRC}
\alias{nf_namesNotHidden}
\alias{nfPFstep}
\alias{nfPFStepVir}
\alias{nfProcessing}
\alias{nf_processSetupFunctionBody}
\alias{nf_refClassLabelMaker}
\alias{nfWriteCompileAndLoadSO}
\alias{nimArrMapExpr}
\alias{NimbleCodeDir}
\alias{nimbleEquals}
\alias{nimbleFunctionList}
\alias{nimbleIncludeFile}
\alias{nimbleOrRfunctionNames}
\alias{nimbleProjectClass}
\alias{nimbleStep}
\alias{nimbleUniqueID}
\alias{nimCopyAccess}
\alias{nimCppKeywordsThatFillSemicolon}
\alias{nimDeparse}
\alias{nimGenerateCpp}
\alias{nimKeyWords}
\alias{nimPointerList}
\alias{nl_addIndicesToVariables}
\alias{nl_aggregateConsecutiveBlocks}
\alias{nl_checkNodeNamesInModel}
\alias{nl_checkVarNamesInModel}
\alias{nl_createVarsAndFlatIndexRanges}
\alias{nl_determineFlatIndex}
\alias{nl_expandNodeIndex}
\alias{nl_expandNodeIndexExpr}
\alias{nl_expandNodeNames}
\alias{nl_getVarNameFromNodeName}
\alias{nl_nodeVectorReadyNodes}
\alias{nl_removeNodeNamesNotInSymbolTable}
\alias{nl_vectorizedExpandNodeIndex}
\alias{nl_vectorizedExpandNodeIndexExprs}
\alias{node_determ}
\alias{nodeFunction}
\alias{nodeFunctionVector}
\alias{node_stoch_dbern}
\alias{node_stoch_dbeta}
\alias{node_stoch_dbin}
\alias{node_stoch_dcat}
\alias{node_stoch_dchisq}
\alias{node_stoch_ddirch}
\alias{node_stoch_dexp}
\alias{node_stoch_dgamma}
\alias{node_stoch_dlnorm}
\alias{node_stoch_dlogis}
\alias{node_stoch_dmnorm}
\alias{node_stoch_dmulti}
\alias{node_stoch_dnegbin}
\alias{node_stoch_dnorm}
\alias{node_stoch_dpois}
\alias{node_stoch_dt}
\alias{node_stoch_dunif}
\alias{node_stoch_dweib}
\alias{node_stoch_dwish}
\alias{nonNativeEigenCalls}
\alias{nonNativeEigenProxyCalls}
\alias{numericList}
\alias{NumericListBase}
\alias{operatorRank}
\alias{outputCppArrayIndex2}
\alias{outputCppBinaryOperation2}
\alias{outputCppCommentBlock2}
\alias{outputCppCout2}
\alias{outputCppDereference2}
\alias{outputCppExit2}
\alias{outputCppFor2}
\alias{outputCppIfElse2}
\alias{outputCppLiteral}
\alias{outputCppOpenBracket2}
\alias{outputCppOpenBracketInvisible}
\alias{outputCppOpenParentheses2}
\alias{outputCppParseTree2}
\alias{outputCppReference2}
\alias{outputCppScopeResolution2}
\alias{outputCppTempArray2}
\alias{outputCppWhile2}
\alias{outputEigenOperation2}
\alias{outputGeneralCppFunction2}
\alias{outputTypeCast2}
\alias{parseTreeSubstitute}
\alias{passThroughOperators}
\alias{pasteSemicolon}
\alias{phi}
\alias{pointAt}
\alias{populateManyModelValuesAccess}
\alias{populateManyModelVarAccess}
\alias{populateNodeFxnVec}
\alias{posteriorClass}
\alias{pow}
\alias{probit}
\alias{processModelFile}
\alias{processVarBlock}
\alias{projectNameCreator}
\alias{putCodeLinesInBrackets}
\alias{rCalcNodes}
\alias{rcat}
\alias{RCfunction}
\alias{RCfunctionCompileClass}
\alias{RCfunctionDef}
\alias{RCfunInfoClass}
\alias{rcFunLabelMaker}
\alias{RCfunProcessing}
\alias{RCvirtualFunProcessing}
\alias{rdirch}
\alias{recurseSetSizes}
\alias{reductionBinaryOperators}
\alias{reductionBinaryOperatorsEither}
\alias{reductionUnaryOperators}
\alias{reductionUnaryOperatorsArray}
\alias{reductionUnaryOperatorsEither}
\alias{removeExprClassLayer}
\alias{removeIndexing}
\alias{removeLayerHandler}
\alias{removeLogProbName}
\alias{removeNodeFxn}
\alias{removeNonScalarElementsFromList}
\alias{removeSingleModelValuesAccess}
\alias{removeSingleVarAccess}
\alias{replaceAllCodeSuccessfully}
\alias{replaceConstantsRecurse}
\alias{replacementHandler}
\alias{replaceWhatWeCan}
\alias{resizeCModelAccessors}
\alias{resizeCModelValuesAccessors}
\alias{resizeManyModelValuesAccessor}
\alias{resizeManyModelVarAccessor}
\alias{resizeNodeFxnVec}
\alias{rFunHandler}
\alias{rGetLogProbsNodes}
\alias{RHSonlyInit}
\alias{RHSonlyInit_virtual}
\alias{RMakeCustomModelClass}
\alias{rmFunHandler}
\alias{rmnorm_chol}
\alias{RModelBaseClass}
\alias{rmulti}
\alias{Rname2CppName}
\alias{RNumericList}
\alias{rop}
\alias{RparseTree2ExprClasses}
\alias{R_setDoublePtrFromSinglePtr}
\alias{R_setOnePtrVectorOfPtrs}
\alias{R_setPtrVectorOfPtrs}
\alias{R_setSinglePtrFromSinglePtr}
\alias{rSimNodes}
\alias{rt_nonstandard}
\alias{rwish_chol}
\alias{sampler_BASE}
\alias{sampler_conjugate_beta}
\alias{sampler_conjugate_gamma}
\alias{sampler_conjugate_norm}
\alias{sampler_crossLevel}
\alias{sampler_end}
\alias{sampler_RW}
\alias{sampler_RW_block}
\alias{sampler_RW_llFunction}
\alias{sampler_slice}
\alias{samplerSpec}
\alias{scalarOutputTypes}
\alias{seqAlongHandler}
\alias{setArg}
\alias{setAsRowOrCol}
\alias{setBoolValue}
\alias{setCaller}
\alias{setCNumericListValue}
\alias{setDoublePtrFromSinglePtr}
\alias{setDoubleValue}
\alias{setIntValue}
\alias{setMVElement}
\alias{setNimValues}
\alias{setNodeElement}
\alias{setNodeModelPtr}
\alias{setOnePtrVectorOfPtrs}
\alias{setPtrVectorOfPtrs}
\alias{setSinglePtrFromSinglePtr}
\alias{setValuesAccess}
\alias{SEXPscalarConvertFunctions}
\alias{singleModelValuesAccess}
\alias{singleModelValuesAccessClass}
\alias{singleVarAccess}
\alias{singleVarAccessClass}
\alias{sizeasDoublePtr}
\alias{sizeAsRowOrCol}
\alias{sizeAssign}
\alias{sizeAssignAfterRecursing}
\alias{sizeBinaryCwise}
\alias{sizeBinaryCwiseLogical}
\alias{sizeBinaryReduction}
\alias{sizeBinaryUnaryCwise}
\alias{sizecallC}
\alias{sizeCalls}
\alias{sizeChainedCall}
\alias{sizeColonOperator}
\alias{sizeDimOperator}
\alias{sizeFor}
\alias{sizeforceEigenize}
\alias{sizeIndexingBracket}
\alias{sizeInsertIntermediate}
\alias{sizemap}
\alias{sizeMatrixMult}
\alias{sizeMatrixSquareReduction}
\alias{sizemvAccessBracket}
\alias{sizeNFvar}
\alias{sizeNimbleFunction}
\alias{sizeOneEigenCommand}
\alias{sizeRCfunction}
\alias{sizeResizeNoPtr}
\alias{sizeReturn}
\alias{sizeRmultivarFirstArg}
\alias{sizeRMVNorm}
\alias{sizeScalar}
\alias{sizeScalarRecurse}
\alias{sizeSetSize}
\alias{sizeSolveOp}
\alias{sizeTranspose}
\alias{sizeUnaryCwise}
\alias{sizeUnaryCwiseSquare}
\alias{sizeUnaryNonaryCwise}
\alias{sizeUnaryReduction}
\alias{sizeUndefined}
\alias{sizeValues}
\alias{specificCallHandlers}
\alias{specificCallReplacements}
\alias{symbolBase}
\alias{symbolBasic}
\alias{symbolDouble}
\alias{symbolEigenMap}
\alias{symbolInt}
\alias{symbolMemberFunction}
\alias{symbolModel}
\alias{symbolModelValues}
\alias{symbolModelValuesAccessorVector}
\alias{symbolModelVariableAccessorVector}
\alias{symbolNimArrDoublePtr}
\alias{symbolNimbleFunction}
\alias{symbolNimbleFunctionList}
\alias{symbolNimPtrList}
\alias{symbolNodeFunctionVector}
\alias{symbolNumericList}
\alias{symbolPtr}
\alias{symbolTable}
\alias{symbolTable2cppVars}
\alias{symbolVecNimArrPtr}
\alias{toSEXPscalarConvertFunctions}
\alias{unaryOperators}
\alias{unaryOrNonaryOperators}
\alias{UseLibraryMakevars}
\alias{v}
\alias{varAndIndices2flatIndex}
\alias{varAndIndices2mapParts}
\alias{varInfoClass}
\alias{virtualNFprocessing}
\alias{virtualNodeFunctionDefinitions}
\alias{wrapAsNumeric}
\alias{writeCode}
\description{
Functions and classes used internally in NIMBLE and not expected to be called directly by users. For the moment all functions and classes are exported while we are sorting out questions about the use of non-exported functions in Reference Class methods.
}
\author{NIMBLE Development Team}