\name{nimble-internal}
\title{Functions and Classes Internal to NIMBLE}
\alias{accessorVectorLength_setupCodeTemplate}
\alias{activeBindingTemplate}
\alias{activeBindingTemplateLength1NonScalar}
\alias{addArg}
\alias{addBlankModelValueRows}
\alias{addBlockFromCppName}
\alias{addDistKeywordProcessors}
\alias{addDistList2matchFunctions}
\alias{addIndentToList}
\alias{addMissingIndexingRecurse}
\alias{addModelDollarSign}
\alias{addNecessaryIndexingToNewNode}
\alias{addNecessarySetupCode}
\alias{addNewCode}
\alias{addSamplerToSpec}
\alias{addSetupCodeNames}
\alias{addSingleModelValuesAccess}
\alias{addSingleVarAccess}
\alias{addToTypeEnv}
\alias{allLHSNodes_SetupTemplate}
\alias{allModelNodes_SetupTemplate}
\alias{allModelValuesVars_SetupTemplate}
\alias{allParamEfficiencyComparisonComponent}
\alias{anyNonScalar}
\alias{areMVSymTabsEqual}
\alias{argInfo2PointerStaticCast}
\alias{argType2symbol}
\alias{argTypeList2symbolTable}
\alias{arithmeticOutputType}
\alias{as.matrix.CmodelValues}
\alias{as.matrix.modelValuesBaseClass}
\alias{assignLogProbName}
\alias{assignmentAsFirstArgFuns}
\alias{assignmentOperators}
\alias{assignmentTypeWarn}
\alias{assignNumberedObject}
\alias{autoBlockClass}
\alias{autoBlockModel}
\alias{autoBlockParamDefaults}
\alias{auxFStep}
\alias{auxFuncVirtual}
\alias{auxLookFunc}
\alias{auxSimFunc}
\alias{auxStepVirtual}
\alias{binaryLeftOperators}
\alias{binaryMidLogicalOperators}
\alias{binaryMidOperators}
\alias{binaryOperators}
\alias{binaryOrUnaryOperators}
\alias{bootFStep}
\alias{bootStepVirtual}
\alias{brackOperators}
\alias{BUGScode}
\alias{BUGScontextClass}
\alias{BUGSdistToRdist}
\alias{BUGSmodel}
\alias{BUGSsingleContextClass}
\alias{buildCopyLineFromSEXP}
\alias{buildCopyLineToSEXP}
\alias{buildIntermCalls}
\alias{buildIntermNone}
\alias{buildModelInterface}
\alias{buildNeededObjects}
\alias{buildNimbleFxnInterface}
\alias{buildSimpleIntermCall}
\alias{buildSymbolTable}
\alias{C_dcat}
\alias{C_ddirch}
\alias{C_dinterval}
\alias{C_dmnorm_chol}
\alias{C_dmulti}
\alias{C_dt_nonstandard}
\alias{C_dwish_chol}
\alias{C_rcat}
\alias{C_rdirch}
\alias{C_rinterval}
\alias{C_rmnorm_chol}
\alias{C_rmulti}
\alias{C_rt_nonstandard}
\alias{C_rwish_chol}
\alias{cAddBlank}
\alias{calc_asympVar}
\alias{calc_dmnormAltParams}
\alias{calc_dmnormConjugacyContributions}
\alias{calc_E_llk_gen}
\alias{calcAdaptationFactor}
\alias{calculate_keywordInfo}
\alias{calculateDiff_keywordInfo}
\alias{callsFromExternalUnaries}
\alias{callToPackVPtrs}
\alias{callToSkipInEigenization}
\alias{cc_checkLinearity}
\alias{cc_combineExprsAddition}
\alias{cc_combineExprsDivision}
\alias{cc_combineExprsMultiplication}
\alias{cc_combineExprsSubtraction}
\alias{cc_createStructureExpr}
\alias{cc_expandDetermNodesInExpr}
\alias{cc_getNodesInExpr}
\alias{cc_linkCheck}
\alias{cc_makeConjugateSamplerName}
\alias{cc_makeRDistributionName}
\alias{cc_makeSamplerTypeName}
\alias{cc_negateExpr}
\alias{cc_nodeInExpr}
\alias{cc_otherParamsCheck}
\alias{cc_structureExprName}
\alias{cc_vectorizedComponentCheck}
\alias{cCopyVariableRows}
\alias{cGetNRow}
\alias{character2index}
\alias{checkDistributionsFunctions}
\alias{checkDistributionsInput}
\alias{checkForDeterministicDorR}
\alias{checkForDuplicateNodeDeclaration}
\alias{checkNimbleFunctionListCpp}
\alias{checkNimbleOrRfunctionNames}
\alias{checkRHSonlyInit}
\alias{cleanWhite}
\alias{cloglog}
\alias{Cmatrix2mvOneVar}
\alias{CmodelValues}
\alias{CmultiNimbleFunctionClass}
\alias{CNumericList}
\alias{code2Name_fromArgList}
\alias{codeProcessIfThenElse}
\alias{codeSubstitute}
\alias{collectEdges}
\alias{collectInferredVertexEdges}
\alias{combineIndices2Flat}
\alias{compareConjugacyLists}
\alias{compileBUGSmodel}
\alias{compileNimbleFunction}
\alias{compileRCfunction}
\alias{conjugacyClass}
\alias{conjugacyRelationshipsClass}
\alias{conjugacyRelationshipsInputList}
\alias{conjugacyRelationshipsObject}
\alias{copierVector}
\alias{copierVector_setupCodeTemplate}
\alias{copyFromRobject}
\alias{copyFromRobjectViaActiveBindings}
\alias{copyModelValuesElements}
\alias{CppBinaryOperators2}
\alias{cppBUGSmodelClass}
\alias{cppCasts}
\alias{cppClassDef}
\alias{cppCodeBlock}
\alias{cppCodeFileClass}
\alias{cppCPPfileClass}
\alias{cppDefinition}
\alias{cppDouble}
\alias{cppEigenMap}
\alias{cppFunctionDef}
\alias{cppHfileClass}
\alias{cppInt}
\alias{cppKeywordsThatFillSemicolons}
\alias{cppLiteral}
\alias{cppMidOperators}
\alias{cppMinusOne}
\alias{cppModelValuesAccessorVector}
\alias{cppModelValuesClass}
\alias{cppModelValuesMapAccessorVector}
\alias{cppModelVariableAccessorVector}
\alias{cppModelVariableMapAccessorVector}
\alias{cppNamedObjectsClass}
\alias{cppNamespace}
\alias{cppNimArr}
\alias{cppNimArrPtr}
\alias{cppNimbleFunctionClass}
\alias{cppNodeFunctionVector}
\alias{cppOptimObject}
\alias{cppOutputBinaryOrUnary}
\alias{cppOutputBlank}
\alias{cppOutputBracket}
\alias{cppOutputCall}
\alias{cppOutputCallAsIs}
\alias{cppOutputCalls}
\alias{cppOutputCallWithName}
\alias{cppOutputCast}
\alias{cppOutputChainedCall}
\alias{cppOutputCout}
\alias{cppOutputDereference}
\alias{cppOutputDoubleBracket}
\alias{cppOutputEigBlank}
\alias{cppOutputEigenMapAssign}
\alias{cppOutputEigExternalUnaryFunction}
\alias{cppOutputEigMemberFunction}
\alias{cppOutputEigMemberFunctionNoTranslate}
\alias{cppOutputFor}
\alias{cppOutputGetNodeFunctionIndexedInfo}
\alias{cppOutputGetParam}
\alias{cppOutputIfWhile}
\alias{cppOutputMemberFunction}
\alias{cppOutputMemberFunctionDeref}
\alias{cppOutputMemberFunctionGeneric}
\alias{cppOutputMidOperator}
\alias{cppOutputNFmethod}
\alias{cppOutputNFvar}
\alias{cppOutputNimFunListAccess}
\alias{cppOutputNimSwitch}
\alias{cppOutputNonNativeEigen}
\alias{cppOutputNumList}
\alias{cppOutputParen}
\alias{cppOutputPow}
\alias{cppOutputPtrType}
\alias{cppOutputReturn}
\alias{cppOutputTemplate}
\alias{cppOutputVoidPtr}
\alias{cppProjectClass}
\alias{cppSEXP}
\alias{cppStrideType}
\alias{cppVar}
\alias{cppVarFull}
\alias{cppVecNimArr}
\alias{cppVecNimArrPtr}
\alias{cppVecVoidPtr}
\alias{cppVirtualNimbleFunctionClass}
\alias{cppVoid}
\alias{createDefault_isDataObj}
\alias{createDFfromABlist}
\alias{createDynamicConjugateSamplerName}
\alias{createMakevars}
\alias{createNamedObjectsFromList}
\alias{cube}
\alias{d_dist_keywordInfo}
\alias{d_exp_nimble_keywordInfo}
\alias{d_gamma_keywordInfo}
\alias{declareHandler}
\alias{deparse}
\alias{dependentClass}
\alias{determineContextSize}
\alias{determineNdimFromOneCase}
\alias{determineNdimsFromNfproc}
\alias{determineNodeIndexSizes}
\alias{dimOrLength}
\alias{distClass}
\alias{distfun}
\alias{distribution_dFuns}
\alias{distribution_pFuns}
\alias{distribution_qFuns}
\alias{distribution_rFuns}
\alias{distributionAliases}
\alias{distributionFuns}
\alias{distributions}
\alias{distributionsClass}
\alias{distributionsInputList}
\alias{dmFunHandler}
\alias{dollarSign_keywordInfo}
\alias{doPars}
\alias{double}
\alias{doubleBracket_keywordInfo}
\alias{dropSingleSizes}
\alias{dynamicConjugateSamplerAdd}
\alias{dynamicConjugateSamplerDefinitionsEnv}
\alias{dynamicConjugateSamplerExists}
\alias{dynamicConjugateSamplerFunctionsEnv}
\alias{dynamicConjugateSamplerGet}
\alias{dynamicConjugateSamplerWrite}
\alias{effectiveSizeHO}
\alias{effectiveSizeStan}
\alias{efficiencyDetailsComparisonComponent}
\alias{eigCalls}
\alias{eigenize_asRowOrCol}
\alias{eigenize_assign_before_recurse}
\alias{eigenize_chainedCall}
\alias{eigenize_cWiseAddSub}
\alias{eigenize_cWiseBinaryArray}
\alias{eigenize_cWiseBinaryEitherMatch}
\alias{eigenize_cWiseBinaryMatrix}
\alias{eigenize_cWiseByScalarArray}
\alias{eigenize_cWiseMultDiv}
\alias{eigenize_cWiseUnaryArray}
\alias{eigenize_cWiseUnaryEither}
\alias{eigenize_cWiseUnaryMatrix}
\alias{eigenize_doNotRecurse}
\alias{eigenize_matrixOps}
\alias{eigenize_nfVar}
\alias{eigenize_reductionArray}
\alias{eigenize_reductionBinaryEither}
\alias{eigenize_reductionEither}
\alias{eigenizeArrayize}
\alias{eigenizeCalls}
\alias{eigenizeCallsBeforeRecursing}
\alias{eigenizeMatricize}
\alias{eigenizeName}
\alias{eigenizeNameStrided}
\alias{eigenizeTranslate}
\alias{EigenNewExpr}
\alias{eigOtherMemberFunctionCalls}
\alias{eigProxyCalls}
\alias{eigProxyCallsExternalUnary}
\alias{eigProxyTranslate}
\alias{eigProxyTranslateExternalUnary}
\alias{embedInRbracket}
\alias{embedListInRbracket}
\alias{emptyTypeInfo}
\alias{ENKFFuncVirtual}
\alias{enkfMultFunc}
\alias{enkfScalFunc}
\alias{ENKFStep}
\alias{ENKFStepVirtual}
\alias{ETentriesFromExternalUnaries}
\alias{evalBracketArgs}
\alias{evalBracketArgsKnownBracket}
\alias{evalInDistsMatchCallEnv}
\alias{example_fillInMissingIndices}
\alias{example_getMissingDimensions}
\alias{existsFunctionEnvVar}
\alias{expandContextAndReplacements}
\alias{expandIndexSet4sapply}
\alias{expandMVNames}
\alias{expit}
\alias{exprAsListDrop2}
\alias{exprClass}
\alias{exprClasses_addDebugMarks}
\alias{exprClasses_buildInterms}
\alias{exprClasses_eigenize}
\alias{exprClasses_initSizes}
\alias{exprClasses_insertAssertions}
\alias{exprClasses_labelForEigenization}
\alias{exprClasses_liftMaps}
\alias{exprClasses_processSpecificCalls}
\alias{exprClasses_setSizes}
\alias{exprClassProcessingErrorMsg}
\alias{exprName2Cpp}
\alias{exprTypeInfoClass}
\alias{extractAnyVectorizedIndexExprs}
\alias{extractFlatIndices_wVarInfo}
\alias{funName2OptimFunName}
\alias{gd_allNeighbors}
\alias{gd_getDependencies_IDs}
\alias{generalFunSizeHandler}
\alias{generateAll}
\alias{generateClassHeader}
\alias{generateFunctionHeader}
\alias{generateObjectDefs}
\alias{genIndexVarValues_recurse}
\alias{genLogProbNodeExprAndReplacements}
\alias{genName2}
\alias{genReplacementsAndCodeRecurse}
\alias{getAdditonalArgsListForOptim}
\alias{getArgInfoFromNFSym}
\alias{getArgsFromOptimCallForTargetFun}
\alias{getAvailableNames}
\alias{getBoolValue}
\alias{getBUGSexampleDir}
\alias{getCallText}
\alias{getCharacterValue}
\alias{getCharacterVectorValue}
\alias{getCNumericListValue}
\alias{getCRows}
\alias{getDependencyPaths}
\alias{getDistribution}
\alias{getDistribution2}
\alias{getDistributionList}
\alias{getDistributionsInfo}
\alias{getDoubleValue}
\alias{getFunctionEnvVar}
\alias{getFxnVectorPtr}
\alias{getIntValue}
\alias{getLoadingNamespace}
\alias{getLogProb_keywordInfo}
\alias{getMaxDim}
\alias{getModelAccessorValues}
\alias{getModelObjectPtr}
\alias{getModelValuesAccessorValues}
\alias{getModelValuesPtrFromModel}
\alias{getModelValuesSpec}
\alias{getMVElement}
\alias{getMVName}
\alias{getMVptr}
\alias{getNamedObjected}
\alias{getNimbleFunctionEnvironment}
\alias{getNimbleProject}
\alias{getNimbleTypes}
\alias{getNimValues}
\alias{getNodeFunctionIndexedInfo}
\alias{getNRow}
\alias{getNumberedObject}
\alias{getParam_keywordInfo}
\alias{getPosteriorDensityFromConjSampler}
\alias{getPosteriorDensityFromConjSampler_virtual}
\alias{getSetCharacter}
\alias{getSetCharacterVector}
\alias{getSetDoubleScalar}
\alias{getSetIntegerScalar}
\alias{getSetLogicalScalar}
\alias{getSetModelValues}
\alias{getSetModelValuesPtr}
\alias{getSetModelVarPtr}
\alias{getSetNimbleFunction}
\alias{getSetNimPtrList}
\alias{getSetNumericVector}
\alias{getSize_NumberedObjects}
\alias{getSymbolicParentNodes}
\alias{getSymbolicParentNodesRecurse}
\alias{getValueDim}
\alias{getValues}
\alias{getValuesAccess}
\alias{getVarAndIndices}
\alias{graphNode}
\alias{handleScaleAndRateForExpNimble}
\alias{handleScaleAndRateForGamma}
\alias{hasBracket}
\alias{icloglog}
\alias{identityAssert}
\alias{ifOrWhile}
\alias{ilogit}
\alias{IncludeCodeDir}
\alias{indexedNodeInfoTableClass}
\alias{inner_populateNodeFxnVec}
\alias{inprod}
\alias{insertExprClassLayer}
\alias{insertIndexingBracket}
\alias{insertSingleIndexBrackets}
\alias{insertSubIndexExpr}
\alias{IntermLabelMaker}
\alias{inverse}
\alias{iprobit}
\alias{is.blank}
\alias{is.Cnf}
\alias{is.indexed}
\alias{is.na.vec}
\alias{is.nan.vec}
\alias{is.nf}
\alias{is.nfGenerator}
\alias{is.rcf}
\alias{is.vectorized}
\alias{isCodeArgBlank}
\alias{isCodeScalar}
\alias{isEigScalar}
\alias{isExprLiftable}
\alias{isModelValuesSpec}
\alias{isNameInExpr}
\alias{isNameInExprList}
\alias{isRbracket}
\alias{isSetupCodeGenerated}
\alias{isSymbolType}
\alias{isValid}
\alias{keywordInfoClass}
\alias{keywordList}
\alias{keywordListModelMemberFuns}
\alias{keywordOnlyMatchDistList}
\alias{labelFunctionCreator}
\alias{length.nimPointerList}
\alias{linkInverses}
\alias{logdet}
\alias{logfact}
\alias{loggam}
\alias{logit}
\alias{LWparFunc}
\alias{LWSetMeanVirtual}
\alias{LWSetParVirtual}
\alias{LWStep}
\alias{LWStepVirtual}
\alias{make_example_html}
\alias{make_main}
\alias{makeBUGSactiveBindingDef}
\alias{makeBUGSclassFields}
\alias{makeCallList}
\alias{makeCNumericListPtr}
\alias{makeCppNIMBLEfunction}
\alias{makeCSingleModelValuesAccessor}
\alias{makeCSingleVariableAccessor}
\alias{MakeCustomModelClass}
\alias{makeCustomModelValuesClass}
\alias{makeDeclareSizeField}
\alias{makeDefaultDirName}
\alias{makeDSLCallforVoidPtr_fromArgInfAndCall}
\alias{makeEigenArgsMatch}
\alias{makeEigenName}
\alias{makeEigenTypeLabel}
\alias{makeEnvName}
\alias{makeFillValue}
\alias{makeFixedParameterCallCodeForOptim}
\alias{makefnScaleLine}
\alias{makeGetCodeFromAccessorVector}
\alias{makeIndexedVariable}
\alias{makeIndexNamePieces}
\alias{makeLogProbEnvName}
\alias{makeLogProbName}
\alias{makeMapAccessExpr}
\alias{makeMapExprFromBrackets}
\alias{makeMapInfoFromAccessorVector}
\alias{makeMapInfoFromAccessorVectorFaster}
\alias{makeMapSetupCodeNames}
\alias{makeModelBindingFields}
\alias{makeModelCppCopyTypes}
\alias{makeModelValuesClassFields}
\alias{makeMV_GID_Map}
\alias{makeNameName}
\alias{makeNewParsCharCopyLine}
\alias{makeNewSpecFromOldSpec}
\alias{makeNFBindingFields}
\alias{makeNimbleFxnCppCopyTypes}
\alias{makeNimbleFxnInterfaceCallMethodCode}
\alias{makeNumericList}
\alias{makeOffsetRexpr}
\alias{makeOptimFunctionCallLine}
\alias{makeRowName}
\alias{makeSetCodeFromAccessorVector}
\alias{makeSingleIndexAccessCodeNames}
\alias{makeSingleIndexAccessExpr}
\alias{makeSizeExpressions}
\alias{makeSpaces}
\alias{makeStaticCastingChunk}
\alias{makeStrideRexprs}
\alias{makeVecName}
\alias{makeVertexNamesFromIndexArray}
\alias{makeVertexNamesFromIndexArray2}
\alias{makeVoidPointerName_fromObjName}
\alias{map_SetupTemplate}
\alias{mapsClass}
\alias{matchAndFill.call}
\alias{matchDistList}
\alias{matchFunctions}
\alias{matchKeywordCode}
\alias{matchKeywordCodeMemberFun}
\alias{matchKeywords_recurse}
\alias{matchModelMemberFunctions}
\alias{matrix2mv}
\alias{matrixEigenOperators}
\alias{matrixFlipOperators}
\alias{matrixMultOperators}
\alias{matrixSolveOperators}
\alias{matrixSquareOperators}
\alias{matrixSquareReductionOperators}
\alias{mcmc_findControlListNamesInCode}
\alias{mcmc_listContentsToStr}
\alias{MCMCefficiencyHO}
\alias{MCMCefficiencyHOlist}
\alias{mergeMultiLineStatements}
\alias{midOperators}
\alias{minMaxHandler}
\alias{minMeanAllComparisonComponent}
\alias{minMeanComparisonComponent}
\alias{modelDefInfoClass}
\alias{modelMemberFun_keywordInfo}
\alias{modelValuesAccessorVector}
\alias{modelValuesAccessorVector_setupCodeTemplate}
\alias{modelValuesElement2Matrix}
\alias{modelVariableAccessorVector}
\alias{modelVariableAccessorVector_setupCodeTemplate}
\alias{mvAccessHandler}
\alias{mvFirstArgCheckLists}
\alias{mvInfoClass}
\alias{ndf_addModelDollarSignsToMethods}
\alias{ndf_createContains}
\alias{ndf_createMethodList}
\alias{ndf_createSetupFunction}
\alias{ndf_createSingleMethod}
\alias{ndf_createStochCalculate}
\alias{ndf_createStochCalculateTrunc}
\alias{ndf_createStochSimulate}
\alias{ndf_createStochSimulateTrunc}
\alias{ndf_createVirtualNodeFunctionDefinition}
\alias{ndf_createVirtualNodeFunctionDefinitionsList}
\alias{ndf_generateGetParamFunction}
\alias{ndf_generateGetParamSwitchFunction}
\alias{NeedMakevarsFile}
\alias{neverReplaceable}
\alias{newAssignmentExpression}
\alias{newBracketExpr}
\alias{newEPT}
\alias{newManyModelValuesAccess}
\alias{newManyVarAccess}
\alias{newNumberedObjects}
\alias{newObjElementPtr}
\alias{newSpacesFunction}
\alias{nf_assignmentLHSvars}
\alias{nf_changeNimKeywords}
\alias{nf_changeNimKeywordsOne}
\alias{nf_createAList}
\alias{nf_createAllNamesFromMethodList}
\alias{nf_createGeneratorFunctionArgs}
\alias{nf_createGeneratorFunctionDef}
\alias{nf_createRefClassDef}
\alias{nf_createRefClassDef_fields}
\alias{nf_createSetupOutputNames}
\alias{nf_getGeneratorFunction}
\alias{nf_getInstances}
\alias{nf_getMethodList}
\alias{nf_getNamesFromSetupOutputDeclaration}
\alias{nf_getRefClassObject}
\alias{nf_getSetupOutputNames}
\alias{nf_getVarFromAssignmentLHScode}
\alias{nf_namesNotHidden}
\alias{nf_preProcessMemberDataObject}
\alias{nf_processSetupFunctionBody}
\alias{nf_refClassLabelMaker}
\alias{nf_substituteExceptFunctionsAndDollarSigns}
\alias{nfBuildCInterface}
\alias{nfCompilationInfoClass}
\alias{nfGetDefVar}
\alias{nfMethodErrorHandler}
\alias{nfMethodRC}
\alias{nfProcessing}
\alias{nfWriteCompileAndLoadSO}
\alias{nimArrayGeneralHandler}
\alias{nimArrMapExpr}
\alias{NimbleCodeDir}
\alias{nimbleFunctionList}
\alias{nimbleGeneralParseDeparse}
\alias{nimbleGraphClass}
\alias{nimbleIncludeFile}
\alias{nimbleOrRfunctionNames}
\alias{nimbleProjectClass}
\alias{nimbleUniqueID}
\alias{nimbleUserNamespace}
\alias{nimCopy_keywordInfo}
\alias{nimCppKeywordsThatFillSemicolon}
\alias{nimDeparse}
\alias{nimEquals}
\alias{nimGenerateCpp}
\alias{nimKeyWords}
\alias{nimOptim}
\alias{nimOptim_keywordInfo}
\alias{nimPointerList}
\alias{nimRound}
\alias{nimStep}
\alias{nimSwitch}
\alias{nl_addIndicesToVariables}
\alias{nl_aggregateConsecutiveBlocks}
\alias{nl_checkNodeNamesInModel}
\alias{nl_checkVarNamesInModel}
\alias{nl_createVarsAndFlatIndexRanges}
\alias{nl_determineFlatIndex}
\alias{nl_expandNodeIndex}
\alias{nl_expandNodeIndexExpr}
\alias{nl_expandNodeNames}
\alias{nl_getVarNameFromNodeName}
\alias{nl_nodeVectorReadyNodes}
\alias{nl_removeNodeNamesNotInSymbolTable}
\alias{nl_vectorizedExpandNodeIndex}
\alias{nl_vectorizedExpandNodeIndexExprs}
\alias{nm_seq_noDecrease}
\alias{nndf_addModelDollarSignsToMethods}
\alias{nndf_createMethodList}
\alias{nndf_createSetupFunction}
\alias{nndf_generateGetParamSwitchFunction}
\alias{nndf_makeNodeFunctionIndexAccessCall}
\alias{nndf_makeNodeFunctionIndexLabels}
\alias{nndf_replaceSetupOutputsWithIndexedNodeInfo}
\alias{node_determ}
\alias{node_stoch_dbern}
\alias{node_stoch_dbeta}
\alias{node_stoch_dbin}
\alias{node_stoch_dcat}
\alias{node_stoch_dchisq}
\alias{node_stoch_dconstraint}
\alias{node_stoch_ddirch}
\alias{node_stoch_dexp}
\alias{node_stoch_dgamma}
\alias{node_stoch_dinterval}
\alias{node_stoch_dlnorm}
\alias{node_stoch_dlogis}
\alias{node_stoch_dmnorm}
\alias{node_stoch_dmulti}
\alias{node_stoch_dmvt}
\alias{node_stoch_dnegbin}
\alias{node_stoch_dnorm}
\alias{node_stoch_dpois}
\alias{node_stoch_dt}
\alias{node_stoch_dunif}
\alias{node_stoch_dweib}
\alias{node_stoch_dwish}
\alias{nodeFunction}
\alias{nodeFunctionNew}
\alias{nodeFunctionVector}
\alias{nodeFunctionVector_SetupTemplate}
\alias{nodeInit_virtual}
\alias{nonNativeEigenCalls}
\alias{nonNativeEigenProxyCalls}
\alias{numberedModelValuesAccessors}
\alias{numberedModelVariableAccessors}
\alias{numberedObjects}
\alias{numericList}
\alias{NumericListBase}
\alias{operatorRank}
\alias{operatorsAllowedBeforeIndexBracketsWithoutLifting}
\alias{OptimFun_argDefs}
\alias{optimReadyFun_setupCodeTemplate}
\alias{OptimReadyFunction}
\alias{outputCppArrayIndex2}
\alias{outputCppBinaryOperation2}
\alias{outputCppCommentBlock2}
\alias{outputCppCout2}
\alias{outputCppDereference2}
\alias{outputCppExit2}
\alias{outputCppFor2}
\alias{outputCppIfElse2}
\alias{outputCppLiteral}
\alias{outputCppOpenBracket2}
\alias{outputCppOpenBracketInvisible}
\alias{outputCppOpenParentheses2}
\alias{outputCppParseTree2}
\alias{outputCppReference2}
\alias{outputCppScopeResolution2}
\alias{outputCppTempArray2}
\alias{outputCppWhile2}
\alias{outputEigenOperation2}
\alias{outputGeneralCppFunction2}
\alias{outputTypeCast2}
\alias{paramInfo_SetupTemplate}
\alias{paramMean}
\alias{parseEvalCharacter}
\alias{parseEvalCharacterMany}
\alias{parseEvalNumeric}
\alias{parseEvalNumericMany}
\alias{parseTreeSubstitute}
\alias{passThroughOperators}
\alias{pasteSemicolon}
\alias{phi}
\alias{plotABS}
\alias{plotMinMeanAll}
\alias{pointAt}
\alias{populateCopierVector}
\alias{populateIndexedNodeInfoTable}
\alias{populateManyModelValuesMapAccess}
\alias{populateManyModelVarMapAccess}
\alias{populateNodeFxnVec}
\alias{populateNodeFxnVecNew}
\alias{posteriorClass}
\alias{posteriorSummaryComparisonComponent}
\alias{pow}
\alias{pq_exp_nimble_keywordInfo}
\alias{pq_gamma_keywordInfo}
\alias{pqAvail}
\alias{pqDists}
\alias{printMinTimeABS}
\alias{probit}
\alias{processDistributionAliases}
\alias{processKeyword}
\alias{processKeywordCodeMemberFun}
\alias{processKeywords_recurse}
\alias{processModelFile}
\alias{processMonitorNames}
\alias{processNonParseableCode}
\alias{processVarBlock}
\alias{projectNameCreator}
\alias{putCodeLinesInBrackets}
\alias{qp_dist_keywordInfo}
\alias{R_setDoublePtrFromSinglePtr}
\alias{R_setOnePtrVectorOfPtrs}
\alias{R_setPtrVectorOfPtrs}
\alias{rCalcDiffNodes}
\alias{rCalcNodes}
\alias{RCfunction}
\alias{RCfunctionCompileClass}
\alias{RCfunctionDef}
\alias{RCfunInfoClass}
\alias{rcFunLabelMaker}
\alias{RCfunProcessing}
\alias{RCvirtualFunProcessing}
\alias{recurseSetSizes}
\alias{reduceDF}
\alias{reductionBinaryOperators}
\alias{reductionBinaryOperatorsEither}
\alias{reductionUnaryOperators}
\alias{reductionUnaryOperatorsArray}
\alias{reductionUnaryOperatorsEither}
\alias{removeColonOperator}
\alias{removeExprClassLayer}
\alias{removeIndexing}
\alias{removeLayerHandler}
\alias{removeLogProbName}
\alias{removeNonScalarElementsFromList}
\alias{removeSingleModelValuesAccess}
\alias{removeSingleVarAccess}
\alias{removeVariableFromEnv}
\alias{replaceAllCodeSuccessfully}
\alias{replaceConstantsRecurse}
\alias{replaceDistributionAliases}
\alias{replaceDistributionAliasesNameOnly}
\alias{replacementHandler}
\alias{replaceWhatWeCan}
\alias{reprioritizeColonOperator}
\alias{resize_NumberedObjects}
\alias{resizeManyModelValuesAccessor}
\alias{resizeManyModelVarAccessor}
\alias{rexp_nimble_keywordInfo}
\alias{rFunHandler}
\alias{rgamma_keywordInfo}
\alias{rGetLogProbsNodes}
\alias{RMakeCustomModelClass}
\alias{Rmatrix2mvOneVar}
\alias{rmFunHandler}
\alias{Rname2CppName}
\alias{RNumericList}
\alias{rop}
\alias{RparseTree2ExprClasses}
\alias{rSimNodes}
\alias{sampler_BASE}
\alias{sampler_binary}
\alias{sampler_ess}
\alias{samplerSpec}
\alias{scalarOutputTypes}
\alias{seqAlongHandler}
\alias{setAndCalculateDiff}
\alias{setArg}
\alias{setAsRowOrCol}
\alias{setBoolValue}
\alias{setCaller}
\alias{setCharacterValue}
\alias{setCharacterVectorValue}
\alias{setCNumericListValue}
\alias{setDoublePtrFromSinglePtr}
\alias{setDoubleValue}
\alias{setIntValue}
\alias{setMVElement}
\alias{setNimbleOption}
\alias{setNimValues}
\alias{setOnePtrVectorOfPtrs}
\alias{setPtrVectorOfPtrs}
\alias{setupCodeTemplateClass}
\alias{setValues}
\alias{setValuesAccess}
\alias{SEXPscalarConvertFunctions}
\alias{simulate_keywordInfo}
\alias{singleBracket_keywordInfo}
\alias{singleModelIndexAccess_SetupTemplate}
\alias{singleModelValuesAccess}
\alias{singleModelValuesAccessor_SetupTemplate}
\alias{singleVarAccess}
\alias{singleVarAccess_SetupTemplate}
\alias{sizeasDoublePtr}
\alias{sizeAsRowOrCol}
\alias{sizeAssign}
\alias{sizeAssignAfterRecursing}
\alias{sizeBinaryCwise}
\alias{sizeBinaryCwiseLogical}
\alias{sizeBinaryReduction}
\alias{sizeBinaryUnaryCwise}
\alias{sizecallC}
\alias{sizeCalls}
\alias{sizeChainedCall}
\alias{sizeColonOperator}
\alias{sizeDimOperator}
\alias{sizeFor}
\alias{sizeforceEigenize}
\alias{sizeGetParam}
\alias{sizeIndexingBracket}
\alias{sizeInsertIntermediate}
\alias{sizemap}
\alias{sizeMatrixMult}
\alias{sizeMatrixSquareReduction}
\alias{sizemvAccessBracket}
\alias{sizeNFvar}
\alias{sizeNimArrayGeneral}
\alias{sizeNimbleFunction}
\alias{sizeOneEigenCommand}
\alias{sizeRCfunction}
\alias{sizeResizeNoPtr}
\alias{sizeReturn}
\alias{sizeRmultivarFirstArg}
\alias{sizeRMVNorm}
\alias{sizeScalar}
\alias{sizeScalarModelOp}
\alias{sizeScalarRecurse}
\alias{sizeSetSize}
\alias{sizeSimulate}
\alias{sizeSolveOp}
\alias{sizeSwitch}
\alias{sizeTranspose}
\alias{sizeUnaryCwise}
\alias{sizeUnaryCwiseSquare}
\alias{sizeUnaryNonaryCwise}
\alias{sizeUnaryReduction}
\alias{sizeUndefined}
\alias{sizeValues}
\alias{sizeVoidPtr}
\alias{specificCallHandlers}
\alias{specificCallReplacements}
\alias{spectrum0.ar.big}
\alias{splitCompletionForOrigNodes}
\alias{splitVertexIDsToElementIDs}
\alias{splitVertices}
\alias{stochNodeInit}
\alias{stripArgs}
\alias{stripPrefix}
\alias{symbolBase}
\alias{symbolBasic}
\alias{symbolCopierVector}
\alias{symbolDouble}
\alias{symbolEigenMap}
\alias{symbolGetParamInfo}
\alias{symbolIndexedNodeInfoTable}
\alias{symbolInt}
\alias{symbolInternalType}
\alias{symbolMemberFunction}
\alias{symbolModel}
\alias{symbolModelValues}
\alias{symbolModelValuesAccessorVector}
\alias{symbolModelVariableAccessorVector}
\alias{symbolNimArrDoublePtr}
\alias{symbolNimbleFunction}
\alias{symbolNimbleFunctionList}
\alias{symbolNimPtrList}
\alias{symbolNodeFunctionVector}
\alias{symbolNumericList}
\alias{symbolOptimReadyFunction}
\alias{symbolPtr}
\alias{symbolString}
\alias{symbolTable}
\alias{symbolTable2cppVars}
\alias{symbolVecNimArrPtr}
\alias{symbolVoidPtr}
\alias{symTypeFromSymTab}
\alias{timeComparisonComponent}
\alias{toSEXPscalarConvertFunctions}
\alias{unaryOperators}
\alias{unaryOrNonaryOperators}
\alias{UseLibraryMakevars}
\alias{v}
\alias{valueInCompiledNimbleFunction}
\alias{values_keywordInfo}
\alias{valuesAccessorVector}
\alias{varAndIndices2flatIndex}
\alias{varAndIndices2mapParts}
\alias{varInfoClass}
\alias{vectorIndex_2_flat}
\alias{virtualNFprocessing}
\alias{virtualNodeFunctionDefinitions}
\alias{whyInvalid}
\alias{wrapAsNumeric}
\alias{writeCode}
\alias{[,numberedModelValuesAccessors-method}
\alias{[<-,numberedModelValuesAccessors-method}
\alias{[,numberedObjects-method}
\alias{[<-,numberedObjects-method}
\alias{[[,CNumericList-method}
\alias{[[<-,CNumericList-method}
\alias{[[,RNumericList-method}
\alias{[[<-,RNumericList-method}
\alias{[[,nimPointerList-method}
\alias{[[<-,nimPointerList-method}
\alias{[[<-,nimbleFunctionList-method}
\alias{[[,distributionsClass-method}
\alias{[[,conjugacyRelationshipsClass-method}
\alias{as.matrix.modelValuesBaseClass-Class}
\alias{autoBlockClass-Class}
\alias{BUGScontextClass-Class}
\alias{BUGSsingleContextClass-Class}
\alias{CmultiNimbleFunctionClass-Class}
\alias{conjugacyClass-Class}
\alias{conjugacyRelationshipsClass-Class}
\alias{cppBUGSmodelClass-Class}
\alias{cppCodeFileClass-Class}
\alias{cppCPPfileClass-Class}
\alias{cppHfileClass-Class}
\alias{cppModelValuesClass-Class}
\alias{cppNamedObjectsClass-Class}
\alias{cppNimbleFunctionClass-Class}
\alias{cppProjectClass-Class}
\alias{cppVirtualNimbleFunctionClass-Class}
\alias{dependentClass-Class}
\alias{distClass-Class}
\alias{distributionsClass-Class}
\alias{exprClass-Class}
\alias{exprTypeInfoClass-Class}
\alias{indexedNodeInfoTableClass-Class}
\alias{keywordInfoClass-Class}
\alias{MakeCustomModelClass-Class}
\alias{makeCustomModelValuesClass-Class}
\alias{mapsClass-Class}
\alias{modelDefInfoClass-Class}
\alias{mvInfoClass-Class}
\alias{nfCompilationInfoClass-Class}
\alias{nimbleGraphClass-Class}
\alias{nimbleProjectClass-Class}
\alias{posteriorClass-Class}
\alias{RCfunctionCompileClass-Class}
\alias{RCfunInfoClass-Class}
\alias{RMakeCustomModelClass-Class}
\alias{setupCodeTemplateClass-Class}
\alias{varInfoClass-Class}
\description{
Functions and classes used internally in NIMBLE and not expected to be called directly by users. For the moment many functions and classes not intended for direct use are exported because they are used within Reference Class methods for classes programmatically generated by NIMBLE.
}
\author{NIMBLE Development Team}