\name{nimble-internal}
\title{Functions and Classes Internal to NIMBLE}
\alias{as.list.modelValuesBaseClass-Class}
\alias{as.matrix.modelValuesBaseClass-Class}
\alias{autoBlockClass-Class}
\alias{BUGScontextClass-Class}
\alias{BUGSsingleContextClass-Class}
\alias{CmultiNimbleFunctionClass-Class}
\alias{CmultiNimbleListClass-Class}
\alias{CmultiNimbleObjClass-Class}
\alias{conjugacyClass-Class}
\alias{conjugacyRelationshipsClass-Class}
\alias{copyExprClass-Class}
\alias{cppBUGSmodelClass-Class}
\alias{cppCodeFileClass-Class}
\alias{cppCPPfileClass-Class}
\alias{cppHfileClass-Class}
\alias{cppModelValuesClass-Class}
\alias{cppNamedObjectsClass-Class}
\alias{cppNimbleClassClass-Class}
\alias{cppNimbleFunctionClass-Class}
\alias{cppNimbleListClass-Class}
\alias{cppProjectClass-Class}
\alias{cppVirtualNimbleFunctionClass-Class}
\alias{dependentClass-Class}
\alias{distClass-Class}
\alias{distributionsClass-Class}
\alias{eigenize_nimbleNullaryClass-Class}
\alias{exprClass-Class}
\alias{exprTypeInfoClass-Class}
\alias{findClass-Class}
\alias{findMethodsInExprClass-Class}
\alias{getClass-Class}
\alias{getRefClass-Class}
\alias{indexedNodeInfoTableClass-Class}
\alias{isClass-Class}
\alias{isSealedClass-Class}
\alias{isVirtualClass-Class}
\alias{isXS3Class-Class}
\alias{keywordInfoClass-Class}
\alias{MakeCustomModelClass-Class}
\alias{makeCustomModelValuesClass-Class}
\alias{makeStaticInitClass-Class}
\alias{mapsClass-Class}
\alias{modelDefInfoClass-Class}
\alias{mvInfoClass-Class}
\alias{nfCompilationInfoClass-Class}
\alias{nimbleGraphClass-Class}
\alias{nimbleListDefClass-Class}
\alias{nimbleProjectClass-Class}
\alias{nlCompilationInfoClass-Class}
\alias{posteriorClass-Class}
\alias{promptClass-Class}
\alias{R6Class-Class}
\alias{RCfunctionCompileClass-Class}
\alias{RCfunInfoClass-Class}
\alias{removeClass-Class}
\alias{resetClass-Class}
\alias{RMakeCustomModelClass-Class}
\alias{S3Class-Class}
\alias{sealClass-Class}
\alias{setClass-Class}
\alias{setOldClass-Class}
\alias{setRefClass-Class}
\alias{setupCodeTemplateClass-Class}
\alias{showClass-Class}
\alias{singleModelValuesAccessClass-Class}
\alias{varInfoClass-Class}
\alias{[,numberedModelValuesAccessors-method}
\alias{[<-,numberedModelValuesAccessors-method}
\alias{[,numberedObjects-method}
\alias{[<-,numberedObjects-method}
\alias{[[,CNumericList-method}
\alias{[[<-,CNumericList-method}
\alias{[[,RNumericList-method}
\alias{[[<-,RNumericList-method}
\alias{[[,nimPointerList-method}
\alias{[[<-,nimPointerList-method}
\alias{[[<-,nimbleFunctionList-method}
\alias{[,distributionsClass-method}
\alias{[[,distributionsClass-method}
\alias{[[,conjugacyRelationshipsClass-method}
\alias{dsqrtinvgamma}
\alias{rsqrtinvgamma}
\alias{is.Cmodel}
\alias{is.Cnf}
\alias{is.Rmodel}
\alias{is.model}
\alias{nf_preProcessMemberDataObject}
\alias{samplesSummary}
\alias{calc_dmnormConjugacyContributions}
\alias{calc_dmnormAltParams}
\alias{calc_dwishAltParams}
\alias{calc_dcatConjugacyContributions}
\alias{CAR_calcM}
\alias{CAR_calcC}
\alias{CAR_calcCmatrix}
\alias{CAR_calcEVs2}
\alias{CAR_calcEVs3}
\alias{getNodeFunctionIndexedInfo}
\alias{singleModelValuesAccess}
\alias{getNimbleProject}
\alias{cc_getNodesInExpr}
\alias{calcAdaptationFactor}
\alias{is.nfGenerator}
\alias{nimbleInternalFunctions}
\alias{nimbleUserNamespace}
\description{
Functions and classes used internally in NIMBLE and not expected to be called directly by users. Some functions and classes not intended for direct use are documented and/or exported because they are used within Reference Class methods for classes programmatically generated by NIMBLE.
}
\author{NIMBLE Development Team}