% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79eeb11e58b
=======
% Please edit documentation in RtmpLwENQY/file616168dd5c38
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb36b0aae49
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b04f887170
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d671cdd5c4
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file4976be42076
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78717189d576
>>>>>>> added line to remove .o files after create
\name{nimbleCode}
\alias{nimbleCode}
\title{Turn BUGS model code into an object for use in \code{nimbleModel} or \code{readBUGSmodel}}
\usage{
nimbleCode(code)
}
\arguments{
  \item{code}{expression providing the code for the model}}\description{
Simply keeps model code as an R call object, the form
needed by \link{nimbleModel} and optionally usable by
\link{readBUGSmodel}
}
\details{
It is equivalent to use the R function \code{quote}.
\code{nimbleCode} is simply provided as a more readable
alternative for NIMBLE users not familiar with \code{quote}.
}
\examples{
code <- nimbleCode({
    x ~ dnorm(mu, sd = 1)
    mu ~ dnorm(0, sd = prior_sd)
})
}
\author{
Daniel Turek
}

