% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmphInSZr/filecb5b1ba2078d
\docType{class}
\name{nimbleFunctionBase-class}
\alias{nimbleFunctionBase}
\alias{nimbleFunctionBase-class}
\title{Class \code{nimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

