% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpVY8pk7/file3afa216412b6
\docType{class}
\name{nimbleFunctionBase-class}
\alias{nimbleFunctionBase}
\alias{nimbleFunctionBase-class}
\title{Class \code{nimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

