% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp8XMcGu/file9e31665ccb79
\docType{class}
\name{nimbleFunctionBase-class}
\alias{nimbleFunctionBase}
\alias{nimbleFunctionBase-class}
\title{Class \code{nimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

