% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_core.R
\docType{class}
\name{nimbleFunctionBase-class}
\alias{nimbleFunctionBase-class}
\alias{nimbleFunctionBase}
\title{Class \code{nimbleFunctionBase}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

