% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpHcXJC5/file6550613c639a
\docType{class}
\name{nimbleFunctionList-class}
\alias{nimbleFunctionList}
\alias{nimbleFunctionList-class}
\title{Create a list of nimbleFunctions}
\description{
Create an empty list of nimbleFunctions that all will inherit from a base class.
}
\details{
See the User Manual for information about creating and populating a \code{nimbleFunctionList}.
}
\author{
NIMBLE development team
}

