% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_nimbleFunctionList.R
\docType{class}
\name{nimbleFunctionList-class}
\alias{nimbleFunctionList-class}
\alias{nimbleFunctionList}
\title{Create a list of nimbleFunctions}
\description{
Create an empty list of nimbleFunctions that all will inherit from a base class.
}
\details{
See the User Manual for information about creating and populating a \code{nimbleFunctionList}.
}

\author{
NIMBLE development team
}
