% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_core.R
\name{nimbleFunctionVirtual}
\alias{nimbleFunctionVirtual}
\title{create a virtual nimbleFunction, a base class for other nimbleFunctions}
\usage{
nimbleFunctionVirtual(
  contains = NULL,
  run = function() { },
  methods = list(),
  name = NA,
  methodControl = list()
)
}
\arguments{
\item{contains}{Not yet functional}

\item{run}{A NIMBLE function that will only be used to inspect its argument types and returnType.}

\item{methods}{An optional named list of NIMBLE functions that will also only be used for inspecting argument types and returnTypes.}

\item{name}{An optional name used internally by the NIMBLE compiler.  This is usually omitted and NIMBLE provides one.}

\item{methodControl}{An optional list that allows specification of methods with defaults.}
}
\value{
An object that can be passed as the \code{contains} argument to \code{nimbleFunction} or as the argument to \code{nimbleFunctionList}
}
\description{
define argument types and returnType for the \code{run} function and any \code{methods}, to be used in the \code{contains} argument of \code{nimbleFunction}
}
\details{
See the NIMBLE \href{https://r-nimble.org/html_manual/cha-welcome-nimble.html}{User Manual} section on nimbleFunctionLists for explanation of how to use a virtual nimbleFunction.
}
\seealso{
\code{\link{nimbleFunction}}
}
\author{
NIMBLE development team
}
