% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{nimbleOptions}
\alias{nimbleOptions}
\title{options used for NIMBLE package}
\format{\preformatted{List of 10
 $ convertSingleVectorsToScalarsInSetupArgs  : logi TRUE
 $ messagesWhenBuildingOrFinalizingCppObjects: logi FALSE
 $ indexDrop                                 : logi TRUE
 $ debugRCfunProcessing                      : logi FALSE
 $ debugNFProcessing                         : logi FALSE
 $ compileAltParamFunctions                  : logi TRUE
 $ verifyConjugatePosteriors                 : logi TRUE
 $ includeCPPdists                           : logi TRUE
 $ processBackwardsModelIndexRanges          : logi FALSE
 $ prioritizeColonLikeBUGS                   : logi TRUE
}}
\usage{
nimbleOptions
}
\description{
options used for NIMBLE package
}
\details{
These options are for development use at this point.
}
\keyword{datasets}

