% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpLwENQY/file61612b823b49
\name{nimbleOptions}
\alias{nimbleOptions}
\title{NIMBLE Options Settings}
\usage{
nimbleOptions(...)
}
\arguments{
  \item{...}{any options to be defined as one or more 'name
  = value' pairs. Options can also be passed by giving a
  single unnamed argument that is a named list.}
}
\value{
When invoked with no arguments, a list with the current
values of all options.
}
\description{
Allow the user to set and examine a variety of global
_options_ that affect the way in which NIMBLE operates
}
\details{
\code{nimbleOptions} mimics \code{options}. Invoking
\code{nimbleOptions()} with no arguments returns a list with
the current values of the options. To access the value of a
single option, one should use \code{getNimbleOption()}.
}
\examples{
nimbleOptions(verifyConjugatePosteriors = FALSE)
}
\author{
Christopher Paciorek
}

