% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_implementations.R
\name{nimDerivs}
\alias{nimDerivs}
\title{Nimble Derivatives}
\usage{
nimDerivs(nimFxn = NA, order = nimC(0, 1, 2))
}
\arguments{
\item{nimFxn}{a call to a \code{nimbleFunction} method with arguments included.}

\item{order}{an integer vector with values within the set {0, 1, 2}, corresponding to whether the function value, gradient, and Hessian should be returned respectively.}
}
\description{
EXPERIMENTAL Computes the value, gradient, and Hessian of a given  \code{nimbleFunction} method.  The R version is currently unimplemented.
}
