% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
=======
% Please edit documentation in RtmpLwENQY/file616162260d5d
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3359ae819
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0b253e8a
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6320f3e98
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497656247499
>>>>>>> updated suite and autoblock roxygen
=======
% Please edit documentation in Rtmp60KOwh/file78712a8c67
>>>>>>> added line to remove .o files after create
=======
% Please edit documentation in Rtmpgi4ISD/file4c3443f77500
>>>>>>> various cleanups of roxygen stuff
=======
% Please edit documentation in RtmpkxJCni/file2c0b48daab2c
>>>>>>> removed top level tests dir
=======
% Please edit documentation in RtmpaEoKPd/file20ba30b2868d
>>>>>>> fixed up some roxygen stuff and
=======
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
>>>>>>> added man files from devel to ease rebase
=======
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
>>>>>>> 6f0bc2e7a0885e03527d78f21be97fad5c54074a
\name{nimDim}
\alias{nimDim}
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nimDim(obj)
}
\arguments{
<<<<<<< HEAD
<<<<<<< HEAD
\item{obj}{objects for which the sizes are requested}}\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.  It is useful to have this function that treats a vector similarly to a matrix or array.  Works in R and NIMBLE.  In NIMBLE \code{dim} is identical to \code{nimDim}, not to R's \code{dim}
=======
=======
>>>>>>> 6f0bc2e7a0885e03527d78f21be97fad5c54074a
  \item{obj}{objects for which the sizes are requested}}\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.
It is useful to have this function that treats a vector
similarly to a matrix or array. Works in R and NIMBLE. In
NIMBLE \code{dim} is identical to \code{nimDim}, not to R's
\code{dim}
>>>>>>> added man files from devel to ease rebase
}
\examples{
x <- rnorm(4)
dim(x)
nimDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nimDim(y)
}
\author{
NIMBLE development team
}

