% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
% Please edit documentation in Rtmpzdod9z/file79ee404fec36
=======
% Please edit documentation in RtmpLwENQY/file616162260d5d
>>>>>>> updated S3 methods to follow
=======
% Please edit documentation in Rtmps9T35v/file3fb3359ae819
>>>>>>> various updates to roxygen
=======
% Please edit documentation in RtmpnG0e98/file11b0b253e8a
>>>>>>> took Makefiles out of
=======
% Please edit documentation in RtmpAPakJJ/file50d6320f3e98
>>>>>>> fixed error in MCMCspec roxygen example
=======
% Please edit documentation in RtmpmO7lLg/file497656247499
>>>>>>> updated suite and autoblock roxygen
\name{nimDim}
\alias{nimDim}
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nimDim(obj)
}
\arguments{
  \item{obj}{objects for which the sizes are requested}}\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.
It is useful to have this function that treats a vector
similarly to a matrix or array. Works in R and NIMBLE. In
NIMBLE \code{dim} is identical to \code{nimDim}, not to R's
\code{dim}
}
\examples{
x <- rnorm(4)
dim(x)
nimDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nimDim(y)
}
\author{
NIMBLE development team
}

