% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{nimDim}
\alias{dim}
\alias{nimDim}
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nimDim(obj)
}
\arguments{
\item{obj}{objects for which the sizes are requested}
}
\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.  It is useful to have this function that treats a vector similarly to a matrix or array.  Works in R and NIMBLE.  In NIMBLE \code{dim} is identical to \code{nimDim}, not to R's \code{dim}
}
\examples{
x <- rnorm(4)
dim(x)
nimDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nimDim(y)

}
\author{
NIMBLE development team
}

