% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in Rtmp8XMcGu/file9e311a42e2ff
=======
% Please edit documentation in RtmpOW946u/file1481d45fb9821
>>>>>>> reduceExport
\name{nimInteger}
\alias{integer}
\alias{nimInteger}
\title{creates an integer vector}
\usage{
nimInteger(length = 0, value = 0, init = TRUE)
}
\arguments{
\item{length}{the length of the vector (default = 0)}

<<<<<<< HEAD
\item{value}{the initial value for each element of the vector (default = 0L)}

\item{init}{logical, whether to initialize elements of the vector (default = TRUE)}
=======
\item{value}{the initial value for each element of the vector (default = 0L).  Only used if \code{init} is \code{TRUE}.}

\item{init}{logical, whether to initialize elements of the vector (default = TRUE).}
>>>>>>> reduceExport
}
\description{
creates an integer vector
}
\details{
When used in a \code{nimbleFunction} (in \code{run} or other member function), \code{integer} is a synonym for \code{nimInteger}.  When used with only the \code{length} argument, this behaves similarly to R's \code{integer} function.  NIMBLE provides additional arguments to control the initialization value and whether or not initialization will occur.  Using \code{init=FALSE} when initialization is not necessary can make compiled nimbleFunctions a bit faster.
}
\author{
Daniel Turek
}
\seealso{
\link{numeric} \link{matrix} \link{array}
}

