% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmphInSZr/filecb5b764d5a4c
\name{nimInteger}
\alias{integer}
\alias{nimInteger}
\title{creates an integer vector}
\usage{
nimInteger(length = 0, value = 0, init = TRUE)
}
\arguments{
\item{length}{the length of the vector (default = 0)}

\item{value}{the initial value for each element of the vector (default = 0L)}

\item{init}{logical, whether to initialize elements of the vector (default = TRUE)}
}
\description{
creates an integer vector
}
\author{
Daniel Turek
}
\seealso{
\link{numeric} \link{matrix} \link{array}
}

