% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d45fb9821
\name{nimMatrix}
\alias{matrix}
\alias{nimMatrix}
\title{creates a matrix}
\usage{
nimMatrix(value = 0, nrow = 1, ncol = 1, init = TRUE, type = "double")
}
\arguments{
\item{value}{the initial value for each element of the matrix (default = 0)}

\item{nrow}{the number of rows in the matrix (default = 1)}

\item{ncol}{the number of columns in the matrix (default = 1)}

\item{init}{logical, whether to initialize elements of the matrix (default = TRUE)}

\item{type}{character representing the data type, i.e. 'double' or 'integer' (default = 'double')}
}
\description{
creates a matrix
}
\details{
When used in a \code{nimbleFunction} (in \code{run} or other member function), \code{matrix} is a synonym for \code{nimMatrix}.  When used with only the first three arguments, this behaves similarly to R's \code{matrix} function.  NIMBLE provides additional arguments to control the initialization value, whether or not initialization will occur, and the type of scalar elements.  Using \code{init=FALSE} when initialization is not necessary can make compiled nimbleFunctions a bit faster.
}
\author{
Daniel Turek
}
\seealso{
\link{numeric} \link{integer} \link{array}
}

