% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmp29bxL0/file74da464eb95f
\name{nimMatrix}
\alias{matrix}
\alias{nimMatrix}
\title{creates a matrix}
\usage{
nimMatrix(value = 0, nrow = 1, ncol = 1, init = TRUE, type = "double")
}
\arguments{
\item{value}{the initial value for each element of the matrix (default = 0)}\item{nrow}{the number of rows in the matrix (default = 1)}\item{ncol}{the number of columns in the matrix (default = 1)}\item{init}{logical, whether to initialize elements of the matrix (default = TRUE)}\item{type}{character representing the data type, i.e. 'double' or 'integer' (default = 'double')}}\description{
creates a matrix
}
\author{
Daniel Turek#'

Daniel Turek
}
\seealso{
\link{numeric} \link{integer} \link{array}
}

