% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpowSbmr/filed7686395a88c
\name{nimNumeric}
\alias{nimNumeric}
\alias{numeric}
\title{creates a numeric vector}
\usage{
nimNumeric(length = 0, value = 0, init = TRUE)
}
\arguments{
\item{length}{the length of the vector (default = 0)}

\item{value}{the initial value for each element of the vector (default = 0)}

\item{init}{logical, whether to initialize elements of the vector (default = TRUE)}
}
\description{
creates a numeric vector
}
\author{
Daniel Turek
}
\seealso{
\link{integer} \link{matrix} \link{array}
}

