% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp6hGOZh/file12a1f7e20bd7a
\name{nimPrint}
\alias{nimPrint}
\title{print function for use in nimbleFunctions, where it is identical to \code{print}, but not R's \code{print}.}
\usage{
nimPrint(...)
}
\arguments{
\item{...}{an abitrary set of arguments that will be printed in sequence}
}
\description{
print function for use in nimbleFunctions, where it is identical to \code{print}.  Works in R or NIMBLE, not quite identically.
}
\details{
The keyword \code{print} in nimbleFunction run-time code will be automatically turned into \code{nimPrint}.  This is a function that prints its arguments int order using \code{cat} in R, or using \code{std::cout} in C++ code generated from compiling nimbleFunctions.
Non-scalar numeric objects can be included, although their output will be formatted slightly different in uncompiled and compiled nimbleFunctions.
}
\examples{
ans <- matrix(1:4, nrow = 2) ## R code, not NIMBLE code
nimPrint('Answer is ', ans, '\\n') ## would work in R or NIMBLE

}

