% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmprbuFLj/file2aa178d90550
\name{nimPrint}
\alias{nimPrint}
\title{print function for use in nimbleFunctions}
\usage{
nimPrint(...)
}
\arguments{
\item{...}{an abitrary set of arguments that will be printed in sequence.}
}
\description{
print function for use in nimbleFunctions
}
\details{
The keyword \code{print} in nimbleFunction run-time code will be automatically turned into \code{nimPrint}.  This is a function that prints its arguments in order using \code{cat} in R, or using \code{std::cout} in C++ code generated by compiling nimbleFunctions.
Non-scalar numeric objects can be included, although their output will be formatted slightly different in uncompiled and compiled nimbleFunctions.
}
\examples{
ans <- matrix(1:4, nrow = 2) ## R code, not NIMBLE code
nimPrint('Answer is ', ans) ## would work in R or NIMBLE
}
\seealso{
\code{\link{cat}}
}

