% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nimPrint}
\alias{nimPrint}
\alias{print}
\title{print function for use in nimbleFunctions}
\usage{
nimPrint(...)
}
\arguments{
\item{...}{an abitrary set of arguments that will be printed in sequence.}
}
\description{
print function for use in nimbleFunctions
}
\details{
The keyword \code{print} in nimbleFunction run-time code will be automatically turned into \code{nimPrint}.  This is a function that prints its arguments in order using \code{cat} in R, or using \code{std::cout} in C++ code generated by compiling nimbleFunctions.
Non-scalar numeric objects can be included, although their output will be formatted slightly different in uncompiled and compiled nimbleFunctions.

For numeric values, the number of digits printed is controlled by the system option \code{nimbleOptions('digits')}.
}
\examples{
ans <- matrix(1:4, nrow = 2) ## R code, not NIMBLE code
nimPrint('Answer is ', ans) ## would work in R or NIMBLE

}
\seealso{
\code{\link{cat}}
}
