% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_Rexecution.R
\name{nimStop}
\alias{nimStop}
\alias{stop}
\title{Halt execution of a nimbleFunction function method.  Part of the NIMBLE language}
\usage{
nimStop(msg)
}
\arguments{
\item{msg}{Character object to be output as an error message}
}
\description{
Halt execution of a nimbleFunction function method.  Part of the NIMBLE language
}
\details{
The NIMBLE \code{stop} is similar to the native R \code{stop}, but it takes only one argument, the error message to be output.  During uncompiled NIMBLE execution, \code{nimStop} simply calls R's stop funtion. During compiled execution it calls the error function from the R headers.  \code{stop} is an alias for \code{nimStop} in the NIMBLE language
}
\author{
Perry de Valpine
}
