% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmptLDKMD/file136533acb599
\docType{class}
\name{nodeInfoClass-class}
\alias{nodeInfoClass}
\alias{nodeInfoClass-class}
\title{nodeInfoClass contains information for a node}
\description{
nodeInfoClass contains information for a node
}

