% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpoHax5j/file73321f00f1e1
\docType{class}
\name{nodeInfoClass-class}
\alias{nodeInfoClass}
\alias{nodeInfoClass-class}
\title{nodeInfoClass contains information for a node}
\description{
nodeInfoClass contains information for a node
}

