% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTransform.R
\name{parameterTransform}
\alias{parameterTransform}
\title{Automated transformations of model nodes to unconstrained scales}
\usage{
parameterTransform(model, nodes)
}
\arguments{
\item{model}{A nimble model object.}

\item{nodes}{Vector of node names to which automated transformation will e applied.}
}
\description{
Automated transformations of model nodes to unconstrained scales
}
\author{
Daniel Turek
}
