% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_comparisons.R
\name{rename_MCMC_comparison_method}
\alias{rename_MCMC_comparison_method}
\title{Rename a method in an object returned by compareMCMCs}
\usage{
rename_MCMC_comparison_method(oldname, newname, comparison)
}
\arguments{
\item{oldname}{(character) Existing name for the method (which would have been determined from inputs to \code{\link{compareMCMCs}}).}

\item{newname}{(character) New name for the method}

\item{comparison}{An object returned by compareMCMCs.}
}
\value{
An object in the same format as \code{comparison}.
}
\description{
Switches one label for a new one in the timing, efficiency, summary, and samples (if present) elements of a compareMCMCs result.
}
\details{
One of the main purposes of renaming a method is that the method names are used in html results generated by \code{\link{make_MCMC_comparison_pages}}.
}
\seealso{
\code{\link{compareMCMCs}}, code{\link{combine_MCMC_comparison_results}}, \code{\link{make_MCMC_comparison_pages}}, \code{\link{reshape_comparison_results}}.
}
