% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpnJPcqB/file3e4950e12ca4
\name{resize}
\alias{resize}
\title{Resizes a modelValues object}
\usage{
resize(container, k)
}
\arguments{
\item{container}{modelValues object}\item{k}{number of rows that modelValues is set to}}\description{
Adds or removes rows to a modelValues object. If rows are added to a modelValues object, the default values are \code{NA}. Works in both R and NIMBLE.
}
\details{
See the User Manual or \code{help(modelValuesBaseClass)} for infomation about modelValues objects
}
\examples{
mvSpec <- modelValuesSpec(vars = c('a', 'b'),
             types = c('double', 'double'),
             sizes = list(a = 1, b = c(2,2) ) )
mv <- modelValues(mvSpec)
as.matrix(mv)
resize(mv, 3)
as.matrix(mv)
}
\author{
Clifford Anderson-Bergman
}

