% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{resize}
\alias{resize}
\title{Resizes a modelValues object}
\usage{
resize(container, k)
}
\arguments{
\item{container}{modelValues object}

\item{k}{number of rows that modelValues is set to}
}
\description{
Adds or removes rows to a modelValues object. If rows are added to a modelValues object, the default values are \code{NA}. Works in both R and NIMBLE.
}
\details{
See the \href{https://r-nimble.org/html_manual/cha-welcome-nimble.html}{User Manual} or \code{help(modelValuesBaseClass)} for infomation about modelValues objects
}
\examples{
mvConf <- modelValuesConf(vars = c('a', 'b'),
             types = c('double', 'double'),
             sizes = list(a = 1, b = c(2,2) ) ) 
mv <- modelValues(mvConf)
as.matrix(mv)
resize(mv, 3)
as.matrix(mv)

}
\author{
Clifford Anderson-Bergman
}
