% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmphInSZr/filecb5b4fd2c90c
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\alias{RmodelBaseClass-class}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

