% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpHcXJC5/file65505ce4ee5e
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\alias{RmodelBaseClass-class}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

