% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmph6b7yW/file26124374d5d3
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\alias{RmodelBaseClass-class}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

