% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpOrtTO2/file33513ab0de39
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\alias{RmodelBaseClass-class}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

