% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUGS_model.R
\docType{class}
\name{RmodelBaseClass-class}
\alias{RmodelBaseClass-class}
\alias{RmodelBaseClass}
\title{Class \code{RmodelBaseClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

