% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpPe12xI/file6551506edef4
\name{run.time}
\alias{run.time}
\title{Time execution of NIMBLE code}
\usage{
run.time(code)
}
\arguments{
\item{code}{code to be timed}
}
\description{
Time execution of NIMBLE code
}
\details{
Function for use in nimbleFunction run code; when nimbleFunctions are run in R, this simply wraps \code{system.time}.
}
\author{
NIMBLE Development Team
}

