% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in RtmptLDKMD/file13656e6afbaa
\name{setSize}
\alias{setSize}
\title{set the size of a numeric variable in NIMBLE}
\usage{
setSize(numList, row, d1 = NA, d2 = NA, d3 = NA)
}
\arguments{
  \item{numList}{This is the object to be resized}

  \item{row}{Optional argument that is not currently used}

  \item{d1}{first size}

  \item{d2}{second size}

  \item{d3}{third size}
}
\description{
set the size of a numeric variable in NIMBLE.  This works
in R and NIMBLE, but in R it usually has no effect.
}
\details{
This function is part of the NIMBLE language.  Its purpose
is to explicitly resize a multivariate object (vector,
matrix or array), currently up to 3 dimensions.  Explicit
resizing is not needed when an entire object is assigned
to.  For example, in \code{Y <- A \%*\% B}, where A and B
are matrices, \code{Y} will be resized automatically.
Explicit resizing is necessary when assignment will be by
indexed elements or blocks, or prior to a call to
getValues.

When executed in R, this works but usually has no effect,
because R objects will be dynamically resized anyway.  It
is important for it to work so that nimbleFunctions can be
executed in R.

At the moment the version that executes in R is primarily
defined for numericLists, which are a type to be provided
in future versions of NIMBLE.
}
\author{
NIMBLE development team
}

