% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{setupMargNodes}
\alias{setupMargNodes}
\title{Process model to organize nodes for marginalization as by Laplace or adaptive
Gauss-Hermite quadrature (AGHQuad) approximations}
\usage{
setupMargNodes(
  model,
  paramNodes,
  randomEffectsNodes,
  calcNodes,
  calcNodesOther,
  split = TRUE,
  warn = TRUE
)
}
\description{
Process model to organize nodes for marginalization as by Laplace or adaptive
Gauss-Hermite quadrature (AGHQuad) approximations
}
