% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpoXeYG3/file5fe4ef6ca39
\docType{class}
\name{singleModelValuesAccessClass-class}
\alias{singleModelValuesAccessClass}
\alias{singleModelValuesAccessClass-class}
\title{Class \code{singleModelValuesAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

