% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d8c0c26c
\docType{class}
\name{singleModelValuesAccessClass-class}
\alias{singleModelValuesAccess}
\alias{singleModelValuesAccessClass}
\alias{singleModelValuesAccessClass-class}
\title{Class \code{singleModelValuesAccessClass}}
\usage{
singleModelValuesAccess(modelValues, var)
}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

