% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp8XMcGu/file9e3134fd5ed4
\docType{class}
\name{singleModelValuesAccessClass-class}
\alias{singleModelValuesAccessClass}
\alias{singleModelValuesAccessClass-class}
\title{Class \code{singleModelValuesAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

