% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in RtmpNnT5lv/file751f7c8d7868
\docType{class}
\name{singleModelValuesAccessClass-class}
\alias{singleModelValuesAccess}
\alias{singleModelValuesAccessClass}
\alias{singleModelValuesAccessClass-class}
\title{Class \code{singleModelValuesAccessClass}}
\usage{
singleModelValuesAccess(modelValues, var)
}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

