% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp6hGOZh/file12a1f1777d547
\docType{class}
\name{singleModelValuesAccessClass-class}
\alias{singleModelValuesAccess}
\alias{singleModelValuesAccessClass}
\alias{singleModelValuesAccessClass-class}
\title{Class \code{singleModelValuesAccessClass}}
\usage{
singleModelValuesAccess(modelValues, var)
}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

