% Generated by roxygen2: do not edit by hand
% Please edit documentation in RtmpOW946u/file1481d8c0c26c
\docType{class}
\name{singleVarAccessClass-class}
\alias{singleVarAccessClass}
\alias{singleVarAccessClass-class}
\title{Class \code{singleVarAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

