% Generated by roxygen2: do not edit by hand
% Please edit documentation in Rtmp6hGOZh/file12a1f1777d547
\docType{class}
\name{singleVarAccessClass-class}
\alias{singleVarAccessClass}
\alias{singleVarAccessClass-class}
\title{Class \code{singleVarAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

