% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in Rtmp3xd4YE/filefb228f81eb0
\docType{class}
\name{singleVarAccessClass-class}
\alias{singleVarAccessClass}
\alias{singleVarAccessClass-class}
\title{Class \code{singleVarAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

