% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleFunction_compile.R
\docType{class}
\name{singleVarAccessClass-class}
\alias{singleVarAccessClass-class}
\alias{singleVarAccessClass}
\title{Class \code{singleVarAccessClass}}
\description{
Classes used internally in NIMBLE and not expected to be called directly by users.
}

