% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{svdNimbleList}
\alias{svdNimbleList}
\title{svdNimbleList definition}
\format{An object of class \code{list} of length 1.}
\usage{
svdNimbleList
}
\description{
\code{nimbleList} definition for the type of \code{nimbleList} returned by \code{\link{nimSvd}}.
}
\author{
NIMBLE development team
}
\seealso{
\code{\link{nimSvd}}
}
\keyword{datasets}

