% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleList_core.R
\docType{data}
\name{waicList}
\alias{waicList}
\title{waicList definition}
\format{
An object of class \code{list} of length 1.

An object of class \code{list} of length 1.
}
\usage{
waicList

waicList
}
\description{
\code{waicList} definition for the \code{nimbleList} type returned by WAIC
computation.

\code{waicList} definition for the \code{nimbleList} type returned by WAIC
computation.
}
\details{
See \code{help(waic)} for details on the elements of the list.

See \code{help(waic)} for details on the elements of the list.
}
\author{
NIMBLE development team

NIMBLE development team
}
\keyword{datasets}
